% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMobjects.R
\name{AAVY}
\alias{AAVY}
\title{Performance Metric: Probability AAVY < 0.2}
\usage{
AAVY(MSEobj = NULL, Ref = 0.2)
}
\arguments{
\item{MSEobj}{An object of class MSE}

\item{Ref}{Reference Point for AAVY (fraction)}
}
\value{
An object of class PMobj
}
\description{
Performance Metric: Probability AAVY < 0.2
}
\examples{
\dontrun{
AAVY(myMSE)
}
}
