% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{AJ_AJ_Rec_comb_GOM}
\alias{AJ_AJ_Rec_comb_GOM}
\title{AJ_AJ_Rec_comb_GOM OM}
\format{An object of class \code{OM} of length 1.}
\usage{
AJ_AJ_Rec_comb_GOM
}
\description{
An object of class OM
}
\keyword{datasets}
