% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_Exported.R
\name{alphaconv}
\alias{alphaconv}
\title{Calculate alpha parameter for beta distribution from mean and standard deviation}
\usage{
alphaconv(m, sd)
}
\arguments{
\item{m}{mean}

\item{sd}{standard deviation}
}
\value{
numeric
}
\description{
Calculate alpha parameter for beta distribution from mean and standard deviation
}
\author{
T. Carruthers
}
