% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Functions.R
\name{applyMP}
\alias{applyMP}
\title{Apply Management Procedures to an object of class Data}
\usage{
applyMP(Data, MPs = NA, reps = 100, nsims = NA, silent = FALSE)
}
\arguments{
\item{Data}{An object of class Data}

\item{MPs}{Name(s) of the MPs to run}

\item{reps}{Number of samples}

\item{nsims}{Optional. Number of simulations.}

\item{silent}{Logical. Should messages be suppressed?}
}
\value{
A list with the first element a list of management recommendations,
and the second the updated Data object
}
\description{
Apply Management Procedures to an object of class Data
}
