% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{Arrowtooth_Flounder_BC_DFO_iSCAM}
\alias{Arrowtooth_Flounder_BC_DFO_iSCAM}
\title{Arrowtooth_Flounder_BC_DFO_iSCAM OM}
\format{An object of class \code{OM} of length 1.}
\usage{
Arrowtooth_Flounder_BC_DFO_iSCAM
}
\description{
An object of class OM (built using iSCAM2DLM)
 No source provided. Author: No author provided.
}
\keyword{datasets}
