% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{Atlantic_mackerel}
\alias{Atlantic_mackerel}
\title{Atlantic_mackerel DLM_data}
\format{An object of class \code{DLM_data} of length 1.}
\usage{
Atlantic_mackerel
}
\description{
An object of class DLM_data
}
\keyword{datasets}
