% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFuns.r
\name{avail}
\alias{avail}
\title{What objects of this class are available}
\usage{
avail(classy)
}
\arguments{
\item{classy}{A class of object (character string, e.g. 'Fleet')}
}
\description{
Generic class finder
}
\details{
Finds objects of the specified class in the global environment or the
package:DLMtool
}
\author{
T. Carruthers
}
