% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Awatea2OM.R
\name{Awatea2OM}
\alias{Awatea2OM}
\title{Reads MCMC estimates from Awatea (Paul Starr) processed r file structure into an operating model}
\usage{
Awatea2OM(AwateaDir, nsim = 48, proyears = 50, Name = NULL,
  Source = "No source provided", Author = "No author provided",
  verbose = T)
}
\arguments{
\item{AwateaDir}{A folder with Awatea files}

\item{nsim}{The number of simulations to take for parameters with uncertainty (for OM@cpars custom parameters)}

\item{proyears}{The number of projection years for MSE}

\item{Name}{The name of the operating model}

\item{Source}{Reference to assessment documentation e.g. a url}

\item{Author}{Who did the assessment}

\item{verbose}{Should the r4ss function SS_ouput return detailed messages?}
}
\description{
A function that uses the file location of a fitted Awatea model post-processed into a set of rmd files
}
\author{
T. Carruthers
}
