% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFuns.r
\name{betaconv}
\alias{betaconv}
\title{Calculate beta parameter for beta distribution from mean and standard deviation}
\usage{
betaconv(m, sd)
}
\arguments{
\item{m}{mean}

\item{sd}{standard deviation}
}
\value{
numeric
}
\description{
Calculate beta parameter for beta distribution from mean and standard deviation
}
\author{
T. Carruthers
}
