% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineClasses.r
\docType{class}
\name{BioEco-class}
\alias{BioEco-class}
\title{Class \code{'BioEco'}}
\description{
An operating model component that specifies the bio-economic parameters
}
\section{Slots}{

\describe{
\item{\code{Name}}{The name of the Bio-Economic object. Single value. Character string.}

\item{\code{CostCurr}}{Cost of current effort. Uniform distribution lower and upper bounds. Non-negative real numbers}

\item{\code{RevCurr}}{Revenue of current catch. Uniform distribution lower and upper bounds. Non-negative real numbers}

\item{\code{CostInc}}{Annual increase in cost-per-unit-effort (percent). Uniform distribution lower and upper bounds. Non-negative real numbers}

\item{\code{RevInc}}{Annual increase in price-per-unit-catch (percent). Uniform distribution lower and upper bounds. Non-negative real numbers}

\item{\code{Response}}{Responsiveness of fishing effort to profitability. Uniform distribution lower and upper bounds. Non-negative real numbers}

\item{\code{LatentEff}}{Limited entry fishery - fraction of total allowable effort that is not currently being used. Optional - leave blank for open-access. Uniform distribution lower and upper bounds. Fraction <= 1}
}}

\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new('BioEco')}#'
}

\examples{

showClass('BioEco')

}
\author{
A. Hordyk
}
\keyword{classes}
