% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sketch_Functions.R
\name{BlankSelPlot}
\alias{BlankSelPlot}
\title{Internal function to create a blank plot for mapping selectivity at length}
\usage{
BlankSelPlot(Stock = NULL, Yr = NULL, N = NULL, isRel)
}
\arguments{
\item{Stock}{Stock object. If provided, average length-at-maturity
is included on plot for reference.}

\item{Yr}{Year number}

\item{N}{Year index}

\item{isRel}{Logical. Are selectivity parameters relative to length of maturity?}
}
\description{
Internal function to create a blank plot for mapping selectivity at length
}
\author{
A. Hordyk
}
\keyword{internal}
