% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{Blue_shark}
\alias{Blue_shark}
\title{Blue_shark Stock}
\format{An object of class \code{Stock} of length 1.}
\usage{
Blue_shark
}
\description{
An object of class Stock
}
\keyword{datasets}

