% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{Bluefin_tuna}
\alias{Bluefin_tuna}
\title{Bluefin_tuna Stock}
\format{An object of class \code{Stock} of length 1.}
\usage{
Bluefin_tuna
}
\description{
An object of class Stock
}
\keyword{datasets}

