% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{Bluefin_tuna_WAtl}
\alias{Bluefin_tuna_WAtl}
\title{Bluefin_tuna_WAtl Stock}
\format{An object of class \code{Stock} of length 1.}
\usage{
Bluefin_tuna_WAtl
}
\description{
An object of class Stock
}
\keyword{datasets}
