% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{Butterfish}
\alias{Butterfish}
\title{Butterfish Stock}
\format{An object of class \code{Stock} of length 1.}
\usage{
Butterfish
}
\description{
An object of class Stock
}
\keyword{datasets}
