% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineClasses.r
\name{calcMean}
\alias{calcMean}
\title{Calculate Mean Probabilty}
\usage{
calcMean(Prob, MSEobj)
}
\arguments{
\item{Prob}{Prob slot from an object of class PMobj}

\item{MSEobj}{An object of class MSE}
}
\description{
Calculate Mean Probabilty
}
