% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefinePMClasses.R
\name{calcProb}
\alias{calcProb}
\title{Calculate Probabilty}
\usage{
calcProb(PM, MSEobj)
}
\arguments{
\item{PM}{A PM method}

\item{MSEobj}{An object of class MSE}
}
\description{
Calculate Probabilty
}
