% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data2csv.R
\name{CALsimp}
\alias{CALsimp}
\title{Simplifies the CAL slot of data object}
\usage{
CALsimp(Data, nbins = 10, simno = 1)
}
\arguments{
\item{Data}{An object of class 'Data'.}

\item{nbins}{Integer. The target number of catch at length bins}

\item{simno}{Integer. An optional argument to specify the simulation number if writing simulated data}
}
\description{
A function that condenses the number of catch-at-length bins in a data object
}
\author{
T. Carruthers
}
