% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{Canary_Rockfish_BC_Canary_Rockfish_BC_DFO}
\alias{Canary_Rockfish_BC_Canary_Rockfish_BC_DFO}
\title{Canary_Rockfish_BC_Canary_Rockfish_BC_DFO OM}
\format{An object of class \code{OM} of length 1.}
\usage{
Canary_Rockfish_BC_Canary_Rockfish_BC_DFO
}
\description{
An object of class OM
}
\keyword{datasets}
