% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_Internal.R
\name{CanMP}
\alias{CanMP}
\title{What Data objects can be used to run this MP?}
\usage{
CanMP(MP)
}
\arguments{
\item{MP}{Name of the MP}
}
\description{
What Data objects can be used to run this MP?
}
\keyword{internal}
