% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_SupportingFunctions.R
\name{CC}
\alias{CC}
\title{Age-based Catch Curve}
\usage{
CC(x, Data, reps = 100, plot = FALSE)
}
\arguments{
\item{x}{Iteration number}

\item{Data}{An object of class \code{Data}}

\item{reps}{Number of reps}

\item{plot}{Logical. Show the plot?}
}
\value{
A vector of length \code{reps} of samples of the negative slope of the catch-curve (Z)
}
\description{
Age-based Catch Curve
}
\examples{
CC(1, DLMtool::SimulatedData, plot=TRUE)
}
