% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output.r
\name{CC4}
\alias{CC4}
\title{Constant catch management procedure of Geromont and Butterworth (2014)}
\usage{
CC4(x, Data, reps = 100, yrsmth = 5, xx=0.3)
}
\arguments{
\item{x}{A position in data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of TAC samples}

\item{yrsmth}{Years over which to average catches}

\item{xx}{Parameter controlling the TAC. Mean catches are multiplied by
(1-xx)}
}
\value{
A numeric vector of TAC recommendations
}
\description{
The TAC is the average catch over last yrsmth years reduced by 30%.
}
\details{
This is one of four constant catch MPs of Geromont and Butterworth 2014.
}
\references{
Geromont, H.F., Butterworth, D.S. 2014. Generic management
procedures for data-poor fisheries; forecasting with few data. ICES J. Mar.
Sci. doi:10.1093/icesjms/fst232
}
\author{
T. Carruthers
}
