% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_functions.R
\name{ChangeTS}
\alias{ChangeTS}
\title{Change the time-step for an Operating Model}
\usage{
ChangeTS(OM, ts = 4, silent = FALSE)
}
\arguments{
\item{OM}{An object of class \code{OM}}

\item{ts}{The number of sub-year time-steps. e.g, \code{ts=4} means a model with
quarterly time-steps. Must be an integer}

\item{silent}{Logical. Suppress messages?}
}
\value{
An object of class \code{OM} with sub-year time-steps
}
\description{
All rate parameters and other values relating to time-scale
are assumed to be in annual units in the input OM. For example, \code{OM@interval}
is in \emph{years}, as are \code{OM@M} and \code{OM@K} etc.
}
\details{
Note: OMs with \code{cpars} are not currently supported
}
