% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_Exported.R
\name{cheatsheets}
\alias{cheatsheets}
\title{Opens the DLMtool Cheat-Sheets (requires internet connection)}
\usage{
cheatsheets()
}
\description{
Opens the DLMtool Cheat-Sheets (requires internet connection)
}
\examples{
\dontrun{
cheatsheets()
}
}
