% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSEObj.r
\name{CheckConverg}
\alias{CheckConverg}
\title{MSE convergence diagnostic}
\usage{
CheckConverg(MSEobj,thresh=2, Plot=TRUE)
}
\arguments{
\item{MSEobj}{An object of class 'MSE'}

\item{thresh}{The convergence threshold (percentage). If mean perforamnce
metrics are within thresh percent of the second to last interation, the MSE
can be considered to have converged.}

\item{Plot}{Should figures be plotted?}
}
\description{
Have I undertaken enough simulations (nsim)? Has my MSE converged on stable
(reliable) peformance metrics?
}
\author{
A. Hordyk
}
