% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChooseSelect.r
\name{CheckSelect}
\alias{CheckSelect}
\title{Rough Plot of Historical Selectivity Patterns}
\usage{
CheckSelect(Fleet, Stock = NULL)
}
\description{
Rough Plot of Historical Selectivity Patterns
}
\author{
A. Hordyk
}
\keyword{internal}
