% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{China_rockfish}
\alias{China_rockfish}
\title{China_rockfish DLM_data}
\format{An object of class \code{DLM_data} of length 1.}
\usage{
China_rockfish
}
\description{
An object of class DLM_data
}
\keyword{datasets}

