% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output.r
\name{ChkDatNA}
\alias{ChkDatNA}
\title{Check Data object is valid for a MP}
\usage{
ChkDatNA(Data, dependencies)
}
\arguments{
\item{Data}{An object of class Data}

\item{dependencies}{A string of slots in the Data object required for the MP}
}
\description{
Checks that all slots in Data object required by the MP contain 
finite values
}
\author{
A. Hordyk
}
