% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input.r
\name{ChkInRec}
\alias{ChkInRec}
\title{Check that input control recommendation object is valid}
\usage{
ChkInRec(rec, Data)
}
\arguments{
\item{rec}{An input control recommendation object}

\item{Data}{A data object}
}
\value{
Nothing if all is ok, stops with an error if there is a problem with the input control
recommendation object
}
\description{
Check that input control recommendation object is valid
}
