% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_Internal.R
\name{ChkObj}
\alias{ChkObj}
\title{Check that a DLM object is valid}
\usage{
ChkObj(OM, error = TRUE)
}
\arguments{
\item{OM}{An object of class OM, Stock, Fleet, Obs, or Imp}

\item{error}{Logical. Stop on missing parameter values? FALSE = warning}
}
\description{
Check that all slots in Object are valid and contain values
}
