\name{ChooseEffort}
\alias{ChooseEffort}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
 Manually choose the historical relative fishing effort trajectory.
}
\description{
 Interactive plot which allows users to specify the relative trajectory and 
 variability in the historical fishing effort. 
}
\usage{
ChooseEffort(FleetObj, Years=NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{FleetObj}{
A fleet object.
}
  \item{Years}{
An optional vector of years. Should be nyears long.
}
}
\author{
A. Hordyk
}


