% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChooseSelect.r
\name{ChooseL5}
\alias{ChooseL5}
\title{Internal function for mapping selectivity}
\usage{
ChooseL5(Fleet, Stock, isRel)
}
\description{
Internal function for mapping selectivity
}
\author{
A. Hordyk
}
\keyword{internal}

