% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChooseSelect.r
\name{ChooseLFS}
\alias{ChooseLFS}
\title{Internal function for mapping selectivity}
\usage{
ChooseLFS(L5out, Fleet, Stock, isRel)
}
\description{
Internal function for mapping selectivity
}
\author{
A. Hordyk
}
\keyword{internal}

