% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChooseMort.R
\name{ChooseMort}
\alias{ChooseMort}
\title{Manually map natural mortality at age or size.}
\usage{
ChooseMort(OM, type = c("age", "length"), x = NULL, y = NULL)
}
\arguments{
\item{OM}{An object of class 'OM'}

\item{type}{A character string - is M to be mapped by 'age' or 'length'?}

\item{x}{Optional vector for x-axis}

\item{x}{Optional vector for y-axis}
}
\description{
Interactive plot which allows users to specify M by age or size class
}
\author{
A. Hordyk
}
