% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChooseSelect.r
\name{ChooseVmaxlen}
\alias{ChooseVmaxlen}
\title{Internal function for mapping selectivity}
\usage{
ChooseVmaxlen(Fleet, Stock, isRel)
}
\description{
Internal function for mapping selectivity
}
\author{
A. Hordyk
}
\keyword{internal}
