\name{comp}
\alias{comp}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Comparison plots for individual simulations
}
\description{
A simulation by simulation approach to plotting results
}
\usage{
comp(MSEobj,Meths=NA)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{MSEobj}{
An object of class MSE
}
\item{Meths}{
A character vector of two methods that were applied in making the MSE object
}
}
\author{
T. Carruthers
}
