% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Output.R
\name{CompSRA_}
\alias{CompSRA_}
\title{Internal function for CompSRA MP}
\usage{
CompSRA_(x, Data, reps = 100)
}
\arguments{
\item{x}{Simulation number}

\item{Data}{An object of class Data}

\item{reps}{Number of reps}
}
\value{
A list
}
\description{
Internal function for CompSRA MP
}
\keyword{internal}
