% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE_functions.R
\name{Converge}
\alias{Converge}
\title{Check Convergence}
\usage{
Converge(MSEobj, thresh = 2, Plot = TRUE)
}
\arguments{
\item{MSEobj}{An MSE object of class \code{'MSE'}}

\item{thresh}{The convergence threshold (percentage). If mean performance
metrics from the last 20 simulations are within thresh percent of the last interation, the MSE
can be considered to have converged.}

\item{Plot}{Should figures be plotted?}
}
\description{
Have I undertaken enough simulations (nsim)? Has my MSE converged on stable
(reliable) peformance metrics?
}
\author{
A. Hordyk
}
