% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFO.R
\name{COSEWIC_Blow}
\alias{COSEWIC_Blow}
\title{COSEWIC Blow plot}
\usage{
COSEWIC_Blow(MSEobj, syear = 2017, qcol = rgb(0.4, 0.8, 0.95),
  lcol = "dodgerblue4", quants = c(0.05, 0.25, 0.5, 0.75, 0.95), nGT = 3)
}
\arguments{
\item{MSEobj}{An object of class MSE created from runMSE() with the argument MPs=c("NFref","FMSYref","curE")}

\item{syear}{Starting year of the projection for graphing purposes}

\item{qcol}{Color of shaded regions (bars, quantiles)}

\item{lcol}{Color of lines}

\item{quants}{Quantiles of the shaded regions (vector 5 long e.g. 0.1, 0.2, 0.5, 0.8, 0.9)}

\item{nGT}{Number of generations for projections. The number of projection years should be greater than \code{MaxAge * nGT}}
}
\value{
A plot
}
\description{
These plots evaluate the likelihood of declining below Blow, by default biomass that it takes 3 generation times to reach half BMSY with zero fishing
}
\author{
T. Carruthers
}
