% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFO.R
\name{COSEWIC_plot}
\alias{COSEWIC_plot}
\title{COSEWIC forward projection plot}
\usage{
COSEWIC_plot(MSEobj, syear = 2015)
}
\arguments{
\item{MSEobj}{An object of class MSE created from runMSE() with the argument MPs=c("NFref","FMSYref","curE")}

\item{syear}{Starting year of the projection for graphing purposes}
}
\value{
A plot
}
\description{
Projection of biomass under three scenarios: no catch, FMSY fishing and status quo fishing
This plot if for an MSE object created from runMSE with the argument MPs=c("NFref","FMSYref","curE")
}
\author{
T. Carruthers
}
