% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFO.R
\name{COSEWIC_Pplot}
\alias{COSEWIC_Pplot}
\title{COSEWIC Projection plot}
\usage{
COSEWIC_Pplot(MSEobj, syear = 2017, qcol = rgb(0.4, 0.8, 0.95),
  lcol = "dodgerblue4", quants = c(0.05, 0.25, 0.5, 0.75, 0.95))
}
\arguments{
\item{MSEobj}{An object of class MSE created from runMSE() with the argument MPs=c("NFref","FMSYref","curE")}

\item{syear}{Starting year of the projection for graphing purposes}

\item{qcol}{Color of shaded regions (bars, quantiles)}

\item{lcol}{Color of lines}

\item{quants}{Quantiles of the shaded regions (vector 5 long e.g. 0.1, 0.2, 0.5, 0.8, 0.9)}
}
\value{
A plot
}
\description{
Projection of spawning stock biomass under three scenarios: no catch, FMSY fishing and status quo fishing effort
This plot if for an MSE object created from runMSE with the argument MPs=c("NFref","FMSYref","curE")
}
\author{
T. Carruthers
}
