% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFO.R
\name{COSEWIC_report}
\alias{COSEWIC_report}
\title{Create a standard DFO COSEWIC report}
\usage{
COSEWIC_report(MSEobj, output_file = NA, author = "Author not specified",
  title = NA)
}
\arguments{
\item{MSEobj}{An object of class MSE}

\item{output_file}{The directory and filename you wish to use for the report e.g. "C:/temp/myMSEreport.html"}

\item{author}{The person who made this report}

\item{title}{The title of the report}
}
\description{
Provides performance plots to inform COSEWIC processes in Canadian fish stocks.
}
\author{
T. Carruthers
}
