% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFO.R
\name{COSEWIC_tab}
\alias{COSEWIC_tab}
\title{Create a standard COSEWIC performance table}
\usage{
COSEWIC_tab(MSEobj, rnd = 0, GTs = c(3, 6), syear = 2017, nGT = 3)
}
\arguments{
\item{MSEobj}{An object of class MSE}

\item{rnd}{The number of significant figures for rounding.}

\item{GTs}{A vector of mean generation times to evaluate performance metrics over}

\item{syear}{Current year, starting year for projections (e.g. 2017)}

\item{nGT}{Number of generation times for moving window of SSB chance (metrics A1 and A2)}
}
\description{
P_Cr is the probability of being in the critical zone (less than 20% depletion)
P_Ct is the probability of being in the cautious zone (between 20% and 40% depletion)
P_H is the probability of being in the healthy zone (above 40% depletion)
P_Cr_MSY is the probability of being in the critical zone (less than 40% BMSY)
P_Ct_MSY is the probability of being in the cautious zone (between 40% and 80% BMSY)
P_H_MSY is the probability of being in the healthy zone (above 80% BMSY)
Caut is the probability of being in the cautious zone in the last 10 projected years
P_A1 is the probability of being designated threatened according to COSEWIC Indicator A1 (Spawning biomass less than 70% that three generation times previously)
P_A2 is the probability of being designated threatened according to COSEWIC Indicator A2 (Spawning biomass less than 50% that three generation times previously)
Blow is the probability that the stock is below the biomass for which it takes 3 generation times to reach 50% BMSY with zero fishing
}
\author{
T. Carruthers
}
