% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFO.R
\name{COSEWIC_tab_formatted}
\alias{COSEWIC_tab_formatted}
\title{A formatted version of the standard COSEWIC performance plot, color coded by thresholds}
\usage{
COSEWIC_tab_formatted(Ptab1, thresh = c(10, 40, 50, 10, 40, 50, 20, 20, 5))
}
\arguments{
\item{Ptab1}{A COSEWIC performance table made by COSEWIC_tab()}

\item{thresh}{A vector of thresholds for each column Health, Yield and Reb are 'greater than threshold' conditions}
}
\description{
P_Cr is the probability of being in the critical zone (less than 20% depletion)
P_Ct is the probability of being in the cautious zone (between 20% and 40% depletion)
P_H is the probability of being in the healthy zone (above 40% depletion)
P_Cr_MSY is the probability of being in the critical zone (less than 40% BMSY)
P_Ct_MSY is the probability of being in the cautious zone (between 40% and 80% BMSY)
P_H_MSY is the probability of being in the healthy zone (above 80% BMSY)
Caut is the probability of being in the cautious zone in the last 10 projected years
P_A1 is the probability of being designated threatened according to COSEWIC Indicator A1 (Spawning biomass less than 70% that three generation times previously)
P_A2 is the probability of being designated threatened according to COSEWIC Indicator A2 (Spawning biomass less than 50% that three generation times previously)
Blow is the probability that the stock is below the biomass for which it takes 3 generation times to reach 50% BMSY with zero fishing
}
\author{
T. Carruthers
}
