% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMSE.r
\name{cparscheck}
\alias{cparscheck}
\title{Internal function of runMSE for checking that the OM slot cpars slot is formatted correctly}
\usage{
cparscheck(cpars)
}
\arguments{
\item{cpars}{a list of model parameters to be sampled (single parameters are a vector nsim long, time series are matrices nsim x nyears)}
}
\value{
either an error and the length of the first dimension of the various cpars list items or passes and returns the number of simulations
}
\description{
Internal function of runMSE for checking that the OM slot cpars slot is formatted correctly
}
\author{
T. Carruthers
}
