% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_Exported.R
\name{cv}
\alias{cv}
\title{Calculate CV from vector of values}
\usage{
cv(x)
}
\arguments{
\item{x}{vector of numeric values}
}
\value{
numeric
}
\description{
Calculate CV from vector of values
}
\author{
T. Carruthers
}
\keyword{internal}
