% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output.r
\name{DAAC}
\alias{DAAC}
\title{Depletion Adjusted Average Catch}
\usage{
DAAC(x, Data, reps = 100)
}
\arguments{
\item{x}{A position in a data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of stochastic samples of the TAC recommendation}
}
\description{
Essentially DCAC multiplied by 2*depletion and divided by BMSY/B0 (Bpeak)
}
\references{
MacCall, A.D., 2009. Depletion-corrected average catch: a simple
formula for estimating sustainable yields in data-poor situations. ICES J.
Mar. Sci. 66, 2267-2271. Harford W. and Carruthers, T. 2016. Simulation
testing novel catch-based fisheries management. In draft, intended for Fish.
Bull.
}
\author{
W. Harford and T. Carruthers
}
