% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data2csv.R
\name{Data2csv}
\alias{Data2csv}
\title{Converts a Data object into a .csv data file}
\usage{
Data2csv(Data, file = NULL, simno = 1, overwrite = F, keepNAs = T)
}
\arguments{
\item{Data}{An object of class 'Data'.}

\item{file}{Character string. The name of the location and file you wish to create (e.g. "C:/temp/mydata.csv")}

\item{simno}{Integer. An optional argument to specify the simulation number if writing simulated data}

\item{overwrite}{Boolean. Should existing data files be automatically overwritten.}

\item{keepNAs}{Boolean. Should slots with NAs still be written to the data file.}
}
\description{
A function that writes a correctly formatted .csv file from a DLMtool / MSEtool Data object
}
\author{
T. Carruthers
}
