% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_init_doc.R
\name{Data_xl}
\alias{Data_xl}
\title{Read in Data object from Excel spreadsheet}
\usage{
Data_xl(fname, stkname, fpath = "", saveCSV = FALSE)
}
\arguments{
\item{fname}{Name of the Excel spreadsheet file. Must include file
extension.}

\item{stkname}{Name of the Stock.}

\item{fpath}{Full file path, if file is not in current working directory}

\item{saveCSV}{Do you also want to the Data parameters to a CSV file?}
}
\value{
A object of class Data
}
\description{
A function to read in Data object from an Excel spreadsheet
with tabs named following specific convention.
}
\details{
The Excel spreadsheet must have tabs named with the following convention.
For example if \code{stkname} is 'myFish', the Data parameters are in a tab
named 'myFishData'.
}
\examples{

\dontrun{
OM <- OM_xl(fname='OMTables.xlsx', stkname='myFish')
}

}
\author{
A. Hordyk
}
