% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{DataDescription}
\alias{DataDescription}
\title{DataDescription}
\format{An object of class \code{data.frame} with 78 rows and 2 columns.}
\usage{
DataDescription
}
\description{
A data.frame with description of slots for class Data
}
\keyword{datasets}
