% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Output.R
\name{DBSRA_}
\alias{DBSRA_}
\title{Depletion-based SRA internal function}
\usage{
DBSRA_(x, Data, reps = 100, plot = FALSE, depo = NULL, hcr = NULL)
}
\arguments{
\item{x}{A position in a data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of stochastic samples of the TAC recommendation}

\item{plot}{Logical. Show the plot?}

\item{depo}{Optional fixed depletion (single value)}

\item{hcr}{Optional harvest control rule for throttling catch as a function of B/B0.
Numeric vector of length 2 specifying HCR break points - e.g c(0.4, 0.1) for 40-10 HCR}
}
\description{
Depletion-based SRA internal function
}
\keyword{internal}
