% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Output.R
\name{DBSRAopt}
\alias{DBSRAopt}
\title{Internal optimization function}
\usage{
DBSRAopt(lnK, C_hist, nys, Mdb, FMSY_M, BMSY_K, Bt_K, adelay, opt = 1)
}
\arguments{
\item{lnK}{log unfished biomass}

\item{C_hist}{Historical catches}

\item{nys}{number of years (length of C_hist)}

\item{Mdb}{M}

\item{FMSY_M}{FMSY/M}

\item{BMSY_K}{BMSY/K}

\item{Bt_K}{Current depletion}

\item{adelay}{age of maturity}

\item{opt}{Numeric. 1 for optimization. 2 to return Biomass trend}
}
\description{
Internal optimization function
}
\keyword{internal}
