% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Output.R
\name{DCAC_}
\alias{DCAC_}
\title{Depletion Corrected Average Catch}
\usage{
DCAC_(x, Data, reps = 100, Bt_K = NULL, updateD = FALSE)
}
\arguments{
\item{x}{Simulation number}

\item{Data}{An object of class \code{Data}}

\item{reps}{The number of reps}

\item{Bt_K}{Optional value of depletion (otherwise taken from Data object)}

\item{updateD}{Logical. Should depletion be updated in projection years?}
}
\value{
A list with dcac, Bt_K, and BMSY_K
}
\description{
Internal code to calculate dcac
}
\keyword{internal}
