% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output.r
\name{DCAC_40}
\alias{DCAC_40}
\title{Depletion Corrected Average Catch assuming 40 per cent stock depletion}
\usage{
DCAC_40(x, Data, reps = 100)
}
\arguments{
\item{x}{A position in a data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of stochastic samples of the TAC recommendation}
}
\description{
DCAC assuming that current stock biomass is exactly 40 per cent of unfished
levels.
}
\note{
The 40 percent depletion assumption doesn't really affect DCAC that
much as it already makes TAC recommendations that are quite MSY-like.
}
\references{
MacCall, A.D., 2009. Depletion-corrected average catch: a simple
formula for estimating sustainable yields in data-poor situations. ICES J.
Mar. Sci. 66, 2267-2271.
}
\author{
T. Carruthers
}
