% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Output.R
\name{DD4010}
\alias{DD4010}
\title{Delay - Difference Stock Assessment with UMSY and MSY leading coupled with a
40-10 harvest control rule}
\usage{
DD4010(x, Data, reps = 100)
}
\arguments{
\item{x}{A position in a data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of stochastic samples of the TAC recommendation}
}
\value{
A numeric vector of TAC recommendations
}
\description{
A simple delay-difference assessment that estimates the OFL using a
time-series of catches and a relative abundance index. In this version of
the DD MP a 40-10 rule is imposed over the OFL recommendation.
}
\references{
Carruthers, T, Walters, C.J,, and McAllister, M.K. 2012. Evaluating methods that classify
fisheries stock status using only fisheries catch data. Fisheries Research 119-120:66-79.

Hilborn, R., and Walters, C., 1992. Quantitative Fisheries Stock Assessment: Choice,
Dynamics and Uncertainty. Chapman and Hall, New York.
}
\author{
T. Carruthers
}
