% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Output.R
\name{DD_}
\alias{DD_}
\title{Delay-Difference Internal Function}
\usage{
DD_(x, Data, reps = 100, hcr = NULL)
}
\arguments{
\item{x}{iteration number}

\item{Data}{object of class Data}

\item{reps}{number of reps}

\item{hcr}{Optional harvest control rule for throttling catch as a function of B/B0.
Numeric vector of length 2 specifying HCR break points - e.g c(0.4, 0.1) for 40-10 HCR}
}
\description{
Delay-Difference Internal Function
}
\keyword{internal}
