% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Output.R
\name{DD_R}
\alias{DD_R}
\title{Delay-Difference Internal Function}
\usage{
DD_R(params, opty, So_DD, Alpha_DD, Rho_DD, ny_DD, k_DD, wa_DD, E_hist, C_hist,
  UMSYprior)
}
\arguments{
\item{params}{xx}

\item{opty}{xx}

\item{So_DD}{xx}

\item{Alpha_DD}{xx}

\item{Rho_DD}{xx}

\item{ny_DD}{xx}

\item{k_DD}{xx}

\item{wa_DD}{xx}

\item{E_hist}{xx}

\item{C_hist}{xx}

\item{UMSYprior}{xx}
}
\description{
Delay-Difference Internal Function
}
\keyword{internal}
