% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD_TMB.R
\name{DD_TMB}
\alias{DD_TMB}
\title{Delay - Difference Stock Assessment with UMSY and MSY leading using TMB}
\usage{
DD_TMB(x, Data, reps = 100)
}
\arguments{
\item{x}{A position in a data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of stochastic samples of the TAC recommendation}
}
\value{
A numeric vector of TAC recommendations
}
\description{
A simple delay-difference assessment that estimates the TAC using a
time-series of catches and a relative abundance index and coded with TMB
}
\note{
This DD model is observation error only and has does not estimate
process error (recruitment deviations). Similar to many other assessment
models it depends on a whole host of dubious assumptions such as temporally
stationary productivity and proportionality between the abundance index and
real abundance. Unsurprisingly the extent to which these assumptions are
violated tends to be the biggest driver of performance for this method.
}
\references{
Method based on equations of Carl Walters (bug him with
questions and expect colourful responses)
}
\author{
T. Carruthers & Z. Siders. Zach Siders coded the TMB function
}
