% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Input.R
\name{DDe}
\alias{DDe}
\alias{DDes}
\alias{DDe75}
\title{DDe}
\usage{
DDe(x, Data, reps = 100)

DDes(x, Data, reps = 100, LB = 0.9, UB = 1.1)

DDe75(x, Data, reps = 100)
}
\arguments{
\item{x}{rr}

\item{Data}{rr}

\item{reps}{rr}

\item{LB}{The lowest permitted factor of previous fishing effort}

\item{UB}{The highest permitted factor of previous fishing effort}
}
\description{
Description
}
\details{
Details
}
\section{Functions}{
\itemize{
\item \code{DDe}: Effort-control version. The recommended effort is EMSY.

\item \code{DDes}: Variant of \code{DDe} that limits the maximum change in effort to 10 percent.

\item \code{DDe75}: Variant of \code{DDe} where the recommended effort is 75\% EMSY.
}}

\seealso{
Other Delay-Difference MPs: \code{\link{DD}}
}
