% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{DecE_HDom}
\alias{DecE_HDom}
\title{DecE_HDom Fleet}
\format{An object of class \code{Fleet} of length 1.}
\usage{
DecE_HDom
}
\description{
An object of class Fleet
}
\keyword{datasets}
