% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{DecE_NDom}
\alias{DecE_NDom}
\title{DecE_NDom Fleet}
\format{An object of class \code{Fleet} of length 1.}
\usage{
DecE_NDom
}
\description{
An object of class Fleet
}
\keyword{datasets}

