% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Output.R
\name{demographic2}
\alias{demographic2}
\alias{demofn}
\title{Internal demographic function}
\usage{
demographic2(log.r, M, amat, sigma, K, Linf, to, hR, maxage, a, b)

demofn(log.r, M, amat, sigma, K, Linf, to, hR, maxage, a, b)
}
\arguments{
\item{log.r}{internal}

\item{M}{internal}

\item{amat}{internal}

\item{sigma}{internal}

\item{K}{internal}

\item{Linf}{internal}

\item{to}{internal}

\item{hR}{internal}

\item{maxage}{internal}

\item{a}{internal}

\item{b}{internal}
}
\description{
Internal demographic function
}
\section{Functions}{
\itemize{
\item \code{demofn}: Internal function
}}

\keyword{internal}
