% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output.r
\docType{data}
\name{DepF}
\alias{DepF}
\title{Depletion Corrected Fratio}
\format{An object of class \code{Output} of length 1.}
\usage{
DepF(x, Data, reps = 100)
}
\arguments{
\item{x}{A position in data-limited methods data object DLM}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of TAC samples}
}
\value{
A numeric vector of TAC recommendations
}
\description{
The Fratio MP with a harvest control rule that reduces F according to the
production curve given an estimate of current stock depletion.
}
\author{
T. Carruthers
}
\references{
Made-up for this package.
}
\keyword{datasets}

