% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSteepness.R
\name{derive_beta_par}
\alias{derive_beta_par}
\title{This function reduces the CV by 5 per cent until steepness values can be sampled without error}
\usage{
derive_beta_par(mu, sigma)
}
\arguments{
\item{mu}{mean h}

\item{sigma}{sd of h}
}
\description{
This function reduces the CV by 5 per cent until steepness values can be sampled without error
}
\author{
Q. Huynh
}
