% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFO.R
\name{DFO_bar}
\alias{DFO_bar}
\title{Deparment of Fisheries and Oceans stock status bar plot}
\usage{
DFO_bar(MSEobj, yres = 10)
}
\arguments{
\item{MSEobj}{An MSE object of class MSE produced by DLMtool function runMSE}

\item{yres}{Integer: the year interval over which to calculate B/BMSY in future years}
}
\description{
A plot of biomass relative to BMSY over projected years
}
\author{
T. Carruthers
}
