% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFO.R
\name{DFO_plot}
\alias{DFO_plot}
\title{Deparment of Fisheries and Oceans trade-off plot}
\usage{
DFO_plot(MSEobj)
}
\arguments{
\item{MSEobj}{An MSE object of class MSE produced by DLMtool function runMSE}
}
\description{
A plot of mean biomass relative to BMSY and fishing mortality rate relative to FMSY 
over the final 5 years of the projection
http://www.dfo-mpo.gc.ca/reports-rapports/regs/sff-cpd/precaution-eng.htm
}
\author{
T. Carruthers
}
