% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFO.R
\name{DFO_plot2}
\alias{DFO_plot2}
\title{Deparment of Fisheries and Oceans default plot 2}
\usage{
DFO_plot2(MSEobj,nam=NA,type=NA,panel=T)
}
\arguments{
\item{MSEobj}{An object of class MSE}

\item{nam}{Title of plot}

\item{type}{Plots full range of data if NA. Plots a subset that meet thresholds if not NA.}
}
\value{
A table of performance metrics.
}
\description{
A preliminary plot for returning trade-offs plots and performance table for
total yield, variability in yield, probability of overfishing and likelihood
of biomass dropping below 50 per cent BMSY
}
\author{
T. Carruthers
}
