% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFO.R
\name{DFO_quant}
\alias{DFO_quant}
\title{Deparment of Fisheries and Oceans biomass quantile plot}
\usage{
DFO_quant(MSEobj, maxcol = 6, qcol = rgb(0.4, 0.8, 0.95),
  lcol = "dodgerblue4", curyr = 2018, quants = c(0.1, 0.9), addline = T)
}
\arguments{
\item{MSEobj}{An MSE object of class MSE produced by DLMtool function runMSE}

\item{maxcol}{Integer how many columns for panel plots?}

\item{qcol}{A color, the quantile coloration}

\item{lcol}{A color, the mean B/BMSY line}

\item{curyr}{The current calendar year}

\item{quants}{A vector 2 long for the quantiles e.g. 0.1 and 0.9 for the 10th and 90th quantiles}

\item{addline}{Should two individual example simulations be added to the plot?}
}
\description{
A plot of biomass relative to BMSY quantiles over projected years
}
\author{
T. Carruthers
}
