% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_Exported.R
\docType{data}
\name{DLMenv}
\alias{DLMenv}
\title{DLMenv blank environment}
\format{
An object of class \code{environment} of length 0.
}
\usage{
DLMenv
}
\description{
An environment allocated for MPs to print model output during the
management strategy evaluation. Is blank at the beginning of each call to \code{runMSE}.
}
\seealso{
\link{runMSE}
}
\keyword{datasets}
