% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLMextra.R
\name{DLMextra}
\alias{DLMextra}
\title{Load more data from DLMextra package}
\usage{
DLMextra(silent = FALSE)
}
\arguments{
\item{silent}{Logical. Should messages to printed?}
}
\description{
Downloads the DLMextra package from GitHub
}
