% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLMtool-package.R
\docType{package}
\name{DLMtool-package}
\alias{DLMtool-package}
\alias{DLMtool}
\title{Data-Limited Methods Toolkit}
\description{
Simulation testing and implementation of data-limited fishery stock
assessment methods
}
\references{
Carruthers, T.R., Punt, A.E., Walters, C.J., MacCall, A.,
McAllister, M.K., Dick, E.J., Cope, J. 2014. Evaluating methods for setting
catch limits in data-limited fisheries. Fisheries Research. 153: 48-68.

Carruthers, T.R., Kell, L.T., Butterworth, D.S., Maunder, M.N., Geromont,
H.F., Walters, C., McAllister, M.K., Hillary, R., Levontin, P., Kitakado,
T., Davies, C.R. Performance review of simple management procedures. ICES
Journal of Marine Science.
}
\author{
Tom Carruthers <t.carruthers@fisheries.ubc.ca> Adrian Hordyk
<a.hordyk@murdoch.edu.au>
}
\keyword{assessment}
\keyword{data-limited}
\keyword{data-poor}
\keyword{evaluation}
\keyword{fisheries}
\keyword{management}
\keyword{stock}
\keyword{strategy}
