% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_Internal.R
\name{dnormal}
\alias{dnormal}
\title{Double-normal selectivity curve}
\usage{
dnormal(lens, lfs, sl, sr)
}
\arguments{
\item{lens}{Vector of lengths}

\item{lfs}{Length at full selection}

\item{sl}{Sigma of ascending limb}

\item{sr}{Sigma of descending limb}
}
\description{
Double-normal selectivity curve
}
