% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscInt.r
\name{doprojPI}
\alias{doprojPI}
\title{Internal Projection Function}
\usage{
doprojPI(lnF, Mvec, Wac, Mac, Pc, N_c, SSN_c, Biomass_c, VBiomass_c, SSB_c, Vc,
  hc, R0ac, proyears, nareas, maxage, movc, SSBpRc, aRc, bRc, SRrelc, spat_targ)
}
\arguments{
\item{lnF}{internal parameter}

\item{Mvec}{internal parameter}

\item{Wac}{internal parameter}

\item{Pc}{internal parameter}

\item{N_c}{internal parameter}

\item{SSN_c}{internal parameter}

\item{Biomass_c}{internal parameter}

\item{VBiomass_c}{internal parameter}

\item{SSB_c}{internal parameter}

\item{Vc}{internal parameter}

\item{hc}{internal parameter}

\item{R0ac}{internal parameter}

\item{proyears}{internal parameter}

\item{nareas}{internal parameter}

\item{maxage}{internal parameter}

\item{movc}{internal parameter}

\item{SSBpRc}{internal parameter}

\item{aRc}{internal parameter}

\item{bRc}{internal parameter}

\item{SRrelc}{internal parameter}

\item{spat_targ}{internal parameter}
}
\description{
Internal Projection Function
}
\keyword{internal}
