% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{doprojPI_cpp}
\alias{doprojPI_cpp}
\title{Rcpp version of the Projection function for calculating Reference Yield}
\usage{
doprojPI_cpp(lnF, Mvec, Wac, Mac, Pc, N_c, SSN_c, Biomass_c, VBiomass_c, SSB_c,
  Vc, hc, R0ac, proyears, nareas, maxage, movc, SSBpRc, aRc, bRc, SRrelc,
  Spat_targc)
}
\arguments{
\item{lnF}{internal}

\item{Mvec}{internal}

\item{Wac}{internal}

\item{Mac}{internal}

\item{Pc}{internal}

\item{N_c}{internal}

\item{SSN_c}{internal}

\item{Biomass_c}{internal}

\item{VBiomass_c}{internal}

\item{SSB_c}{internal}

\item{Vc}{internal}

\item{hc}{internal}

\item{R0ac}{internal}

\item{proyears}{internal}

\item{nareas}{internal}

\item{maxage}{internal}

\item{movc}{internal}

\item{SSBpRc}{internal}

\item{aRc}{internal}

\item{bRc}{internal}

\item{SRrelc}{internal}

\item{Spat_targc}{internal}
}
\description{
Rcpp version of the Projection function for calculating Reference Yield
}
\keyword{internal}
