% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{Example_datafile}
\alias{Example_datafile}
\title{Example_datafile DLM_data}
\format{An object of class \code{DLM_data} of length 1.}
\usage{
Example_datafile
}
\description{
An object of class DLM_data
}
\keyword{datasets}

