% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoxyPMobj.r
\docType{data}
\name{F_FMSY}
\alias{F_FMSY}
\title{F_FMSY Performance Metric Object}
\format{An object of class \code{PM} of length 1.}
\usage{
F_FMSY
}
\description{
An object of class PM
}
\keyword{datasets}
