% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Output.R
\name{Fdem_}
\alias{Fdem_}
\title{Internal function for Fdem MP}
\usage{
Fdem_(x, Data, reps, Ac = NULL)
}
\arguments{
\item{x}{internal}

\item{Data}{internal}

\item{reps}{internal}

\item{Ac}{Optional abundance value}
}
\description{
Internal function for Fdem MP
}
\keyword{internal}
