% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output.r
\name{Fdem_ML}
\alias{Fdem_ML}
\title{Demographic FMSY method that uses mean length data to estimate recent Z}
\usage{
Fdem_ML(x, Data, reps = 100)
}
\arguments{
\item{x}{A position in data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of TAC samples}
}
\description{
Demographic F (r/2) method using the mean length estimator to calculate
current abundance.
}
\note{
The mean length extension was programmed by Gary Nelson as part of his
excellent R package 'fishmethods'
}
\references{
McAllister, M.K., Pikitch, E.K., and Babcock, E.A. 2001. Using
demographic methods to construct Bayesian priors for the intrinsic rate of
increase in the Schaefer model and implications for stock rebuilding. Can.
J. Fish. Aquat. Sci. 58: 1871-1890.
}
\author{
T. Carruthers
}
