% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{Feasibility}
\alias{Feasibility}
\alias{Feasibility2}
\alias{Fease-class-objects}
\title{Fease class objects}
\format{An object of class \code{Fease} of length 1.}
\usage{
Feasibility

Feasibility2
}
\description{
Example objects of class Fease
}
\examples{
data(Feasibility2)


avail("Fease")
}
\keyword{datasets}
