% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{Feasibility}
\alias{Feasibility}
\title{Feasibility Fease}
\format{An object of class \code{Fease} of length 1.}
\usage{
Feasibility
}
\description{
An object of class Fease
}
\keyword{datasets}

