% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{Feasibility2}
\alias{Feasibility2}
\title{Feasibility2 Fease}
\format{An object of class \code{Fease} of length 1.}
\usage{
Feasibility2
}
\description{
An object of class Fease
}
\keyword{datasets}
