% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSCAM2DLM.R
\name{fetch.file.names}
\alias{fetch.file.names}
\title{Reads iSCAM Data, Control and Projection files}
\usage{
fetch.file.names(path, filename)
}
\arguments{
\item{path}{File path}

\item{filename}{The filename}
}
\description{
A function for returning the three types of 
iSCAM input and output files
}
\author{
Chris Grandin (DFO PBS)
}
