% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Input.R
\name{FilterSmooth}
\alias{FilterSmooth}
\title{Kalman filter and Rauch-Tung-Striebel smoother}
\usage{
FilterSmooth(RawEsts, R = 0.2, Q = 0.1, Int = 100)
}
\arguments{
\item{RawEsts}{a vector of estimated values}

\item{R}{variance of sampling noise}

\item{Q}{variance of random walk increments}

\item{Int}{covariance of initial uncertainty}
}
\value{
a vector of smoothed values
}
\description{
A function that applies a filter and smoother to estimates
}
\keyword{internal}
