% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{FlatE_Dom}
\alias{FlatE_Dom}
\title{FlatE_Dom Fleet}
\format{An object of class \code{Fleet} of length 1.}
\usage{
FlatE_Dom
}
\description{
An object of class Fleet
}
\keyword{datasets}
