% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{FlatE_HDom}
\alias{FlatE_HDom}
\title{FlatE_HDom Fleet}
\format{An object of class \code{Fleet} of length 1.}
\usage{
FlatE_HDom
}
\description{
An object of class Fleet
}
\keyword{datasets}
