% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{FlatE_NDom}
\alias{FlatE_NDom}
\title{FlatE_NDom Fleet}
\format{An object of class \code{Fleet} of length 1.}
\usage{
FlatE_NDom
}
\description{
An object of class Fleet
}
\keyword{datasets}
