% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{DecE_Dom}
\alias{DecE_Dom}
\alias{DecE_HDom}
\alias{DecE_NDom}
\alias{FlatE_Dom}
\alias{FlatE_HDom}
\alias{FlatE_NDom}
\alias{Generic_DecE}
\alias{Generic_FlatE}
\alias{Generic_Fleet}
\alias{Generic_IncE}
\alias{IncE_HDom}
\alias{IncE_NDom}
\alias{Low_Effort_Non_Target}
\alias{Target_All_Fish}
\alias{Targeting_Small_Fish}
\alias{Fleet-class-objects}
\title{Fleet class objects}
\format{An object of class \code{Fleet} of length 1.}
\usage{
DecE_Dom

DecE_HDom

DecE_NDom

FlatE_Dom

FlatE_HDom

FlatE_NDom

Generic_DecE

Generic_FlatE

Generic_Fleet

Generic_IncE

IncE_HDom

IncE_NDom

Low_Effort_Non_Target

Target_All_Fish

Targeting_Small_Fish
}
\description{
Example objects of class Fleet
}
\examples{
Targeting_Small_Fish@Name

avail("Fleet")
}
\keyword{datasets}
