% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{FleetDescription}
\alias{FleetDescription}
\title{FleetDescription}
\format{An object of class \code{data.frame} with 27 rows and 2 columns.}
\usage{
FleetDescription
}
\description{
A data.frame with description of slots for class Fleet
}
\keyword{datasets}
