% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscInt.r
\name{FMSYopt}
\alias{FMSYopt}
\title{Internal function FMSY and related metrics}
\usage{
FMSYopt(lnF, Mc, hc, Mac, Wac, R0c, Vc, maxage, nyears, proyears, Spat_targc,
  movc, SRrelc, aRc, bRc, Opt = T)
}
\arguments{
\item{lnF}{internal parameter}

\item{Mc}{internal parameter}

\item{hc}{internal parameter}

\item{Mac}{internal parameter}

\item{Wac}{internal parameter}

\item{R0c}{internal parameter}

\item{Vc}{internal parameter}

\item{maxage}{internal parameter}

\item{nyears}{internal parameter}

\item{proyears}{internal parameter}

\item{Spat_targc}{internal parameter}

\item{movc}{internal parameter}

\item{SRrelc}{internal parameter}

\item{aRc}{internal parameter}

\item{bRc}{internal parameter}

\item{Opt}{internal parameter}
}
\description{
Internal function FMSY and related metrics
}
\keyword{internal}

