% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Reference.R
\name{FMSYref50}
\alias{FMSYref50}
\title{A reference FMSY method that fishes at half of FMSY (uses perfect
information about FMSY)}
\usage{
FMSYref50(x, Data, reps = 100)
}
\arguments{
\item{x}{A position in data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of TAC (OFL) samples}
}
\description{
FMSY is taken from the operating model stored at DLM@OM$FMSY
}
\details{
Note that you can out-performm this method easily. The requirement for fixed
F is actually quite strict and is by no means the upper limit in terms of
yield. Don't panic if your method beats this one for yield!

Interesting that the reduction in yield is no way near commensurate with the
reduction in F - as predicted by a yield curve and expressed in the pretty
good yield theory.
}
\author{
T. Carruthers
}
