% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Output.R
\name{DepF}
\alias{DepF}
\alias{Fratio}
\alias{Fratio4010}
\alias{Fratio_CC}
\alias{Fratio_ML}
\title{Depletion Corrected Fratio}
\usage{
DepF(x, Data, reps = 100)

Fratio(x, Data, reps = 100)

Fratio4010(x, Data, reps = 100)

Fratio_CC(x, Data, reps = 100, Fmin = 0.005)

Fratio_ML(x, Data, reps = 100)
}
\arguments{
\item{x}{A position in a data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of samples of the TAC recommendation}

\item{Fmin}{Minimum current fishing mortality rate for the catch-curve
analysis}
}
\value{
A Rec object with TAC recommendations
}
\description{
Depletion Corrected Fratio

Calculates the OFL based on a fixed ratio of FMSY to M multiplied by a
current estimate of abundance.
}
\section{Functions}{
\itemize{
\item \code{DepF}: Depletion Corrected Fratio: the Fratio MP with a harvest control
rule that reduces F according to the production curve given an estimate of current
stock depletion (made-up for this package).

\item \code{Fratio}: A simple method that tends to outperform many other approaches alarmingly
often even when current biomass is relatively poorly known. The low stock
crash potential is largely due to the quite large difference between Fmax
and FMSY for most stocks.

\item \code{Fratio4010}: Paired with the 40-10
rule that throttles back the OFL to zero at 10 percent of unfished biomass.

\item \code{Fratio_CC}: Pairs with a catch curve estimate of current stock size.

\item \code{Fratio_ML}: Pairs with an estimate of
current stock size from a mean-length estimator.
}}

\references{
Gulland, J.A., 1971. The fish resources of the ocean. Fishing
News Books, West Byfleet, UK.

Martell, S., Froese, R., 2012. A simple method for estimating MSY from catch
and resilience. Fish Fish. doi: 10.1111/j.1467-2979.2012.00485.x.
}
\seealso{
\link{DynF}
}
\author{
T. Carruthers
}
