% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Output.R
\name{Fratio_}
\alias{Fratio_}
\title{Fratio internal function}
\usage{
Fratio_(x, Data, reps = 100, Abun = NULL)
}
\arguments{
\item{x}{Iteration number}

\item{Data}{Object of class Data}

\item{reps}{Number of reps}

\item{Abun}{Optional estimate of abundance}
}
\description{
Fratio internal function
}
\keyword{internal}
