% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{Generic_DecE}
\alias{Generic_DecE}
\title{Generic_DecE Fleet}
\format{An object of class \code{Fleet} of length 1.}
\usage{
Generic_DecE
}
\description{
An object of class Fleet
}
\keyword{datasets}
