% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{Generic_FlatE}
\alias{Generic_FlatE}
\title{Generic_FlatE Fleet}
\format{An object of class \code{Fleet} of length 1.}
\usage{
Generic_FlatE
}
\description{
An object of class Fleet
}
\keyword{datasets}

