% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{Generic_Fleet}
\alias{Generic_Fleet}
\title{Generic_Fleet Fleet}
\format{An object of class \code{Fleet} of length 1.}
\usage{
Generic_Fleet
}
\description{
An object of class Fleet
}
\keyword{datasets}
