% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{Generic_IncE}
\alias{Generic_IncE}
\title{Generic_IncE Fleet}
\format{An object of class \code{Fleet} of length 1.}
\usage{
Generic_IncE
}
\description{
An object of class Fleet
}
\keyword{datasets}
