% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{Generic_obs}
\alias{Generic_obs}
\title{Generic_obs Observation}
\format{An object of class \code{Observation} of length 1.}
\usage{
Generic_obs
}
\description{
An object of class Observation
}
\keyword{datasets}

