% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_Internal.R
\name{genSizeCompWrap}
\alias{genSizeCompWrap}
\alias{getfifth}
\title{Wrapper for C++ function to generate length composition}
\usage{
genSizeCompWrap(
  i,
  vn,
  CAL_binsmid,
  retL,
  CAL_ESS,
  CAL_nsamp,
  Linfarray,
  Karray,
  t0array,
  LenCV,
  truncSD = 2
)

getfifth(lenvec, CAL_binsmid)
}
\arguments{
\item{i}{Simulation number}

\item{vn}{Array of vulnerable numbers}

\item{CAL_binsmid}{Mid-points of CAL bins}

\item{retL}{Array of retention-at-length}

\item{CAL_ESS}{CAL effective sample size}

\item{CAL_nsamp}{CAL sample size}

\item{Linfarray}{Matrix of Linf}

\item{Karray}{Matrix of K values}

\item{t0array}{Matrix of t0 values}

\item{LenCV}{Vector of LenCV}

\item{truncSD}{Numeric. Number of standard deviations to truncate normal d
distribution}

\item{lenvec}{Vector of lengths}
}
\value{
Generated length composition from \code{genSizeComp}
}
\description{
And other internal related functions
}
\section{Functions}{
\itemize{
\item \code{getfifth}: Internal function to calculate fifth percentile of size composition
}}

\keyword{internal}
