% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS2DLM.R
\name{getBH}
\alias{getBH}
\title{Predict Beverton-Holt recruitment and return fit to S-R observations}
\usage{
getBH(pars, SSB, rec, SSBpR, mode = 1, plot = F)
}
\arguments{
\item{pars}{an initial guess at model parameters steepness and R0}

\item{SSB}{'observations' of spawning biomass}

\item{rec}{'observations' (model predictions) of recruitment}

\item{SSBpR}{spawning stock biomass per recruit at unfished conditions}

\item{mode}{should fit or recruitment deviations be returned}

\item{plot}{should a plot of the model fit be produced?}
}
\description{
Internal function to optBH
}
\author{
T. Carruthers
}
