% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blow_funcs.R
\name{getBlow}
\alias{getBlow}
\title{Blow parallel optimization function}
\usage{
getBlow(x, SSBMSY, MGThorizon, Find, Perr, Marray, hs, Mat_age, Wt_age, R0, V,
  nyears, maxage, mov, Spat_targ, SRrel, aR, bR, Bfrac = 0.5, ploty = F)
}
\arguments{
\item{x}{position in a vector}

\item{SSBMSY}{vector nsim long of spawning biomass at MSY}

\item{MGThorizon}{vector nsim long of MGT x HZN}

\item{Find}{matrix of fishing mortality rate nsim x nyears}

\item{Perr}{matrix of recruitment devitions nsim x nyears + maxage -1}

\item{Marray}{matrix of natural mortality rate nsim x nyears + proyears}

\item{hs}{vector nsim long of steepness values}

\item{Mat_age}{matrix nsim x nages of maturity at age}

\item{Wt_age}{matrix nsim x nages of weight at age}

\item{R0}{vector nsim long of unfished recruitment}

\item{V}{array of vulnerability nsim x maxage x nyears}

\item{nyears}{integer: number of historical years}

\item{maxage}{integer: maximum age}

\item{mov}{array of movement nsim x 2 x 2}

\item{Spat_targ}{vector of spatial targetting parameters}

\item{SRrel}{integer representing recruitmetn dynamics type 1: Bev Holt 2: Ricker}

\item{aR}{vector of recruitment parameters}

\item{bR}{vector of recruitment parameters}

\item{Bfrac}{fraction of SSBMSY that is the target}

\item{ploty}{logical: should a plot be produced}
}
\description{
Find the current biomass at which it would take HZN mean generation times
to reach Bfrac x SSBMSY biomass level given zero catches
}
\author{
T. Carruthers
}

