% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_Internal.R
\name{getEffhist}
\alias{getEffhist}
\title{Calculate historical fishing mortality}
\usage{
getEffhist(Esd, nyears, EffYears, EffLower, EffUpper)
}
\arguments{
\item{Esd}{vector of standard deviation}

\item{nyears}{number of years}

\item{EffYears}{index of years}

\item{EffLower}{vector of lower bound}

\item{EffUpper}{vector of upper bound}
}
\description{
Calculate historical fishing mortality
}
\keyword{internal}
