% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscInt.r
\name{getFMSY}
\alias{getFMSY}
\title{Calculate FMSY and related metrics}
\usage{
getFMSY(x, Marray, hs, Mat_age, Wt_age, R0, V, maxage, nyears, proyears,
  Spat_targ, mov, SRrel, aR, bR)
}
\arguments{
\item{x}{internal parameter}

\item{Marray}{internal parameter}

\item{hs}{internal parameter}

\item{Mat_age}{internal parameter}

\item{Wt_age}{internal parameter}

\item{R0}{internal parameter}

\item{V}{internal parameter}

\item{maxage}{internal parameter}

\item{nyears}{internal parameter}

\item{proyears}{internal parameter}

\item{Spat_targ}{internal parameter}

\item{mov}{internal parameter}

\item{SRrel}{internal parameter}

\item{aR}{internal parameter}

\item{bR}{internal parameter}
}
\description{
Calculate FMSY and related metrics
}
\keyword{internal}

