% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscInt.r
\name{getFref2}
\alias{getFref2}
\title{Internal Get Reference F using Rcpp}
\usage{
getFref2(x, M_ageArray, Wt_age, Mat_age, Perr, N_s, SSN_s, Biomass_s,
  VBiomass_s, SSB_s, Vn, hs, R0a, nyears, proyears, nareas, maxage, mov, SSBpR,
  aR, bR, SRrel, Spat_targ)
}
\arguments{
\item{x}{internal parameter}

\item{M_ageArray}{internal parameter}

\item{Wt_age}{internal parameter}

\item{Mat_age}{internal parameter}

\item{Perr}{internal parameter}

\item{N_s}{internal parameter}

\item{SSN_s}{internal parameter}

\item{Biomass_s}{internal parameter}

\item{VBiomass_s}{internal parameter}

\item{SSB_s}{internal parameter}

\item{Vn}{internal parameter}

\item{hs}{internal parameter}

\item{R0a}{internal parameter}

\item{nyears}{internal parameter}

\item{proyears}{internal parameter}

\item{nareas}{internal parameter}

\item{maxage}{internal parameter}

\item{mov}{internal parameter}

\item{SSBpR}{internal parameter}

\item{aR}{internal parameter}

\item{bR}{internal parameter}

\item{SRrel}{internal parameter}

\item{Spat_targ}{internal parameter}

\item{N_s}{internal parameter}

\item{retAn}{internal parameter}
}
\description{
Internal Get Reference F using Rcpp
}
\keyword{internal}
