% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popdyn.R
\name{getFs}
\alias{getFs}
\title{Internal function to calculate F-at-age given catch and biomass}
\usage{
getFs(x, y, Vuln, CB, Bio, Mage, Fdist, maxage, nareas, nyears)
}
\arguments{
\item{x}{Simulation}

\item{y}{year}

\item{Vuln}{Vulnerabilty}

\item{CB}{Catch biomass}

\item{Bio}{Biomass}

\item{Mage}{M-at-age}

\item{Fdist}{Fishing distribution}

\item{maxage}{Maximum age}

\item{nareas}{Number of areas}

\item{nyears}{Number of historical years}
}
\description{
Internal function to calculate F-at-age given catch and biomass
}
\author{
A. Hordyk
}
\keyword{internal}
