% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS2DLM.R
\name{getGpars}
\alias{getGpars}
\title{Extracts growth parameters from a SS3 r4ss replist}
\usage{
getGpars(replist, seas = 1)
}
\arguments{
\item{replist}{the list output of the r4ss SS_output function (a list of assessment inputs / outputs)}

\item{seas}{The reference season for the growth (not actually sure what this does yet)}
}
\description{
Extracts growth parameters from a SS3 r4ss replist
}
\author{
T. Carruthers
}
