% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMoreData.R
\name{GetMoreData}
\alias{GetMoreData}
\title{Load more data from DLMdata package}
\usage{
GetMoreData(silent = FALSE)
}
\arguments{
\item{silent}{Logical. Should messages to printed?}
}
\description{
Downloads the DLMdata package from GitHub
}
