% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFuns.r
\name{getq2}
\alias{getq2}
\title{Optimization function that find the catchability (q where F=qE) value
required to get to user-specified stock depletion (current biomass /
unfished biomass) modified for Rcpp}
\usage{
getq2(x, dep, Find, Perr, Marray, hs, Mat_age, Wt_age, R0, V, nyears, maxage,
  mov, Spat_targ, SRrel, aR, bR)
}
\arguments{
\item{x}{internal parameter}

\item{dep}{internal parameter}

\item{Find}{internal parameter}

\item{Perr}{internal parameter}

\item{Marray}{internal parameter}

\item{hs}{internal parameter}

\item{Mat_age}{internal parameter}

\item{Wt_age}{internal parameter}

\item{R0}{internal parameter}

\item{V}{internal parameter}

\item{nyears}{internal parameter}

\item{maxage}{internal parameter}

\item{mov}{internal parameter}

\item{Spat_targ}{internal parameter}

\item{SRrel}{internal parameter}

\item{aR}{internal parameter}

\item{bR}{internal parameter

Paired with qopt}
}
\description{
The user specifies the level of stock depleiton. This funciton takes the
derived effort trajectories and finds the catchabiltiy to get the stock
there.
}
\author{
T. Carruthers
}
\keyword{internal}
