% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Output.R
\name{getr}
\alias{getr}
\title{Internal function to estimate r}
\usage{
getr(x, Data, Mvec, Kvec, Linfvec, t0vec, hvec, maxage, r_reps = 100)
}
\arguments{
\item{x}{iteration number}

\item{Data}{Object of class \code{Data}}

\item{Mvec}{M}

\item{Kvec}{K}

\item{Linfvec}{Linf}

\item{t0vec}{t0}

\item{hvec}{steepness}

\item{maxage}{maximum age}

\item{r_reps}{number of reps}
}
\description{
Internal function to estimate r
}
\keyword{internal}
