% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscInt.r
\name{getroot}
\alias{getroot}
\title{Calculate slope from ageM and age95}
\usage{
getroot(X, ageM, age95)
}
\arguments{
\item{X}{index}

\item{ageM}{age at maturity}

\item{age95}{age at 95 percent maturity}
}
\description{
Calculate slope from ageM and age95
}
\keyword{internal}

