% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscInt.r
\name{getSlope1}
\alias{getSlope1}
\title{Internal function to calculate slope}
\usage{
getSlope1(tst, L1, L0.05)
}
\arguments{
\item{tst}{s2}

\item{L1}{length one}

\item{L0.05}{L0.05}
}
\description{
Internal function to calculate slope
}
\keyword{internal}

