% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscInt.r
\name{getSlope2}
\alias{getSlope2}
\title{Internal function to calculate slope}
\usage{
getSlope2(tst, L1, s1, Linf, MaxSel)
}
\arguments{
\item{tst}{s2}

\item{L1}{length one}

\item{s1}{s1}

\item{Linf}{asymptotic length}

\item{MaxSel}{Maximum selectivity}
}
\description{
Internal function to calculate slope
}
\keyword{internal}

