% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscInt.r
\name{getSlopeFun}
\alias{getSlopeFun}
\title{Internal function to calculate slope}
\usage{
getSlopeFun(SD, age50, age95)
}
\arguments{
\item{SD}{standard deviation}

\item{age50}{age at maturity}

\item{age95}{age at 95 percent maturity}
}
\description{
Internal function to calculate slope
}
\keyword{internal}

