% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxomonic.R
\name{getTaxo}
\alias{getTaxo}
\title{Retrieve taxonomic information from online database}
\usage{
getTaxo(Species)
}
\arguments{
\item{Species}{}
}
\value{
A data.frame with the taxonomic information
}
\description{
Retrieve taxonomic information from online database
}
\author{
A. Hordyk
}
