% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{Gulf_blue_tilefish}
\alias{Gulf_blue_tilefish}
\title{Gulf_blue_tilefish Data}
\format{An object of class \code{Data} of length 1.}
\usage{
Gulf_blue_tilefish
}
\description{
An object of class Data
}
\keyword{datasets}

