% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{Overages}
\alias{Overages}
\alias{Perfect_Imp}
\alias{Imp-class-objects}
\title{Imp class objects}
\format{An object of class \code{Imp} of length 1.}
\usage{
Overages

Perfect_Imp
}
\description{
Example objects of class Imp
}
\examples{


avail("Imp")
}
\keyword{datasets}
