% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineClasses.r
\docType{class}
\name{Imp-class}
\alias{Imp-class}
\title{Class \code{'Imp'}}
\description{
An operating model component that specifies the degree of adherence to management recommendations (Implementation error)
}
\section{Slots}{

\describe{
\item{\code{Name}}{The name of the Implementation error object}

\item{\code{TACvar}}{lognormal standard deviation in fraction of TAC taken (uniform distribution)}

\item{\code{TACdif}}{Mean fraction of TAC taken (uniform distribution)}

\item{\code{Evar}}{lognormal standard deviation in fraction of TAE taken(uniform distribution)}

\item{\code{Edif}}{Mean fraction of recommended effort taken (uniform distribution)}

\item{\code{DiscMort}}{Discard mortality rate (uniform distribution)}

\item{\code{SizeLimvar}}{Degree of error in size limit implementation (uniform distribution)}

\item{\code{Source}}{A reference to a website or article form which parameters were taken to define the operating model}
}}

\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new('Imp', OM)}
}

\examples{

showClass('Imp')

}
\author{
T. Carruthers
}
\keyword{classes}
