% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{ImpDescription}
\alias{ImpDescription}
\title{ImpDescription}
\format{
An object of class \code{data.frame} with 7 rows and 2 columns.
}
\usage{
ImpDescription
}
\description{
A data.frame with description of slots for class Imp
}
\keyword{datasets}
