% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{Imprecise_Biased}
\alias{Imprecise_Biased}
\title{Imprecise_Biased Obs}
\format{An object of class \code{Obs} of length 1.}
\usage{
Imprecise_Biased
}
\description{
An object of class Obs
}
\keyword{datasets}
