% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{Imprecise_Unbiased}
\alias{Imprecise_Unbiased}
\title{Imprecise_Unbiased Obs}
\format{An object of class \code{Obs} of length 1.}
\usage{
Imprecise_Unbiased
}
\description{
An object of class Obs
}
\keyword{datasets}
