% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_functions.R
\name{ImputeLH}
\alias{ImputeLH}
\title{Impute correlated life-history parameters}
\usage{
ImputeLH(OM, samp.m = 100, plot = TRUE, ign.bounds = TRUE)
}
\arguments{
\item{OM}{An object of class 'OM'}

\item{samp.m}{Numeric. A multiple of \code{OM@sim}. The number of initial samples to generate
before filtering to ensure predicted values are within ranges specified in OM (e.g., \code{OM@M})}

\item{plot}{Logical. Should the plot be produced?}

\item{ign.bounds}{Logical. Should the OM bounds be ignored if it is not possible to generate
enough predicted values within the specified bounds}
}
\value{
An OM with \code{OM@cpars} populated with \code{OM@nsim} samples of M, K, Linf and L50
}
\description{
Impute correlated life-history parameters by imputing K and L50 values from M and Linf
}
\examples{
testOM<-ImputeLH(testOM)

}
\author{
A. Hordyk
}
