% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{IncE_HDom}
\alias{IncE_HDom}
\title{IncE_HDom Fleet}
\format{An object of class \code{Fleet} of length 1.}
\usage{
IncE_HDom
}
\description{
An object of class Fleet
}
\keyword{datasets}
