% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{IncE_NDom}
\alias{IncE_NDom}
\title{IncE_NDom Fleet}
\format{An object of class \code{Fleet} of length 1.}
\usage{
IncE_NDom
}
\description{
An object of class Fleet
}
\keyword{datasets}
