% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineClasses.r
\docType{class}
\name{InputRec-class}
\alias{InputRec-class}
\title{Class \code{'InputRec'}}
\description{
An object for storing the recommendation for an input control MP
}
\section{Slots}{

\describe{
\item{\code{Effort}}{A numeric value with the effort recommendation as a fraction of current (nyear) fishing effort}

\item{\code{Spatial}}{A boolean vector of length 'nareas' specifying if area is open (1) or closed (0) to fishing}

\item{\code{Allocate}}{A boolean value describing if effort should be re-allocated from close to open areas}

\item{\code{R5}}{smallest length at 5 per cent retention}

\item{\code{LFR}}{smallest length at full retention}

\item{\code{HS}}{upper harvest slot (no retention above this)}

\item{\code{Rmaxlen}}{retention of the largest size class}

\item{\code{Misc}}{An empty list that can be used to store information and pass on to MPs in future}
}}

\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new('InputRec')}
}

\author{
A. Hordyk
}
\keyword{classes}
