% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Output.R
\name{Iratio}
\alias{Iratio}
\title{Mean index ratio MP from Jardim et al. 2015}
\usage{
Iratio(x, Data, reps, yrs = c(2, 5))
}
\arguments{
\item{x}{A position in data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of TAC samples}

\item{yrs}{Vector of length 2 specifying the reference years}
}
\description{
The TAC is adjusted by the ratio alpha, where the numerator 
being the mean index in the most recent two years of the time series and the denominator
being the mean index in the three years prior to those in the numerator.
}
\details{
This MP is the stochastic version of Method 3.2 used by ICES for Data-Limited Stocks (ICES 2012).
}
\references{
Ernesto Jardim, Manuela Azevedo, Nuno M. Brites, Harvest control rules for 
data limited stocks using length-based reference points and survey biomass indices, 
Fisheries Research, Volume 171, November 2015, Pages 12-19, ISSN 0165-7836, 
https://doi.org/10.1016/j.fishres.2014.11.013

ICES. 2012. ICES Implementation of Advice for Data-limited Stocks in 2012 in its 2012
Advice. ICES CM 2012/ACOM 68. 42 pp.
}
\author{
Coded by Q. Huynh. Developed by Jardim et al. (2015)
}
