% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSCAM2DLM.R
\name{iSCAM2Data}
\alias{iSCAM2Data}
\title{Reads data from iSCAM file structure into a DLMtool Data object}
\usage{
iSCAM2Data(iSCAMdir, Name = NULL, Source = "No source provided",
  length_timestep = 1, Author = "No author provided")
}
\arguments{
\item{iSCAMdir}{A folder with iSCAM input and output files in it}

\item{Name}{The name of the operating model}

\item{Source}{Reference to assessment documentation e.g. a url}

\item{length_timestep}{How long is a model time step in years 
(e.g. a quarterly model is 0.25, a monthly model 1/12)}

\item{Author}{Who did the assessment}
}
\description{
A function that uses the file location of a fitted iSCAM 
model including input files to population the various slots of an 
data object. iSCAM2DLM relies on several functions written by Chris 
Grandin (DFO PBS).
}
\author{
T. Carruthers
}
