% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSCAM2DLM.R
\name{iSCAMcomps}
\alias{iSCAMcomps}
\title{Combines all iSCAM age composition data across fleets}
\usage{
iSCAMcomps(replist, Year)
}
\arguments{
\item{replist}{S3 class object: the output from a read from an iSCAM data folder}

\item{Year}{Integer vector: the years of the DLMtool data object ie Data@Year}
}
\description{
iSCAM assessments are often fitted to numerous fleets that have differing 
age selectivities. iSCAMcomps is a simple way of providing the aggregate catch at age
data. It should be noted that this process is important and in a real application would 
require due diligence (ie peer reviewed data workshop).
}
\author{
T. Carruthers
}
