% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSCAM2DLM.R
\name{iSCAMinds}
\alias{iSCAMinds}
\title{Combines indices into a single index using linear modelling}
\usage{
iSCAMinds(idata, Year, fleeteffect = T)
}
\arguments{
\item{idata}{List: the indices recorded in a read from an iSCAM data folder, e.g. replist$data$indices}

\item{Year}{Integer vector: the years of the DLMtool data object ie Data@Year}

\item{fleeteffect}{Logical: should a fleet effect be added to the linear model?}
}
\description{
iSCAM assessments often make use of multiple indices of abundance. 
The DLMtool data object and MPs currently only make use of a single index. 
combiSCAMinds is a function that creates a single index from many using
linear modelling. It is a simple way of providing initial calculations of 
management recommendations and it should be noted that this process 
is important and in a real application would require due diligence (ie
peer reviewed data workshop).
}
\author{
T. Carruthers
}
