% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Output.R
\name{Islope_}
\alias{Islope_}
\title{Index Slope Internal Function}
\usage{
Islope_(x, Data, reps = 100, yrsmth = 5, lambda = 0.4, xx = 0.2)
}
\arguments{
\item{x}{Iteration number}

\item{Data}{An object of class \code{Data}}

\item{reps}{Number of samples of the TAC}

\item{yrsmth}{Years over which to calculate index}

\item{lambda}{A gain parameter controlling the speed in update in TAC.}

\item{xx}{Parameter controlling the fraction of mean catch to start using in
first year}
}
\description{
Index Slope Internal Function
}
\keyword{internal}
