% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Output.R
\name{IT_}
\alias{IT_}
\title{Index Target Internal Function}
\usage{
IT_(x, Data, reps = 100, plot = FALSE, yrsmth = 5, mc = 0.05)
}
\arguments{
\item{x}{Iteration number}

\item{Data}{An object of class \code{Data}}

\item{reps}{Number of replicates}

\item{plot}{Logical. Show the plot?}

\item{yrsmth}{The number of historical years over which to average the index}

\item{mc}{The maximum fractional change in the TAC among years.}
}
\value{
A named list with TAC recommendations of length \code{reps}
}
\description{
Index Target Internal Function
}
\keyword{internal}
