% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Output.R
\name{Itarget_}
\alias{Itarget_}
\title{Incremental Index Target MP Internal Function}
\usage{
Itarget_(x, Data, reps = 100, plot = FALSE, yrsmth = 5, xx = 0, Imulti = 1.5)
}
\arguments{
\item{x}{Iteration number}

\item{Data}{An object of class \code{Data}}

\item{reps}{Number of replicates}

\item{plot}{Logical. Show the plot?}

\item{yrsmth}{Years over which the average index is calculated.}

\item{xx}{Parameter controlling the fraction of mean catch to start using in
first year}

\item{Imulti}{Parameter controlling how much larger target CPUE / index is
compared with recent levels.}
}
\value{
A named list with TAC of length \code{reps}
}
\description{
Incremental Index Target MP Internal Function
}
\keyword{internal}
