% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Input.R
\name{Itargeteff_}
\alias{Itargeteff_}
\title{Incremental Index Target MP Internal Function - Effort MP}
\usage{
Itargeteff_(x, Data, reps = 100, plot = FALSE, yrsmth = 5, xx = 0,
  Imulti = 1.5)
}
\arguments{
\item{x}{Iteration number}

\item{Data}{An object of class \code{Data}}

\item{reps}{Number of replicates}

\item{plot}{Logical. Show the plot?}

\item{yrsmth}{Years over which to smooth recent estimates of surplus
production}

\item{xx}{Parameter controlling the fraction of mean catch to start using in
first year}

\item{Imulti}{Parameter controlling how much larger target CPUE / index is
compared with recent levels.}
}
\value{
A named list with Effort recommendations
}
\description{
Incremental Index Target MP Internal Function - Effort MP
}
\keyword{internal}
