% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_SupportingFunctions.R
\name{iVB}
\alias{iVB}
\title{Inverse von Bertalanffy}
\usage{
iVB(t0, K, Linf, L)
}
\arguments{
\item{t0}{Hypothetical age when length is 0}

\item{K}{Growth coefficient}

\item{Linf}{Asymptotic length}

\item{L}{Length}
}
\description{
Calculate the age given length from the vB equation
}
\keyword{internal}
