% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE_Plotting.R
\name{join_plots}
\alias{join_plots}
\title{Plot several plots with a shared legend}
\usage{
join_plots(plots, ncol = length(plots), nrow = 1, position = c("bottom",
  "right"))
}
\arguments{
\item{plots}{list of plots}

\item{ncol}{Optional number of columns}

\item{nrow}{Optional number of rows}

\item{position}{position of the legend ("bottom" or "right")}
}
\description{
Plot several plots with a shared legend
}
\note{
modified from https://github.com/tidyverse/ggplot2/wiki/share-a-legend-between-two-ggplot2-graphs
}
\keyword{internal}
