% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Functions.R
\name{joinData}
\alias{joinData}
\title{Join Data objects present in a list}
\usage{
joinData(DataList)
}
\arguments{
\item{DataList}{A list of data objects of identical dimension (except for simulation)}
}
\description{
A function that combined a list of data objects into a single data object (same dimensions but can have different numbers of simulations)
}
\author{
T. Carruthers
}
