% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE_functions.R
\name{joinMSE}
\alias{joinMSE}
\title{Join multiple MSE objects together}
\usage{
joinMSE(MSEobjs = NULL)
}
\arguments{
\item{MSEobjs}{A list of MSE objects. Must all have identical operating
model and MPs. MPs which don't appear in all MSE objects will be dropped.}
}
\value{
An object of class \code{MSE}
}
\description{
Joins two or more MSE objects together. MSE objects must have identical
number of historical years, and projection years.
}
\examples{
\dontrun{
OM1 <- DLMtool::testOM
MSE1 <- runMSE(OM1) 
OM2 <- OM1 
OM2@seed <- OM1@seed + 1
MSE2 <- runMSE(OM2)
MSE <- joinMSE(list(MSE1, MSE2))
MSE@nsim
}
}
\author{
A. Hordyk
}
