% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{LAJ_GOM}
\alias{LAJ_GOM}
\alias{LAJ_GOM}
\title{LAJ_GOM OM}
\format{An object of class \code{OM} of length 1.}
\usage{
LAJ_GOM

LAJ_GOM
}
\description{
An object of class OM

An object of class OM
}
\keyword{datasets}
