% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LBSPR_MPs.r
\name{LBSPR}
\alias{LBSPR}
\title{Apply the Length-based SPR model to DLMtool Data Object}
\usage{
LBSPR(x, DLM_data, yrsmth = 1, reps = 1, lstyrs = 10)
}
\arguments{
\item{x}{Simulation number}

\item{DLM_data}{DLM_data object}

\item{yrsmth}{Number of years to smooth length data - not currently used}

\item{reps}{Number of repetitions}

\item{lstyrs}{Last number of years to run model}
}
\description{
Apply the Length-based SPR model to DLMtool Data Object
}
\author{
A. Hordyk
}
