% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Input.R
\name{LBSPR2}
\alias{LBSPR2}
\title{Length-Based SPR MPs}
\usage{
LBSPR2(
  x,
  Data,
  reps = 1,
  plot = FALSE,
  SPRtarg = 0.4,
  theta1 = 0.3,
  theta2 = 0.05,
  maxchange = 0.3,
  n = 5,
  smoother = TRUE,
  R = 0.2
)
}
\arguments{
\item{x}{A position in the data object}

\item{Data}{A data object}

\item{reps}{The number of stochastic samples of the MP recommendation(s)}

\item{plot}{Logical. Show the plot?}

\item{SPRtarg}{The target SPR}

\item{theta1}{Control parameter for the harvest control rule}

\item{theta2}{Control parameter for the harvest control rule}

\item{maxchange}{Maximum change in effort}

\item{n}{Last number of years to run the model on.}

\item{smoother}{Logical. Should the SPR estimates be smoothed?}

\item{R}{variance of sampling noise for smoother}
}
\value{
An object of class \code{\link[MSEtool]{Rec-class}} with the TAE slot populated
}
\description{
The spawning potential ratio (SPR) is estimated using the LBSPR method
and compared to a target of 0.4.
}
\details{
Effort is modified according to the harvest control rules described in
Hordyk et al. (2015b):
}
\section{Required Data}{

See \code{\link[MSEtool]{Data-class}} for information on the \code{Data} object \cr


\code{LBSPR}: CAL, CAL_bins, L50, L95, LHYear, MPeff, Mort, Year, vbK, vbLinf, wlb
}

\section{Rendered Equations}{

See \href{https://dlmtool.openmse.com/reference/LBSPR.html}{Online Documentation} for correctly rendered equations
}

