% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testLBSPR.R
\name{LBSPR2_}
\alias{LBSPR2_}
\title{Internal Estimation Function for LBSPR MP}
\usage{
LBSPR2_(x, Data, reps, n = 5, smoother = TRUE, R = 0.2)
}
\arguments{
\item{x}{Iteration number}

\item{Data}{An object of class \code{Data}}

\item{reps}{Number of samples. Not used in this method.}

\item{n}{Numeric. Number of historical years to run the model.}

\item{smoother}{Logical. Should estimates be smoothed over multiple years?}

\item{R}{variance of sampling noise}
}
\description{
Internal Estimation Function for LBSPR MP
}
\keyword{internal}
