% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Input.R
\name{LBSPR_}
\alias{LBSPR_}
\title{Internal Estimation Function for LBSPR MP}
\usage{
LBSPR_(x, Data, reps, n = 5, smoother = TRUE)
}
\arguments{
\item{x}{Iteration number}

\item{Data}{An object of class \code{Data}}

\item{reps}{Number of samples. Not used in this method.}

\item{n}{Numeric. Number of historical years to run the model.}

\item{smoother}{Logical. Should estimates be smoothed over multiple years?}
}
\description{
Internal Estimation Function for LBSPR MP
}
\references{
Hordyk, A., Ono, K., Valencia, S., loneragan, N., and Prince J;
A novel length-based empirical estimation method of spawning potential ratio (SPR),
and tests of its performance, for small-scale, data-poor fisheries,
ICES Journal of Marine Science, 72 (1) 2015, 217–231,
https://doi.org/10.1093/icesjms/fsu004
}
\keyword{internal}
