\name{LBSPR_ItSel}
\alias{LBSPR_ItSel}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Length-based SPR model with HCR that iteratively adjusts Selectivity}
\description{Management Procedure which adjusts size-at-selection based on estimated SPR. Entirely untested, and included at to demonstrate MPs of this type. }
\usage{
LBSPR_ItSel(x, DLM_data, yrsmth = 1, reps = reps)
}
\arguments{
  \item{x}{
Simulation number
}
  \item{DLM_data}{
DLM_data object
}
  \item{yrsmth}{
Number of years to smooth length data - not currently used
}

  \item{reps}{
Not currently used
}
}

\author{
A. Hordyk
}
