% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LBSPRgen}
\alias{LBSPRgen}
\title{Internal estimation function for LBSPR MP}
\usage{
LBSPRgen(SL50, SL95, FM, nage, nlen, CVLinf, LenBins, LenMids, x, MK, Linf, P,
  L50, L95, Beta)
}
\arguments{
\item{SL50}{Length at 50 percent selectivity}

\item{SL95}{Length at 95 percent selectivity}

\item{FM}{Ratio of apical fishing mortality to natural mortality}

\item{nage}{Number of pseudo age-classes}

\item{nlen}{Number of length bins}

\item{CVLinf}{CV of length-at-age}

\item{LenBins}{Vector of length bins}

\item{LenMids}{Vector of mid-points of length bins}

\item{x}{Vector of psuedo ages}

\item{MK}{Ratio of M/K}

\item{Linf}{Asymptotic length}

\item{P}{Fraction of initial cohort alive at maximum age under unfished conditions}

\item{L50}{Length at 50 percent maturity}

\item{L95}{Length at 95 percent maturity}

\item{Beta}{Exponent of the length-weight relationship}
}
\description{
Internal estimation function for LBSPR MP
}
\author{
A. Hordyk
}
\keyword{internal}
