% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{LHdata}
\alias{LHdata}
\title{LHdata data.frame}
\format{An object of class \code{data.frame} with 84 rows and 6 columns.}
\usage{
LHdata
}
\description{
An object of class data.frame
}
\keyword{datasets}
