% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS2DLM.R
\name{LinInterp}
\alias{LinInterp}
\title{Linear interpolation of a y value at level xlev based on a vector x and y}
\usage{
LinInterp(x, y, xlev, ascending = F, zeroint = F)
}
\arguments{
\item{x}{A vector of x values}

\item{y}{A vector of y values (identical length to x)}

\item{xlev}{A the target level of x from which to guess y}

\item{ascending}{Are the the x values supposed to be ordered before interpolation}

\item{zeroint}{is there a zero-zero x-y intercept?}
}
\description{
Linear interpolation of a y value at level xlev based on a vector x and y
}
\author{
T. Carruthers
}
