% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineClasses.r
\docType{class}
\name{lmmodel-class}
\alias{lmmodel-class}
\title{Class \code{'lmmodel'}}
\description{
An object for storing fitted linear model objects in this case the
relationship between M, age-at-maturity and the von B. K parameter.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new('lmmodel', stock)}. %% ~~ describe objects here ~~
}

\examples{

newdata<-new('lmmodel','Name',new('list'))

}
\author{
T. Carruthers
}
\keyword{classes}
