% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSCAM2DLM.R
\name{load.iscam.files}
\alias{load.iscam.files}
\title{Reads iSCAM files into a hierarchical R list object}
\usage{
load.iscam.files(model.dir, burnin = 1000, thin = 1, verbose = FALSE)
}
\arguments{
\item{model.dir}{An iSCAM directory}

\item{burnin}{The initial mcmc samples to be discarded}

\item{thin}{The degree of chain thinning 1 in every thin 
iterations is kept}

\item{verbose}{Should detailed outputs be provided.}
}
\description{
A function for reading iSCAM input and output files 
into R
}
\author{
Chris Grandin (DFO PBS)
}
