% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{Low_Effort_Non_Target}
\alias{Low_Effort_Non_Target}
\title{Low_Effort_Non_Target Fleet}
\format{An object of class \code{Fleet} of length 1.}
\usage{
Low_Effort_Non_Target
}
\description{
An object of class Fleet
}
\keyword{datasets}
