% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{LS_GOM}
\alias{LS_GOM}
\alias{LS_GOM}
\title{LS_GOM OM}
\format{An object of class \code{OM} of length 1.}
\usage{
LS_GOM

LS_GOM
}
\description{
An object of class OM

An object of class OM
}
\keyword{datasets}
