% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{LS_LS_rec_PR_GOM}
\alias{LS_LS_rec_PR_GOM}
\title{LS_LS_rec_PR_GOM OM}
\format{An object of class \code{OM} of length 1.}
\usage{
LS_LS_rec_PR_GOM
}
\description{
An object of class OM
}
\keyword{datasets}
