% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StochasticSRA.R
\name{LSRA2}
\alias{LSRA2}
\title{Alternative version of LSRA that's a wrapper for LSRA_opt to return the right type of output (mode) using sapply}
\usage{
LSRA2(x, lnR0s, FF, Chist, M, Mat_age, Wt_age, sel, Recdevs, h, mode = 2)
}
\arguments{
\item{x}{a position in the various arrays and vectors that corresponds with a simulation (for use with sapply)}

\item{lnR0s}{a vector nsim long that are estimated R0 values}

\item{FF}{a vector of recent fishign mortality rates (apical Fs)}

\item{Chist}{a vector of historical catch observations \code{[nyears]}}

\item{M}{a vector of natural mortality rates \code{[nsim]}}

\item{Mat_age}{a matrix of maturity at age \code{[nsim x nage]}}

\item{Wt_age}{a matrix of weight at age \code{[nsim x nage]}}

\item{sel}{a matrix of selectivity at age \code{[nsim x nage]}}

\item{Recdevs}{a matrix of recruitment deviations \code{[nsim x nyears]}}

\item{h}{a vector of steepness values of the Bev-Holt Stock-Recruitment relationship}

\item{mode}{optimization or plotting}
}
\value{
all package data objects are placed in the global namespace \code{dir}
}
\description{
Alternative version of LSRA that's a wrapper for LSRA_opt to return the right type of output (mode) using sapply
}
\author{
T. Carruthers
}
