% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Input.R, R/MPs_Output.R
\name{LstepCE1}
\alias{LstepCE1}
\alias{LstepCE2}
\alias{LstepCC1}
\alias{LstepCC4}
\title{A management procedure that incrementally adjusts the TAC or effort according to the
mean length of recent catches.}
\usage{
LstepCE1(x, Data, reps = 100, yrsmth = 5, xx = 0, stepsz = 0.05,
  llim = c(0.96, 0.98, 1.05))

LstepCE2(x, Data, reps = 100, yrsmth = 5, xx = 0, stepsz = 0.1,
  llim = c(0.96, 0.98, 1.05))

LstepCC1(x, Data, reps = 100, yrsmth = 5, xx = 0, stepsz = 0.05,
  llim = c(0.96, 0.98, 1.05))

LstepCC4(x, Data, reps = 100, yrsmth = 5, xx = 0.3, stepsz = 0.05,
  llim = c(0.96, 0.98, 1.05))
}
\arguments{
\item{x}{A position in data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of TAC samples}

\item{yrsmth}{Years over which to smooth recent estimates of surplus
production}

\item{xx}{Parameter controlling the fraction of mean catch to start using in
first year}

\item{stepsz}{Parameter controlling the size of update increment in TAC or effort.}

\item{llim}{A vector of length reference points that determine the
conditions for increasing, maintaining or reducing the TAC or effort.}
}
\value{
A \linkS4class{Rec} object
}
\description{
Four adaptive length-based MPs proposed by Geromont and Butterworth 2014.
Tested by Carruthers et al. 2015.
}
\section{Functions}{
\itemize{
\item \code{LstepCE1}: The least biologically precautionary effort-based MP.

\item \code{LstepCE2}: The most precautionary effort-based MP.

\item \code{LstepCC1}: The least biologically precautionary TAC-based MP.

\item \code{LstepCC4}: The most precautionary TAC-based MP.
}}

\references{
Carruthers et al. 2015. Performance evaluation of simple
management procedures. ICES J. Mar Sci. 73, 464-482.

Geromont, H.F., Butterworth, D.S. 2014. Generic management procedures for
data-poor fisheries; forecasting with few data. ICES J. Mar. Sci.
doi:10.1093/icesjms/fst232
}
\author{
T. Carruthers
}
