% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input.r
\name{LtargetE1}
\alias{LtargetE1}
\title{A management procedure that incrementally adjusts the Effort to reach a
target mean length in catches.}
\usage{
LtargetE1(x, Data, reps = 100, yrsmth = 5, xx = 0, xL = 1.05)
}
\arguments{
\item{x}{A position in data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of samples}

\item{yrsmth}{Years over which to smooth recent estimates of surplus
production}

\item{xx}{Parameter controlling the fraction of mean catch to start using in
first year}

\item{xL}{Parameter controlling the magnitude of the target mean length of
catches relative to average length in catches.}
}
\value{
A numeric vector of input controls
}
\description{
A effort based version of the least biologically precautionary of four
target length MPs proposed by Geromont and Butterworth 2014.
}
\author{
T. Carruthers
}
