% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2OMfunctions.r
\name{LW2OM}
\alias{LW2OM}
\title{Estimate length-weight parameters from data}
\usage{
LW2OM(data = NULL, OM = NULL, plot = TRUE)
}
\arguments{
\item{data}{A data frame with columnns 'Length' and 'Weight' with numeric data}

\item{OM}{An object of class \code{OM}}

\item{plot}{Logical. Show plot of data and fit?}
}
\value{
An object of class \code{OM} with \code{OM@a} and \code{OM@b} slots populated
}
\description{
Function estimates alpha and beta parameter from length-weight data and populates
the relevant slots in the OM
}
