% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_make_update.R
\name{makeData}
\alias{makeData}
\title{Title}
\usage{
makeData(Biomass, CBret, Cret, N, SSB, VBiomass, StockPars, FleetPars,
  ObsPars, ImpPars, RefPoints, ErrList, OM, SampCpars, control = NULL,
  silent = FALSE)
}
\arguments{
\item{CBret}{Simulated retained catch. Numeric array dim(nsim, maxage, nyears, nareas)}

\item{ObsPars}{}

\item{OM}{}

\item{control}{}

\item{reps}{}
}
\description{
Title
}
\keyword{internal}
