% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Mixed.R
\name{makeMeanMP}
\alias{makeMeanMP}
\title{Create an MP that averages the results of multiple MPs}
\usage{
makeMeanMP(MPs)
}
\arguments{
\item{MPs}{A vector of MPs names}
}
\value{
A function of class MP
}
\description{
Create an MP that averages the results of multiple MPs
}
\examples{

MeanMP <- makeMeanMP(c("AvC", "DCAC"))
MSE <- runMSE(DLMtool::testOM, MPs=c("AvC", "DCAC", "MeanMP"))
Tplot2(MSE)


}
