% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSEObj.r
\name{makeTransparent}
\alias{makeTransparent}
\title{Make colors transparent}
\usage{
makeTransparent(someColor, alpha = 100)
}
\arguments{
\item{someColor}{Character string describing color}

\item{alpha}{transparency}
}
\description{
Make colors transparent
}
\author{
T. Carruthers
}
