\name{matsizlim}
\alias{matsizlim}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
A data-limited method in which fishing vulnerability is set according to the maturity curve
}
\description{
A very naive approach to demonstrate the implementation of input controls in the DLM toolkit
}
\usage{
matsizlim(x, DLM)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A position in a data-limited methods object
}
  \item{DLM}{
A data-limited methods object
}
}

\value{
A vector of fractions maxage long
}
\references{
Made-up for this package
}
\author{
T. Carruthers
}
