% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFuns.r
\name{mconv}
\alias{mconv}
\title{Get log normal mean from transformed space mean and standard deviation}
\usage{
mconv(m, sd)
}
\arguments{
\item{m}{mean}

\item{sd}{standard deviation}
}
\value{
numeric
}
\description{
Get log normal mean from transformed space mean and standard deviation
}
\author{
T. Carruthers
}
