% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessmentToOM.R
\name{movdistil}
\alias{movdistil}
\title{Simplified a multi-area transition matrix into the best 2 x 2 representation}
\usage{
movdistil(movtab)
}
\arguments{
\item{replist}{the list output of the r4ss SS_output function (a list of assessment inputs / outputs)}

\item{seas}{The reference season for the growth (not actually sure what this does yet)}
}
\description{
A Function that takes a larger movement matrix, identifies the most parsimonious representation of 2 non-mixed areas and returns the final unfished movement matrix
}
\author{
T. Carruthers
}
