% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFuns.r
\name{MPclass}
\alias{MPclass}
\title{Return class of MP from MP name}
\usage{
MPclass(MPs)
}
\arguments{
\item{MPs}{list or vector of MP names}
}
\description{
Return class of MP from MP name
}
\keyword{internal}
