% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_Exported.R
\name{MPtype}
\alias{MPtype}
\title{Management Procedure Type}
\usage{
MPtype(MPs = NA)
}
\arguments{
\item{MPs}{A list of MP names}
}
\value{
A data.frame with MP names and management type
}
\description{
Management Procedure Type
}
\examples{
MPtype(c("AvC", "curE", "matlenlim", "MRreal", "FMSYref"))

}
\seealso{
\link{Required}
}
