% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_Exported.R
\name{MPurl}
\alias{MPurl}
\title{Get help topic URL}
\usage{
MPurl(
  topic,
  url = "https://dlmtool.github.io/DLMtool/reference/",
  nameonly = FALSE
)
}
\arguments{
\item{topic}{Name of the functions}

\item{url}{URL for the help documentation}

\item{nameonly}{Logical. Help file name only?}
}
\value{
file path to help file
}
\description{
Get help topic URL
}
\keyword{internal}
