\name{MRnoreal}
\alias{MRnoreal}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
An marine reserve in area 1 with no spatial reallocation of fishing effort
}
\description{
A spatial control that prevents fishing in area 1 and does not reallocate this fishing effort to area 2.  
}
\usage{
MRnoreal(x, DLM_data)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A position in data / simulation object DLM
}
  \item{DLM_data}{
A data limited methods data object
}
}
\author{
T. Carruthers
}


