% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Input.R
\name{MRreal}
\alias{MRreal}
\alias{MRnoreal}
\title{Spatial closure and allocation management procedures}
\usage{
MRreal(x, Data, ...)

MRnoreal(x, Data, ...)
}
\arguments{
\item{x}{A position in data / simulation object DLM}

\item{Data}{A data limited methods data object}

\item{...}{Optional additional arguments that are ignored. Note arguments
\code{reps} or \code{...} are required for all input controls}
}
\description{
Management procedures which can close areas to fishing and reallocate
fishing effort spatially.
}
\section{Functions}{
\itemize{
\item \code{MRreal}: A spatial control that prevents fishing in area 1 and reallocates this
fishing effort to area 2 (or over other areas).

\item \code{MRnoreal}: A spatial control that prevents fishing in area 1
(e.g., An marine reserve) and does not reallocate this fishing effort to area 2.
}}

\author{
T. Carruthers
}
