% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{MSEDescription}
\alias{MSEDescription}
\title{MSEDescription}
\format{An object of class \code{data.frame} with 107 rows and 2 columns.}
\usage{
MSEDescription
}
\description{
A data.frame with description of slots for class MSE
}
\keyword{datasets}
