% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popdyn.R
\name{MSYCalcs}
\alias{MSYCalcs}
\title{Internal function to calculate MSY Reference Points}
\usage{
MSYCalcs(logU, M_at_Age, Wt_at_Age, Mat_at_Age, V_at_Age, maxage, R0x,
  SRrelx, hx, opt = 1)
}
\arguments{
\item{logU}{log exploitation rate}

\item{M_at_Age}{Vector of M-at-age}

\item{Wt_at_Age}{Vector of weight-at-age}

\item{Mat_at_Age}{Vector of maturity-at-age}

\item{V_at_Age}{Vector of selectivity-at-age}

\item{maxage}{Maximum age}

\item{R0x}{R0 for this simulation}

\item{SRrelx}{SRR type for this simulation}

\item{hx}{numeric. Steepness value for this simulation}

\item{opt}{Option. 1 = return -Yield, 2= return all MSY calcs}
}
\value{
See \code{opt}
}
\description{
Internal function to calculate MSY Reference Points
}
\keyword{internal}
