% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{mypars}
\alias{mypars}
\title{mypars DLM_general}
\format{An object of class \code{DLM_general} of length 1.}
\usage{
mypars
}
\description{
An object of class DLM_general
}
\keyword{datasets}

