\name{NAor0}
\alias{NAor0}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Is a value NA or zero.
}
\description{
As title
}
\usage{
NAor0(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A numeric value. 
}
}
\value{
A table of nsim rows and 2 columns (depletion, fishing mortality rate)
}
\author{
T. Carruthers
}

