% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Functions.R
\name{Needed}
\alias{Needed}
\title{Data needed to get MPs running}
\usage{
Needed(Data, timelimit=1)
}
\arguments{
\item{Data}{A data-limited methods data object}

\item{timelimit}{The maximum time (seconds) taken to complete 10 reps}
}
\description{
A funtion that lists what data are needed to run
data-limited methods that are currently not able to run given a Data
object
}
\examples{
Needs <- Needed(DLMtool::Cobia)
}
\seealso{
\link{Can} \link{Cant} \link{avail}
}
\author{
T. Carruthers
}
