% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS2DLM.R
\name{negcorlogspace}
\alias{negcorlogspace}
\title{A function that samples multivariate normal (logspace) variables}
\usage{
negcorlogspace(xmu, ymu, xcv = 0.1, nsim, cor = -0.9, ploty = F)
}
\arguments{
\item{xmu}{The mean (normal space) of the first (x) variable}

\item{ymu}{The mean (normal space) of the second (y) variable}

\item{xcv}{The coefficient of variation (normal space, log normal sd) of the x variable}

\item{nsim}{The number of random draws}

\item{cor}{The off-diagonal (symmetrical) correlation among x and y}

\item{ploty}{Whether a plot of the sampled variables should be produced}
}
\description{
A function that samples multivariate normal (logspace) variables
}
\author{
T. Carruthers
}
