% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Reference.R
\name{NFref}
\alias{NFref}
\title{No Fishing Reference MP}
\usage{
NFref(x, Data, reps = 100)
}
\arguments{
\item{x}{A position in a data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of stochastic samples of the quota recommendation}
}
\value{
A TAC of 0.01
}
\description{
A reference MP that sets annual catch to zero (or very close to it). Used
for looking at variability in stock with no fishing.
}
\author{
A. Hordyk
}
