% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{Generic_Obs}
\alias{Generic_Obs}
\alias{Imprecise_Biased}
\alias{Imprecise_Unbiased}
\alias{Perfect_Info}
\alias{Precise_Biased}
\alias{Precise_Unbiased}
\alias{Obs-class-objects}
\title{Obs class objects}
\format{An object of class \code{Obs} of length 1.}
\usage{
Generic_Obs

Imprecise_Biased

Imprecise_Unbiased

Perfect_Info

Precise_Biased

Precise_Unbiased
}
\description{
Example objects of class Obs
}
\examples{
Precise_Unbiased@Name

avail("Obs")
}
\keyword{datasets}
