% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineClasses.r
\docType{class}
\name{Obs-class}
\alias{Obs-class}
\title{Class \code{'Obs'}}
\description{
An operating model component that controls the observation model
}
\section{Slots}{

\describe{
\item{\code{Name}}{The name of the observation model object. Single value. Character string.}

\item{\code{Cobs}}{Log-normal catch observation error expressed as a coefficient of variation. Uniform distribution lower and upper bounds. Non-negative real numbers}

\item{\code{Cbiascv}}{Log-normal coefficient of variation controlling the sampling of bias in catch observations for each simulation. Uniform distribution lower and upper bounds. Non-negative real numbers}

\item{\code{CAA_nsamp}}{Number of catch-at-age observation per time step. Uniform distribution lower and upper bounds. Positive real numbers}

\item{\code{CAA_ESS}}{Effective sample size (independent age draws) of the multinomial catch-at-age observation error model. Uniform distribution lower and upper bounds. Positive integers}

\item{\code{CAL_nsamp}}{Number of catch-at-length observation per time step. Uniform distribution lower and upper bounds. Positive integers}

\item{\code{CAL_ESS}}{Effective sample size (independent length draws) of the multinomial catch-at-length observation error model. Uniform distribution lower and upper bounds. Positive integers}

\item{\code{CALcv}}{Log-normal, CV of length-at-age. Uniform distribution lower and upper bounds. Positive real numbers}

\item{\code{Iobs}}{Observation error in the relative abundance indices expressed as a coefficient of variation. Uniform distribution lower and upper bounds. Positive real numbers}

\item{\code{Icv}}{Log-normal coefficient of variation controlling error in observations of relative abundance index. Uniform distribution lower and upper bounds. Positive real numbers}

\item{\code{Btcv}}{Log-normal coefficient of variation controlling error in observations of current stock biomass among years. Uniform distribution lower and upper bounds. Positive real numbers}

\item{\code{Btbias}}{Uniform-log bounds for sampling persistent bias in current stock biomass. Uniform-log distribution lower and upper bounds. Positive real numbers}

\item{\code{beta}}{A parameter controlling hyperstability/hyperdepletion where values below 1 lead to hyperstability (an index that decreases slower than true abundance) and values above 1 lead to hyperdepletion (an index that decreases more rapidly than true abundance). Uniform distribution lower and upper bounds. Positive real numbers}

\item{\code{LenMcv}}{Log-normal coefficient of variation for sampling persistent bias in length at 50 percent maturity. Uniform distribution lower and upper bounds. Positive real numbers}

\item{\code{Mcv}}{Log-normal coefficient of variation for sampling persistent bias in observed natural mortality rate. Uniform distribution lower and upper bounds. Positive real numbers}

\item{\code{Kcv}}{Log-normal coefficient of variation for sampling persistent bias in observed growth parameter K. Uniform distribution lower and upper bounds. Positive real numbers}

\item{\code{t0cv}}{Log-normal coefficient of variation for sampling persistent bias in observed t0. Uniform distribution lower and upper bounds. Positive real numbers}

\item{\code{Linfcv}}{Log-normal coefficient of variation for sampling persistent bias in observed maximum length. Uniform distribution lower and upper bounds. Positive real numbers}

\item{\code{LFCcv}}{Log-normal coefficient of variation for sampling persistent bias in observed length at first capture. Uniform distribution lower and upper bounds. Positive real numbers}

\item{\code{LFScv}}{Log-normal coefficient of variation for sampling persistent bias in length-at-full selection. Uniform distribution lower and upper bounds. Positive real numbers}

\item{\code{FMSYcv}}{Log-normal coefficient of variation for sampling persistent bias in FMSY. Uniform distribution lower and upper bounds. Positive real numbers}

\item{\code{FMSY_Mcv}}{Log-normal coefficient of variation for sampling persistent bias in FMSY/M. Uniform distribution lower and upper bounds. Positive real numbers}

\item{\code{BMSY_B0cv}}{Log-normal coefficient of variation for sampling persistent bias in BMSY relative to unfished. Uniform distribution lower and upper bounds. Positive real numbers}

\item{\code{Irefcv}}{Log-normal coefficient of variation for sampling persistent bias in relative abundance index at BMSY. Uniform distribution lower and upper bounds. Positive real numbers}

\item{\code{Brefcv}}{Log-normal coefficient of variation for sampling persistent bias in BMSY. Uniform distribution lower and upper bounds. Positive real numbers}

\item{\code{Crefcv}}{Log-normal coefficient of variation for sampling persistent bias in MSY. Uniform distribution lower and upper bounds. Positive real numbers}

\item{\code{Dbiascv}}{Log-normal coefficient of variation for sampling persistent bias in stock depletion. Uniform distribution lower and upper bounds. Positive real numbers}

\item{\code{Dcv}}{Log-normal coefficient of variation controlling error in observations of stock depletion among years. Uniform distribution lower and upper bounds. Positive real numbers}

\item{\code{hcv}}{Log-normal coefficient of variation for sampling persistent bias in steepness. Uniform distribution lower and upper bounds. Positive real numbers}

\item{\code{Reccv}}{Log-normal coefficient of variation for sampling persistent bias in recent recruitment strength. Uniform distribution lower and upper bounds. Positive real numbers}

\item{\code{B0cv}}{Log-normal coefficient of variation for sampling persistent bias in unfished biomass. Uniform distribution lower and upper bounds. Positive real numbers}

\item{\code{rcv}}{Log-normal coefficient of variation for sampling persistent bias in intrinsic rate of increase. Uniform distribution lower and upper bounds. Positive real numbers}

\item{\code{Fcurbiascv}}{Log-normal coefficient of variation for sampling persistent bias in current fishing mortality rate. Uniform distribution lower and upper bounds. Positive real numbers}

\item{\code{Fcurcv}}{Log-normal coefficient of variation controlling error in observations of current fishing mortality rate among years. Uniform distribution lower and upper bounds. Positive real numbers}

\item{\code{maxagecv}}{Log-normal coefficient of variation for sampling persistent bias in observation of maximum age. Uniform distribution lower and upper bounds. Positive real numbers}
}}

\note{
Its questionable whether the hyperstability/hyperdepletion should be
categorised as an observation model characteristic as it is most often
driven by fleet dynamics (and therefore should be in the fleet object). Oh
well its here and you might want to make it hyperstable beta < 1 or
hyperdeplete beta > 1, only.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new('Obs')}
}

\examples{

showClass('Obs')

}
\author{
T. Carruthers and A. Hordyk
}
\keyword{classes}
