% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{ObsDescription}
\alias{ObsDescription}
\title{ObsDescription}
\format{An object of class \code{data.frame} with 29 rows and 2 columns.}
\usage{
ObsDescription
}
\description{
A data.frame with description of slots for class Obs
}
\keyword{datasets}
