% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{OMDescription}
\alias{OMDescription}
\title{OMDescription}
\format{An object of class \code{data.frame} with 107 rows and 2 columns.}
\usage{
OMDescription
}
\description{
A data.frame with description of slots for class OM
}
\keyword{datasets}
