% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_init_doc.R
\name{OMexample}
\alias{OMexample}
\title{Copy example OM XL and OM Documentation}
\usage{
OMexample(dir = getwd())
}
\arguments{
\item{dir}{the file path to copy the files to.}
}
\description{
Copy example OM XL and OM Documentation
}
\examples{
\dontrun{
OMexample()
}
}
