% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMinit_doc.R
\name{OMinit}
\alias{OMinit}
\title{Initialize Operating Model}
\usage{
OMinit(name = "Example", templates = NULL, overwrite = FALSE)
}
\arguments{
\item{name}{The name of the Excel and text file to be created in the working directory (character)}

\item{templates}{An optional named list of existing DLMtool Stock, Fleet, Obs, or Imp objects
to use as templates for the Operating Model.}

\item{overwrite}{Logical. Should files be overwritten if they already exist?}
}
\value{
An xlsx and text file are created in the working directory.
}
\description{
Generates an Excel spreadsheet and a text file in the current working directory for 
specifying and documenting a DLMtool Operating Model.
}
\examples{
\dontrun{
# Create an Excel OM template and text file called 'myOM.xlxs' and 'myOM.txt': 
OMinit('myOM')

# Create an Excel OM template and text file called 'myOM.xlxs' and 'myOM.txt', using
the Stock object 'Herring' as a template: 
OMinit('myOM', templates=list(Stock='Herring'))

# Create an Excel OM template and text file called 'myOM.xlxs' and 'myOM.txt', using
the Stock object 'Herring', and Obs object 'Generic_obs' as templates: 
OMinit('myOM', templates=list(Stock='Herring', Obs='Generic_obs'), overwrite=TRUE)
}

}
\author{
A. Hordyk
}
