% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataObj.r
\name{OneRep}
\alias{OneRep}
\title{Make stochastic variables certain for only one rep}
\usage{
OneRep(DLM_data)
}
\arguments{
\item{DLM_data}{An object of class DLM_data that has been run though TAC()}
}
\description{
As title.
}
\author{
T. Carruthers
}
\keyword{internal}

