% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS2DLM.R
\name{optBH}
\alias{optBH}
\title{Wrapper for estimating stock recruitment parameters from resampled stock-recruitment data}
\usage{
optBH(x, SSB, rec, SSBpR, R0temp, pars, frac = 0.5, plot = F)
}
\arguments{
\item{x}{position (currently redundant)}

\item{SSB}{'observations' of spawning biomass}

\item{rec}{'observations' (model predictions) of recruitment}

\item{SSBpR}{spawning stock biomass per recruit at unfished conditions}

\item{R0temp}{an initial guess at the level of unfished recruitment}

\item{pars}{an initial guess at model parameters steepness and R0}

\item{frac}{the fraction of observations for resampling}

\item{plot}{should a plot of model fit be produced?}
}
\description{
Wrapper for estimating stock recruitment parameters from resampled stock-recruitment data
}
\author{
T. Carruthers
}
