% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popdyn.R
\name{optF}
\alias{optF}
\title{Internal function to optimize for F}
\usage{
optF(fapic, vuln, catch, bio, mort, fdist, maxage, nareas)
}
\arguments{
\item{fapic}{Apical fishing mortality}

\item{vuln}{Vulnerability}

\item{catch}{Catch}

\item{bio}{Biomass}

\item{mort}{Natural mortality}

\item{fdist}{Fishing distribution}

\item{maxage}{Maximum age}

\item{nareas}{Number of areas}
}
\description{
Internal function to optimize for F
}
\author{
A. Hordyk
}
