% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popdyn.R
\name{optMSY}
\alias{optMSY}
\title{Optimize yield for a single simulation}
\usage{
optMSY(
  logFa,
  Asize_c,
  nareas,
  maxage,
  Ncurr,
  pyears,
  M_age,
  MatAge,
  WtAge,
  Vuln,
  Retc,
  Prec,
  movc,
  SRrelc,
  Effind,
  Spat_targc,
  hc,
  R0c,
  SSBpRc,
  aRc,
  bRc,
  Qc,
  MPA,
  maxF,
  SSB0c,
  plusgroup = 0
)
}
\arguments{
\item{logFa}{log apical fishing mortality}

\item{Asize_c}{A vector of length areas with relative size of areas}

\item{nareas}{Number of area}

\item{maxage}{Maximum age}

\item{Ncurr}{Current N-at-age}

\item{pyears}{Number of projection years}

\item{M_age}{M-at-age}

\item{MatAge}{Maturity-at-age}

\item{WtAge}{Weight-at-age}

\item{Vuln}{Vulnerablity-at-age}

\item{Retc}{Retention-at-age}

\item{Prec}{Recruitment error}

\item{movc}{Movement matrix}

\item{SRrelc}{SR Relationship}

\item{Effind}{Historical effort}

\item{Spat_targc}{Spatial targeting}

\item{hc}{Steepness}

\item{R0c}{Unfished recruitment by area}

\item{SSBpRc}{Unfished spawning stock per recruit by area}

\item{aRc}{Ricker aR}

\item{bRc}{Ricker bR}

\item{Qc}{Catchability}

\item{MPA}{A matrix of spatial closures by year}

\item{maxF}{A numeric value specifying the maximum fishing mortality for any single age class}

\item{SSB0c}{SSB0}

\item{plusgroup}{Integer. Default = 0 = no plus-group. Use 1 to include a plus-group}
}
\description{
Optimize yield for a single simulation
}
\author{
A. Hordyk
}
\keyword{internal}
