% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popdyn.R
\name{optMSY_eq}
\alias{optMSY_eq}
\title{Internal wrapper function to calculate MSY reference points}
\usage{
optMSY_eq(x, M_ageArray, Wt_age, Mat_age, V, maxage, R0_annual, SRrel, hs,
  yr.ind = 1, nts, tyears)
}
\arguments{
\item{x}{Simulation number}

\item{M_ageArray}{Array of M-at-age}

\item{Wt_age}{Array of weight-at-age}

\item{Mat_age}{Array of maturity-at-age}

\item{V}{Array of selectivity-at-age}

\item{maxage}{Vector of maximum age}

\item{SRrel}{SRR type}

\item{hs}{Vector of steepness}

\item{yr.ind}{Year index used in calculations}

\item{nts}{Number of sub-year time-steps}

\item{R0}{Vector of R0s}
}
\value{
Results from \code{MSYCalcs}
}
\description{
Internal wrapper function to calculate MSY reference points
}
\keyword{internal}
