% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popdyn.R
\name{optQ}
\alias{optQ}
\title{Optimize q for a single simulation}
\usage{
optQ(logQ, depc, SSB0c, nareas, maxage, Ncurr, pyears, M_age, MatAge, WtAge,
  Vuln, Retc, Prec, movc, SRrelc, Effind, Spat_targc, hc, R0c, SSBpRc, aRc, bRc,
  maxF, useCPP)
}
\arguments{
\item{logQ}{log q}

\item{depc}{Depletion value}

\item{SSB0c}{Unfished spawning biomass}

\item{nareas}{Number of areas}

\item{maxage}{Maximum age}

\item{Ncurr}{Current N-at-age}

\item{pyears}{Number of years to project population dynamics}

\item{M_age}{M-at-age}

\item{MatAge}{Maturity-at-age}

\item{WtAge}{Weight-at-age}

\item{Vuln}{Vulnerability-at-age}

\item{Retc}{Retention-at-age}

\item{Prec}{Recruitment error by year}

\item{movc}{movement matrix}

\item{SRrelc}{SR parameter}

\item{Effind}{Historical fishing effort}

\item{Spat_targc}{Spatial targetting}

\item{hc}{Steepness}

\item{R0c}{Unfished recruitment by area}

\item{SSBpRc}{Unfished spawning biomass per recruit by area}

\item{aRc}{Ricker aR}

\item{bRc}{Ricker bR}

\item{maxF}{maximum F}

\item{useCPP}{Logical. Use the CPP code?}
}
\description{
Optimize q for a single simulation
}
\author{
A. Hordyk
}
