% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{optQ_cpp}
\alias{optQ_cpp}
\title{Rcpp version of the q Optimizer}
\usage{
optQ_cpp(lnIn, depc, Fc, Perrc, Mc, hc, Mac, Wac, R0c, Vc, nyears, maxage, movc,
  Spat_targc, SRrelc, aRc, bRc)
}
\arguments{
\item{lnIn}{internal}

\item{Fc}{internal}

\item{Perrc}{internal}

\item{Mc}{internal}

\item{hc}{internal}

\item{Mac}{internal}

\item{Wac}{internal}

\item{R0c}{internal}

\item{Vc}{internal}

\item{nyears}{internal}

\item{maxage}{internal}

\item{movc}{internal}

\item{Spat_targc}{internal}

\item{SRrelc}{internal}

\item{aRc}{internal}

\item{bRc}{internal}

\item{movc}{internal}

\item{SSBpRc}{internal}
}
\description{
Optimize for catchability coefficient
}
\keyword{internal}
