\name{ourMSE}
\alias{ourMSE}
\docType{data}
\title{
Example MSE object used in the vignette
}
\description{
A dummy example MSE object, with customized stock and fleet, all MPs, and 16 simulations.
}
\usage{data("ourMSE")}
\format{
  The format is:
Formal class 'MSE' [package "DLMtool"] with 17 slots
  ..@ Name    : chr "Stock:Snapper  Fleet:Generic_FlatE  Observation model:Imprecise_Biased"
  ..@ nyears  : num 50
  ..@ proyears: num 20
  ..@ nMPs    : int 63
  ..@ MPs     : chr [1:63] "AvC" "BK" "BK_CC" "BK_ML" ...
  ..@ nsim    : num 16
  ..@ OM      :'data.frame':	16 obs. of  34 variables:
  .. ..$ RefY        : num [1:16] 50.2 120.1 50.8 67.8 78.4 ...
  .. ..$ M           : num [1:16] 0.213 0.239 0.217 0.243 0.241 ...
  .. ..$ Depletion   : num [1:16] 0.311 0.257 0.257 0.199 0.13 ...
  .. ..$ A           : num [1:16] 144.4 656.6 191.9 72.2 146.1 ...
  .. ..$ BMSY_B0     : num [1:16] 0.325 0.325 0.325 0.296 0.306 ...
  .. ..$ FMSY_M      : num [1:16] 0.579 0.564 0.556 0.664 0.62 ...
  .. ..$ Mgrad       : num [1:16] -0.1677 -0.0985 -0.1874 0.1456 0.0794 ...
  .. ..$ Msd         : num [1:16] 0.0313 0.0253 0.015 0.0364 0.011 ...
  .. ..$ procsd      : num [1:16] 0.475 0.415 0.331 0.343 0.436 ...
  .. ..$ Esd         : num [1:16] 0.293 0.13 0.372 0.106 0.312 ...
  .. ..$ dFfinal     : num [1:16] -0.009183 -0.007048 -0.006787 -0.003407 -0.000569 ...
  .. ..$ MSY         : num [1:16] 55.2 51.1 39.2 48.7 51.8 ...
  .. ..$ qinc        : num [1:16] -0.686 -0.127 1.558 0.78 0.692 ...
  .. ..$ qcv         : num [1:16] 0.159 0.298 0.187 0.241 0.102 ...
  .. ..$ FMSY        : num [1:16] 0.123 0.135 0.121 0.161 0.149 ...
  .. ..$ Linf        : num [1:16] 91.1 89.7 88.9 91.2 89.9 ...
  .. ..$ K           : num [1:16] 0.198 0.205 0.179 0.199 0.21 ...
  .. ..$ t0          : num [1:16] -0.0367 -0.0296 -0.0333 -0.0322 -0.0319 ...
  .. ..$ hs          : num [1:16] 0.856 0.875 0.856 0.937 0.94 ...
  .. ..$ Linfgrad    : num [1:16] -0.185 0.0209 -0.1721 -0.185 -0.1004 ...
  .. ..$ Kgrad       : num [1:16] -0.0662 0.0667 0.1438 0.0726 0.1547 ...
  .. ..$ Linfsd      : num [1:16] 0.01334 0.01203 0.02384 0.01377 0.00404 ...
  .. ..$ recgrad     : num [1:16] -0.731 2.751 -4.075 -0.234 -1.097 ...
  .. ..$ Ksd         : num [1:16] 0.00443 0.01185 0.01474 0.00921 0.02128 ...
  .. ..$ ageM        : num [1:16] 2.26 1.97 2.25 2.51 2.25 ...
  .. ..$ V26         : num [1:16] 12.5 10.3 10.8 13.7 11.1 ...
  .. ..$ V27         : num [1:16] 32 27 31.8 34.8 30.8 ...
  .. ..$ V28         : num [1:16] 0.798 0.681 0.896 0.687 0.902 ...
  .. ..$ LFC         : num [1:16] 13.4 13.3 10.9 12.1 12.6 ...
  .. ..$ OFLreal     : num [1:16] 51.4 28.4 31.1 40 22.3 ...
  .. ..$ Spat_targ   : num [1:16] 1 1 1 1 1 1 1 1 1 1 ...
  .. ..$ Frac_area_1 : num [1:16] 0.136 0.138 0.145 0.07 0.133 ...
  .. ..$ Prob_staying: num [1:16] 0.884 0.911 0.651 0.427 0.658 ...
  .. ..$ AC          : num [1:16] 0.569 0.717 0.763 0.572 0.8 ...
  ..@ Obs     :'data.frame':	16 obs. of  25 variables:
  .. ..$ Cbias      : num [1:16] 0.659 1.007 0.745 1.866 1.27 ...
  .. ..$ Csd        : num [1:16] 0.202 0.311 0.554 0.575 0.589 ...
  .. ..$ CAA_nsamp  : num [1:16] 88 59 86 71 97 81 87 80 97 99 ...
  .. ..$ CAA_ESS    : num [1:16] 13 19 18 19 13 12 19 15 15 18 ...
  .. ..$ CAL_nsamp  : num [1:16] 77.1 71.2 72.3 77.8 74 ...
  .. ..$ CAL_ESS    : num [1:16] 19 19 16 17 15 19 15 14 14 19 ...
  .. ..$ Isd        : num [1:16] 0.512 0.289 0.439 0.543 0.352 ...
  .. ..$ Dbias      : num [1:16] 0.621 2.023 0.765 1.325 0.648 ...
  .. ..$ Derr       : num [1:16] 0.0663 0.0647 0.0921 0.1788 0.1261 ...
  .. ..$ Mbias      : num [1:16] 1.108 1.769 1.271 0.901 0.397 ...
  .. ..$ FMSY_Mbias : num [1:16] 1.208 0.98 1.189 2.058 0.583 ...
  .. ..$ BMSY_B0bias: num [1:16] 0.897 0.947 0.998 0.942 0.953 ...
  .. ..$ lenMbias   : num [1:16] 1.087 0.898 1.067 0.821 0.842 ...
  .. ..$ LFCbias    : num [1:16] 0.939 0.96 0.896 1.042 0.903 ...
  .. ..$ LFSbias    : num [1:16] 0.977 1.006 0.926 1.252 0.76 ...
  .. ..$ Abias      : num [1:16] 0.331 2.987 0.272 0.231 1.854 ...
  .. ..$ Aerr       : num [1:16] 0.449 0.271 0.262 0.302 0.39 ...
  .. ..$ Kbias      : num [1:16] 0.899 1.166 1.03 1.157 0.841 ...
  .. ..$ t0bias     : num [1:16] 1.019 1 0.982 0.954 1.096 ...
  .. ..$ Linfbias   : num [1:16] 0.925 0.927 0.929 1.038 0.971 ...
  .. ..$ hbias      : num [1:16] 1.021 0.977 1.029 0.983 0.969 ...
  .. ..$ Irefbias   : num [1:16] 1.264 0.718 1.281 0.902 0.952 ...
  .. ..$ Crefbias   : num [1:16] 0.839 0.576 0.921 0.908 1.1 ...
  .. ..$ Brefbias   : num [1:16] 0.655 1.141 1.095 0.725 1.58 ...
  .. ..$ betas      : num [1:16] 0.389 0.521 0.366 1.689 0.357 ...
  ..@ B_BMSY  : num [1:16, 1:63, 1:20] 1.096 0.472 0.956 0.828 0.371 ...
  ..@ F_FMSY  : num [1:16, 1:63, 1:20] 0.447 1.344 0.802 2.214 2.207 ...
  ..@ B       : num [1:16, 1:63, 1:20] 579 223 346 305 144 ...
  ..@ FM      : num [1:16, 1:63, 1:20] 0.0552 0.1816 0.0969 0.3566 0.3297 ...
  ..@ C       : num [1:16, 1:63, 1:20] 29.5 38 33.5 116.3 54 ...
  ..@ TAC     : num [1:16, 1:63, 1:20] 29.5 38 33.5 116.3 54 ...
  ..@ SSB_hist: num [1:16, 1:34, 1:50, 1:2] 0.0125 0.0513 0.0463 0.0164 0.0113 ...
  ..@ CB_hist : num [1:16, 1:34, 1:50, 1:2] 0 0 0 0 0 0 0 0 0 0 ...
  ..@ FM_hist : num [1:16, 1:34, 1:50, 1:2] 0 0 0 0 0 0 0 0 0 0 ...
}

\examples{
data(ourMSE)

}
\keyword{datasets}
