\name{ourMSE2}
\alias{ourMSE2}
\docType{data}
\title{
Example MSE object used in the vignette
}
\description{
A dummy example MSE object, with customized stock and fleet, a subset of MPs, and more simulations.
}
\usage{data("ourMSE2")}
\format{
  The format is:
Formal class 'MSE' [package "DLMtool"] with 17 slots
  ..@ Name    : chr "Stock:Snapper  Fleet:Generic_FlatE  Observation model:Imprecise_Biased"
  ..@ nyears  : num 50
  ..@ proyears: num 20
  ..@ nMPs    : int 1
  ..@ MPs     : chr "Fratio4010"
  ..@ nsim    : num 32
  ..@ OM      :'data.frame':	32 obs. of  34 variables:
  .. ..$ RefY        : num [1:32] 76 94.6 88.9 47.5 67.5 ...
  .. ..$ M           : num [1:32] 0.244 0.214 0.245 0.241 0.247 ...
  .. ..$ Depletion   : num [1:32] 0.126 0.189 0.107 0.133 0.325 ...
  .. ..$ A           : num [1:32] 394 907 625 347 867 ...
  .. ..$ BMSY_B0     : num [1:32] 0.295 0.318 0.313 0.289 0.319 ...
  .. ..$ FMSY_M      : num [1:32] 0.652 0.624 0.611 0.64 0.521 ...
  .. ..$ Mgrad       : num [1:32] 0.1869 0.1968 0.0434 -0.0406 0.0975 ...
  .. ..$ Msd         : num [1:32] 0.029002 0.044358 0.008146 0.000748 0.031698 ...
  .. ..$ procsd      : num [1:32] 0.498 0.24 0.23 0.389 0.334 ...
  .. ..$ Esd         : num [1:32] 0.322 0.234 0.261 0.211 0.149 ...
  .. ..$ dFfinal     : num [1:32] 0.00667 0.00268 0.0093 0.00944 -0.00749 ...
  .. ..$ MSY         : num [1:32] 41.9 42.4 53.2 44.8 48.4 ...
  .. ..$ qinc        : num [1:32] -1.699 -0.165 -1.112 -1.265 -1.282 ...
  .. ..$ qcv         : num [1:32] 0.116 0.168 0.192 0.19 0.271 ...
  .. ..$ FMSY        : num [1:32] 0.159 0.134 0.15 0.154 0.129 ...
  .. ..$ Linf        : num [1:32] 88.5 89.1 91.1 88.3 89.5 ...
  .. ..$ K           : num [1:32] 0.193 0.192 0.201 0.176 0.197 ...
  .. ..$ t0          : num [1:32] -0.0257 -0.0283 -0.0271 -0.0374 -0.0333 ...
  .. ..$ hs          : num [1:32] 0.925 0.867 0.863 0.905 0.869 ...
  .. ..$ Linfgrad    : num [1:32] 0.0302 0.2096 0.1537 0.1475 0.1378 ...
  .. ..$ Kgrad       : num [1:32] -0.1926 -0.2262 0.0211 0.1424 0.0862 ...
  .. ..$ Linfsd      : num [1:32] 0.0183 0.01247 0.01639 0.00315 0.02464 ...
  .. ..$ recgrad     : num [1:32] -3.785 0.839 4.931 2.377 1.731 ...
  .. ..$ Ksd         : num [1:32] 0.01624 0.01957 0.02051 0.0234 0.00604 ...
  .. ..$ ageM        : num [1:32] 2.62 2.64 2.49 3.37 2.76 ...
  .. ..$ V26         : num [1:32] 11.8 7.56 12.94 12.63 10.52 ...
  .. ..$ V27         : num [1:32] 32 29.2 36.9 41.2 29.9 ...
  .. ..$ V28         : num [1:32] 0.544 0.562 0.608 0.652 0.996 ...
  .. ..$ LFC         : num [1:32] 11.8 12.2 13.9 10.6 12.6 ...
  .. ..$ OFLreal     : num [1:32] 19.9 24.7 12.7 11.8 34.5 ...
  .. ..$ Spat_targ   : num [1:32] 1 1 1 1 1 1 1 1 1 1 ...
  .. ..$ Frac_area_1 : num [1:32] 0.1219 0.1372 0.1215 0.0901 0.1494 ...
  .. ..$ Prob_staying: num [1:32] 0.518 0.734 0.615 0.659 0.652 ...
  .. ..$ AC          : num [1:32] 0.553 0.744 0.697 0.917 0.913 ...
  ..@ Obs     :'data.frame':	32 obs. of  25 variables:
  .. ..$ Cbias      : num [1:32] 1.39 0.912 0.698 1.014 0.563 ...
  .. ..$ Csd        : num [1:32] 0.449 0.411 0.278 0.296 0.276 ...
  .. ..$ CAA_nsamp  : num [1:32] 80 79 69 81 74 56 83 70 97 60 ...
  .. ..$ CAA_ESS    : num [1:32] 15 11 16 12 20 15 12 17 17 16 ...
  .. ..$ CAL_nsamp  : num [1:32] 90.3 60.1 69.3 77.7 60.2 ...
  .. ..$ CAL_ESS    : num [1:32] 13 14 17 16 18 19 20 14 16 19 ...
  .. ..$ Isd        : num [1:32] 0.271 0.549 0.219 0.283 0.305 ...
  .. ..$ Dbias      : num [1:32] 0.578 0.648 0.935 0.754 0.198 ...
  .. ..$ Derr       : num [1:32] 0.0542 0.1177 0.0535 0.1328 0.0948 ...
  .. ..$ Mbias      : num [1:32] 0.622 0.899 0.82 0.812 0.736 ...
  .. ..$ FMSY_Mbias : num [1:32] 1.422 1.071 0.939 0.853 0.509 ...
  .. ..$ BMSY_B0bias: num [1:32] 1.108 0.905 0.907 0.897 1.254 ...
  .. ..$ lenMbias   : num [1:32] 1.012 0.936 1.077 1.041 1.206 ...
  .. ..$ LFCbias    : num [1:32] 1.083 0.865 1.066 1.113 1.083 ...
  .. ..$ LFSbias    : num [1:32] 0.963 0.964 0.886 1.091 1.045 ...
  .. ..$ Abias      : num [1:32] 3.474 0.352 0.319 0.794 0.819 ...
  .. ..$ Aerr       : num [1:32] 0.473 0.46 0.487 0.399 0.294 ...
  .. ..$ Kbias      : num [1:32] 0.952 0.969 0.875 0.942 1.14 ...
  .. ..$ t0bias     : num [1:32] 0.988 0.993 1.077 1.117 1.088 ...
  .. ..$ Linfbias   : num [1:32] 0.875 0.955 1.028 0.939 1.048 ...
  .. ..$ hbias      : num [1:32] 1.02 1.03 1.01 1.06 1.02 ...
  .. ..$ Irefbias   : num [1:32] 1.099 0.969 0.984 1.27 0.863 ...
  .. ..$ Crefbias   : num [1:32] 0.731 0.973 0.736 1.457 0.968 ...
  .. ..$ Brefbias   : num [1:32] 0.947 0.312 1.685 1.129 0.401 ...
  .. ..$ betas      : num [1:32] 0.358 2.581 0.403 0.721 0.375 ...
  ..@ B_BMSY  : num [1:32, 1, 1:20] 0.435 0.427 0.185 0.147 0.5 ...
  ..@ F_FMSY  : num [1:32, 1, 1:20] 4.18e-15 0.00 2.85e-04 6.91e-02 8.62e-16 ...
  ..@ B       : num [1:32, 1, 1:20] 138.1 173.5 79.6 52.7 204 ...
  ..@ FM      : num [1:32, 1, 1:20] 6.66e-16 0.00 4.27e-05 1.07e-02 1.11e-16 ...
  ..@ C       : num [1:32, 1, 1:20] 8.30e-14 5.83e-15 3.03e-03 4.94e-01 1.58e-14 ...
  ..@ TAC     : num [1:32, 1, 1:20] 8.30e-14 5.83e-15 3.03e-03 4.94e-01 1.58e-14 ...
  ..@ SSB_hist: num [1:32, 1:34, 1:50, 1:2] 0.01726 0.0058 0.02556 0.00565 0.00385 ...
  ..@ CB_hist : num [1:32, 1:34, 1:50, 1:2] 0 0 0 0 0 0 0 0 0 0 ...
  ..@ FM_hist : num [1:32, 1:34, 1:50, 1:2] 0 0 0 0 0 0 0 0 0 0 ...
}

\examples{
data(ourMSE2)
}
\keyword{datasets}
