% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineClasses.r
\docType{data}
\name{ourReefFish}
\alias{ourReefFish}
\title{Example data object}
\usage{
data('ourReefFish')
}
\description{
Example data object with a number of output control MPs run on it, and
includes resulting distributions of TACs
}
\examples{
\dontrun{ 
data(ourReefFish)
str(ourReefFish)  
plot(ourReefFish) 
}

}
\keyword{datasets}
