% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMobjects.R
\name{P100}
\alias{P100}
\title{Performance Metric: Probability B > BMSY}
\usage{
P100(MSEobj = NULL, Ref = 1)
}
\arguments{
\item{MSEobj}{An object of class MSE}

\item{Ref}{Reference Point for biomass (fraction of BMSY)}
}
\value{
An object of class PMobj
}
\description{
Performance Metric: Probability B > BMSY
}
\examples{
\dontrun{
P100(myMSE)
}
}
