% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMobjects.R
\name{P50}
\alias{P50}
\title{Performance Metric: Probability B > 0.5 BMSY}
\usage{
P50(MSEobj = NULL)
}
\arguments{
\item{MSEobj}{An object of class MSE}
}
\value{
An object of class PMobj
}
\description{
Performance Metric: Probability B > 0.5 BMSY
}
\examples{
\dontrun{
P50(myMSE)
}
}
