% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{Perfect_Imp}
\alias{Perfect_Imp}
\title{Perfect_Imp Imp}
\format{An object of class \code{NULL} of length 0.}
\usage{
Perfect_Imp
}
\description{
An object of class Imp
}
\keyword{datasets}
