% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{Perfect_Info}
\alias{Perfect_Info}
\title{Perfect_Info Obs}
\format{An object of class \code{Obs} of length 1.}
\usage{
Perfect_Info
}
\description{
An object of class Obs
}
\keyword{datasets}
