% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineClasses.r
\docType{methods}
\name{plot,DLM_data,ANY-method}
\alias{plot,DLM_data,ANY-method}
\title{Plot DLM_data object}
\usage{
\S4method{plot}{DLM_data,ANY}(x, funcs = NA, maxlines = 6, perc = 0.5,
  xlims = NA)
}
\arguments{
\item{x}{object of class DLM_data}

\item{funcs}{MPs}

\item{maxlines}{maximum number of lines}

\item{perc}{percentile of TAC recommendation}

\item{xlims}{limits of x-axis}
}
\description{
Plot DLM_data object
}
