% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_Plotting.R
\name{plotFleet}
\alias{plotFleet}
\title{Plot the Fleet object parameters}
\usage{
plotFleet(
  x,
  Stock = NULL,
  nsamp = 3,
  nsim = 500,
  proyears = 28,
  col = "darkgray",
  breaks = 10,
  lwd = 2,
  ...
)
}
\arguments{
\item{x}{An object of class Fleet (or of class OM)}

\item{Stock}{An object of class Stock}

\item{nsamp}{Number of random samples for time-series plots}

\item{nsim}{Number of iterations for histograms}

\item{proyears}{Number of projection years}

\item{col}{Color of histograms}

\item{breaks}{Number of breaks for histograms}

\item{lwd}{line width}

\item{...}{Optional additional arguments passed to \code{plot}}
}
\description{
Plot the Fleet object parameters
}
\author{
A. Hordyk
}
