% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_Plotting.R
\name{plotImp}
\alias{plotImp}
\title{Plot the Implementation object parameters}
\usage{
plotImp(x, nsim = 500, nyears = 50, col = "darkgray", breaks = 10, ...)
}
\arguments{
\item{x}{An object of class Imp (or of class OM)}

\item{nsim}{Number of iterations for histograms}

\item{nyears}{Number of historical years}

\item{col}{Color of histograms}

\item{breaks}{Number of breaks for histograms}

\item{...}{Optional additional arguments passed to \code{plot}}
}
\description{
A function that plots histograms of samples from the implementation object parameters,
and time-series plots of \code{nsamp} samples of time-series examples. Used to
visually examine the parameter values and ranges entered into the Obs object.
}
\author{
T. Carruthers and A. Hordyk
}
