% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineClasses.r
\docType{methods}
\name{plot,MSE,ANY-method}
\alias{plot,MSE,ANY-method}
\title{Plot MSE object}
\usage{
\S4method{plot}{MSE,ANY}(x)
}
\arguments{
\item{x}{object of class MSE}
}
\description{
Plot MSE object
}

