% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOM.R
\name{plot.OM}
\alias{plot.OM}
\title{Plot the operating model (OM) object parameters}
\usage{
\method{plot}{OM}(x, ...)
}
\arguments{
\item{x}{An object of class OM}

\item{...}{Optional additional arguments passed to \code{plot}}
}
\description{
A function that plots the parameters and resulting time series of an operating model.
}
\author{
T. Carruthers
}
