% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Plotting.R
\name{plot.Data}
\alias{plot.Data}
\title{Plot Data object}
\usage{
\method{plot}{Data}(x, upq = 0.9, lwq = 0.1, outline = FALSE, ...)
}
\arguments{
\item{x}{object of class Data}

\item{upq}{Upper quantile of TACs for max ylim}

\item{lwq}{Lower quantile of TACs for min ylim}

\item{outline}{Logical. Include outliers in plot?}

\item{...}{Optional additional arguments passed to \code{boxplot}}
}
\description{
Plot Data object
}
