% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFleet.r
\name{plot.Fleet}
\alias{plot.Fleet}
\title{Plot the Fleet object parameters}
\usage{
plot.Fleet(Fleet, Stock, nsamp = 3, nsim = 500, nyears = 50,
  proyears = 28, col = "darkgray", breaks = 10, lwd = 2)
}
\arguments{
\item{Fleet}{An object of class Fleet (or of class Fleet)}

\item{nsamp}{Number of random samples for time-series plots}

\item{nsim}{Number of iterations for histograms}

\item{nyears}{Number of historical years}

\item{proyears}{Number of projection years}

\item{col}{Color of histograms}

\item{breaks}{Number of breaks for histograms}

\item{lwd}{line width}
}
\description{
Plot the Fleet object parameters
}
\author{
A. Hordyk
}
