% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE_Plotting.R
\name{plot.MSE}
\alias{plot.MSE}
\title{Plot MSE object}
\usage{
\method{plot}{MSE}(x, ...)
}
\arguments{
\item{x}{object of class MSE}

\item{...}{other parameters passed to plot (currently ignored)}
}
\description{
Plot MSE object
}
