% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOM.R
\name{plot.OM}
\alias{plot.OM}
\title{Plot the operating model (OM) object parameters}
\usage{
plot.OM(OM)
}
\arguments{
\item{OM}{An object of class OM}
}
\description{
A function that plots the parameters and resulting time series of an operating model.
}
\author{
T. Carruthers
}
