% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_Plotting.R
\name{plotM}
\alias{plotM}
\title{Plot M-at-Age and Size}
\usage{
plotM(Stock, nsim = 5)
}
\arguments{
\item{Stock}{An object of class 'Stock' or 'OM'}

\item{nsim}{The number of simulations to plot}
}
\description{
Plot M-at-Age and Size
}
\examples{
\dontrun{
plotM(Albacore)
}
}
\author{
A. Hordyk
}
