% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataObj.r
\name{plotOFL}
\alias{plotOFL}
\title{A generic OFL plot for NOAA use}
\usage{
plotOFL(DLM_data,xlims=NA,perc=0.5)
}
\arguments{
\item{DLM_data}{An object of class DLM_data that has been run though TAC()}

\item{xlims}{x axis limits}

\item{perc}{The percentile of the OFL distribution to be plotted}
}
\value{
A table of performance metrics.
}
\description{
As title.
}
\author{
T. Carruthers
}
