% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSelect.R
\name{plotSelect}
\alias{plotSelect}
\title{Plot the vulnerability and retention curves}
\usage{
plotSelect(OM)
}
\arguments{
\item{OM}{An object of class 'OM'}
}
\description{
Plot the vulnerability and retention curves
}
\author{
A. Hordyk
}
