% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerfMetrics.r
\docType{class}
\name{PM-class}
\alias{PM-class}
\title{Class \code{'PM'}}
\description{
A Performance Metrics Function Object
}
\details{
ADD DETAILS HERE
}
\section{Slots}{

\describe{
\item{\code{Name}}{The name of the Performance Metric object}
}}

\section{Create Object}{
 Objects of class PM can be created by calls of the form \code{new('PM')}
}

\examples{

showClass('PM')

}
\keyword{classes}
