% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMtable.R
\name{PMLimit}
\alias{PMLimit}
\title{Create a table of Performance Limits}
\usage{
PMLimit(MSE, ..., Prob = NULL, Labels = NULL, FeaseMPs = NULL,
  out.file = NULL, output_format = "html_document", openFile = TRUE,
  quiet = TRUE, dir = NULL, RMDfile = NULL, font_size = 14,
  auto_width = FALSE, enableSearch = TRUE, PMlist = NULL,
  build = TRUE)
}
\arguments{
\item{MSE}{An object of class 'MSE'}

\item{...}{PM objects to be used as performance limits. Characters (i.e names of PM objects)}

\item{Prob}{Minimum probability threshold}

\item{Labels}{Optional named list specifying new labels for MPs. For example: \code{Labels = list(AvC="Average Catch", CC1="Constant Catch")}}

\item{FeaseMPs}{Optional. Character vector of MP names that are considered feasible. e.g. the output from \code{Fease()}}

\item{out.file}{Name of the output file. If none provided, output file will be named 'PerfLimTable'}

\item{output_format}{Output file format. Currently only 'html_document' is supported}

\item{openFile}{Logical. Should the file be opened in browser?}

\item{quiet}{Logical. An option to suppress printing of the pandoc command line.}

\item{dir}{Optional. Directory for output file. Default is working directory.}

\item{RMDfile}{Optional. RMD template file}

\item{font_size}{Numeric. Font size for text in the table}

\item{auto_width}{Logical. Should table be width be automatic?}

\item{enableSearch}{Currently disabled. Logical. Should search be enabled in the html table?}

\item{PMlist}{Optional. List of PM names.}

\item{build}{Logical. Build the html table?}
}
\value{
Invisibly returns names of MPs that pass all performance limits
}
\description{
Create a table of Performance Limits
}
