\name{PMs}
\alias{PMs}
\docType{data}
\title{
Built-in Performance metrics
}
\description{
A list of built-in performance metrics which can be used in \code{TradePlot} for creating trade-off plots. 
}
\usage{data("PMs")}
\format{
  A data frame with 6 performance metrics:
  \describe{
    \item{\code{PMs}}{a character vector}
  }
}

