% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerfMetrics.r
\name{PMtrade}
\alias{PMtrade}
\title{Calculate Trade-Off Performance Metrics}
\usage{
PMtrade(PM1, PM2, MSEobj, MSEname = NULL, Can = NULL, Class = NULL)
}
\arguments{
\item{PM1}{An object of class 'PM'}

\item{PM2}{An object of class 'PM'}

\item{MSEobj}{An object of class 'MSE' or a list of objects of class 'MSE'}

\item{MSEname}{An optional character string of names for the MSE objects if 'MSEobj' is 
a list}

\item{Can}{An optional vector with the output of the 'Can' function, or a logical vector
indicating which MPs can be run}

\item{Class}{An optional vector indicating the class of MPs.}
}
\value{
An object of class 'PMtrade' used in the trade-off plot function
}
\description{
Function takes two performance metrics objects and returns a data.frame of calculated
performance metrics - used internally by the trade-off plot function
}
