% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMobjects.R
\name{POF}
\alias{POF}
\title{Performance Metric: Probability F < FMSY}
\usage{
POF(MSEobj = NULL, Ref = 1)
}
\arguments{
\item{MSEobj}{An object of class MSE}

\item{Ref}{Reference Point for fishing mortality (fraction of FMSY)}
}
\value{
An object of class PMobj
}
\description{
Performance Metric: Probability F < FMSY
}
\examples{
\dontrun{
POF(myMSE)
}
}
