\name{PorgMSE}
\alias{PorgMSE}
\docType{data}
\title{Example MSE object used in the vignette}
\description{
A dummy example MSE object, based on porgy, generic fleet and imprecise and unbiased observation model, four new MPs, and 20 simulations.
}
\usage{data("PorgMSE")}
\format{
  The format is:
Formal class 'MSE' [package "DLMtool"] with 17 slots
  ..@ Name    : chr "Stock:Porgy  Fleet:Generic_IncE  Observation model:Imprecise_Unbiased"
  ..@ nyears  : num 50
  ..@ proyears: num 20
  ..@ nMPs    : int 4
  ..@ MPs     : chr [1:4] "AvC" "THC" "matlenlim" "area1_50"
  ..@ nsim    : num 20
  ..@ OM      :'data.frame':	20 obs. of  34 variables:
  .. ..$ RefY        : num [1:20] 44.8 101.2 119.9 84.3 126 ...
  .. ..$ M           : num [1:20] 0.242 0.242 0.24 0.22 0.211 ...
  .. ..$ Depletion   : num [1:20] 0.0791 0.37 0.5231 0.087 0.3618 ...
  .. ..$ A           : num [1:20] 53.7 659.3 1074.2 240.2 281.7 ...
  .. ..$ BMSY_B0     : num [1:20] 0.429 0.419 0.421 0.394 0.414 ...
  .. ..$ FMSY_M      : num [1:20] 0.217 0.243 0.194 0.404 0.22 ...
  .. ..$ Mgrad       : num [1:20] 0.0556 0.0192 0.0365 0.08 -0.1671 ...
  .. ..$ Msd         : num [1:20] 0.08302 0.02084 0.03384 0.03472 0.00318 ...
  .. ..$ procsd      : num [1:20] 0.492 0.31 0.401 0.3 0.432 ...
  .. ..$ Esd         : num [1:20] 0.391 0.194 0.194 0.366 0.38 ...
  .. ..$ dFfinal     : num [1:20] 0.0276 0.0166 0.0156 0.0249 0.0181 ...
  .. ..$ MSY         : num [1:20] 58.1 55.8 56.5 54.6 51.9 ...
  .. ..$ qinc        : num [1:20] 0.501 -0.656 1.35 -1.478 1.554 ...
  .. ..$ qcv         : num [1:20] 0.128 0.288 0.163 0.163 0.263 ...
  .. ..$ FMSY        : num [1:20] 0.0526 0.059 0.0465 0.0889 0.0465 ...
  .. ..$ Linf        : num [1:20] 53.5 51 50.2 51.2 51.7 ...
  .. ..$ K           : num [1:20] 0.189 0.209 0.2 0.186 0.19 ...
  .. ..$ t0          : num [1:20] -1.02 -1.01 -1.04 -1.02 -1.01 ...
  .. ..$ hs          : num [1:20] 0.397 0.408 0.383 0.459 0.366 ...
  .. ..$ Linfgrad    : num [1:20] 0.0781 0.1421 0.1478 0.1272 0.0708 ...
  .. ..$ Kgrad       : num [1:20] 0.0737 -0.1356 0.2391 -0.0225 -0.1415 ...
  .. ..$ Linfsd      : num [1:20] 0.00464 0.01446 0.0104 0.00358 0.01344 ...
  .. ..$ recgrad     : num [1:20] -6.94 -6.77 3.21 -8.52 1.53 ...
  .. ..$ Ksd         : num [1:20] 0.00384 0.02322 0.00456 0.01538 0.01945 ...
  .. ..$ ageM        : num [1:20] 1.47 2.59 3.16 2.75 3.15 ...
  .. ..$ V26         : num [1:20] 4.42 7.94 6.13 6 10.87 ...
  .. ..$ V27         : num [1:20] 21 20.3 24 23.1 24.7 ...
  .. ..$ V28         : num [1:20] 0.855 0.536 0.927 0.107 0.626 ...
  .. ..$ LFC         : num [1:20] 12.4 12.1 12.3 12.3 13.6 ...
  .. ..$ OFLreal     : num [1:20] 9.24 40.85 44.74 11.03 37.62 ...
  .. ..$ Spat_targ   : num [1:20] 1 1 1 1 1 1 1 1 1 1 ...
  .. ..$ Frac_area_1 : num [1:20] 0.1034 0.1014 0.1038 0.0979 0.1012 ...
  .. ..$ Prob_staying: num [1:20] 0.976 0.976 0.937 0.986 0.926 ...
  .. ..$ AC          : num [1:20] 0.322 0.614 0.761 0.478 0.31 ...
  ..@ Obs     :'data.frame':	20 obs. of  25 variables:
  .. ..$ Cbias      : num [1:20] 1.06 1.01 1 1.04 1.01 ...
  .. ..$ Csd        : num [1:20] 0.55 0.399 0.288 0.406 0.373 ...
  .. ..$ CAA_nsamp  : num [1:20] 92 53 83 70 76 68 71 95 79 76 ...
  .. ..$ CAA_ESS    : num [1:20] 14 19 11 19 14 19 20 13 18 16 ...
  .. ..$ CAL_nsamp  : num [1:20] 90 66.6 86.8 58.2 65.1 ...
  .. ..$ CAL_ESS    : num [1:20] 15 17 16 13 12 15 13 18 13 17 ...
  .. ..$ Isd        : num [1:20] 0.439 0.563 0.571 0.21 0.512 ...
  .. ..$ Dbias      : num [1:20] 1.038 0.936 0.77 0.873 0.846 ...
  .. ..$ Derr       : num [1:20] 0.1293 0.1501 0.0756 0.1422 0.0647 ...
  .. ..$ Mbias      : num [1:20] 0.935 1.048 0.986 1.009 1.028 ...
  .. ..$ FMSY_Mbias : num [1:20] 1.121 0.768 0.881 1.063 0.862 ...
  .. ..$ BMSY_B0bias: num [1:20] 1.056 0.978 0.781 1.111 0.819 ...
  .. ..$ lenMbias   : num [1:20] 1.041 0.939 1.029 1.019 1.02 ...
  .. ..$ LFCbias    : num [1:20] 0.898 1.007 0.903 0.989 0.973 ...
  .. ..$ LFSbias    : num [1:20] 0.971 1.013 1.059 1.027 1.034 ...
  .. ..$ Abias      : num [1:20] 1.045 1.079 0.525 0.826 2.767 ...
  .. ..$ Aerr       : num [1:20] 0.423 0.211 0.499 0.493 0.299 ...
  .. ..$ Kbias      : num [1:20] 1.056 0.979 0.939 1.01 1.019 ...
  .. ..$ t0bias     : num [1:20] 0.994 0.943 0.985 0.993 0.995 ...
  .. ..$ Linfbias   : num [1:20] 0.93 1.04 1.076 0.944 0.994 ...
  .. ..$ hbias      : num [1:20] 0.961 0.925 1.028 1.054 1.002 ...
  .. ..$ Irefbias   : num [1:20] 1.117 0.979 0.869 1.027 1.146 ...
  .. ..$ Crefbias   : num [1:20] 0.969 1.137 0.873 1.035 0.941 ...
  .. ..$ Brefbias   : num [1:20] 1.015 1.015 0.871 1.023 1.004 ...
  .. ..$ betas      : num [1:20] 0.967 0.815 0.813 1.357 0.74 ...
  ..@ B_BMSY  : num [1:20, 1:4, 1:20] 0.113 0.644 0.516 0.171 0.581 ...
  ..@ F_FMSY  : num [1:20, 1:4, 1:20] 8.68 1.68 1.62 5.01 1.99 ...
  ..@ B       : num [1:20, 1:4, 1:20] 135 783 665 190 804 ...
  ..@ FM      : num [1:20, 1:4, 1:20] 0.4565 0.0988 0.0752 0.4454 0.0923 ...
  ..@ C       : num [1:20, 1:4, 1:20] 73.9 65.2 50.1 61.6 64.1 ...
  ..@ TAC     : num [1:20, 1:4, 1:20] 73.9 65.2 50.1 61.6 64.1 ...
  ..@ SSB_hist: num [1:20, 1:34, 1:50, 1:2] 2.37 2.28 1.18 1.48 1.64 ...
  ..@ CB_hist : num [1:20, 1:34, 1:50, 1:2] 0 0 0 0 0 0 0 0 0 0 ...
  ..@ FM_hist : num [1:20, 1:34, 1:50, 1:2] 0 0 0 0 0 0 0 0 0 0 ...
}

\examples{
data(PorgMSE)

}
\keyword{datasets}
