\name{PorgMSE2}
\alias{PorgMSE2}
\docType{data}
\title{
Example MSE object used in the vignette
}
\description{
A dummy example MSE object, based on porgy, generic fleet and imprecise and unbiased observation model, four new MPs, and 20 simulations.
}
\usage{data("PorgMSE2")}
\format{
  The format is:
Formal class 'MSE' [package "DLMtool"] with 17 slots
  ..@ Name    : chr "Stock:Porgy  Fleet:Generic_IncE  Observation model:Imprecise_Unbiased"
  ..@ nyears  : num 50
  ..@ proyears: num 20
  ..@ nMPs    : int 4
  ..@ MPs     : chr [1:4] "AvC" "THC" "matlenlim" "area1_50"
  ..@ nsim    : num 20
  ..@ OM      :'data.frame':	20 obs. of  34 variables:
  .. ..$ RefY        : num [1:20] 95.66 42.58 59.75 123.37 8.89 ...
  .. ..$ M           : num [1:20] 0.206 0.205 0.227 0.232 0.218 ...
  .. ..$ Depletion   : num [1:20] 0.21 0.115 0.286 0.309 0.071 ...
  .. ..$ A           : num [1:20] 112.5 232.2 549.3 292.9 97.2 ...
  .. ..$ BMSY_B0     : num [1:20] 0.417 0.424 0.421 0.41 0.411 ...
  .. ..$ FMSY_M      : num [1:20] 0.222 0.228 0.23 0.299 0.216 ...
  .. ..$ Mgrad       : num [1:20] 0.193 -0.091 -0.175 -0.236 0.115 ...
  .. ..$ Msd         : num [1:20] 0.039 0.0503 0.0683 0.0777 0.0555 ...
  .. ..$ procsd      : num [1:20] 0.21 0.306 0.534 0.569 0.31 ...
  .. ..$ Esd         : num [1:20] 0.244 0.221 0.118 0.165 0.357 ...
  .. ..$ dFfinal     : num [1:20] 0.01516 0.01958 0.00852 0.00811 0.02392 ...
  .. ..$ MSY         : num [1:20] 66 60.6 60.4 55.6 54.5 ...
  .. ..$ qinc        : num [1:20] 1.973 -0.36 0.251 0.995 1.363 ...
  .. ..$ qcv         : num [1:20] 0.266 0.254 0.173 0.165 0.103 ...
  .. ..$ FMSY        : num [1:20] 0.0458 0.0467 0.0523 0.0695 0.0471 ...
  .. ..$ Linf        : num [1:20] 52.4 50.2 52.2 50.9 52.4 ...
  .. ..$ K           : num [1:20] 0.219 0.215 0.196 0.188 0.214 ...
  .. ..$ t0          : num [1:20] -1.02 -1.04 -1.03 -1 -1.01 ...
  .. ..$ hs          : num [1:20] 0.369 0.365 0.384 0.413 0.361 ...
  .. ..$ Linfgrad    : num [1:20] 0.0874 -0.0825 -0.0851 0.1722 -0.2189 ...
  .. ..$ Kgrad       : num [1:20] 0.11521 0.00481 0.2199 0.05355 -0.1189 ...
  .. ..$ Linfsd      : num [1:20] 0.008613 0.000114 0.016192 0.014209 0.002195 ...
  .. ..$ recgrad     : num [1:20] -7.337 -2.032 0.445 3.315 -7.648 ...
  .. ..$ Ksd         : num [1:20] 0.00127 0.01015 0.01971 0.0106 0.00485 ...
  .. ..$ ageM        : num [1:20] 2.2 1.81 2.18 2.95 2.6 ...
  .. ..$ V26         : num [1:20] 10 5.56 5.14 7.06 7.46 ...
  .. ..$ V27         : num [1:20] 22.3 24.8 19.3 27.5 22.6 ...
  .. ..$ V28         : num [1:20] 0.992 0.677 0.51 0.215 0.773 ...
  .. ..$ LFC         : num [1:20] 13.1 12.7 11.8 10.8 11.7 ...
  .. ..$ OFLreal     : num [1:20] 36.4 16.3 28.4 22.2 14.3 ...
  .. ..$ Spat_targ   : num [1:20] 1 1 1 1 1 1 1 1 1 1 ...
  .. ..$ Frac_area_1 : num [1:20] 0.098 0.0967 0.0955 0.0996 0.1006 ...
  .. ..$ Prob_staying: num [1:20] 0.988 0.952 0.936 0.928 0.99 ...
  .. ..$ AC          : num [1:20] 0.681 0.337 0.197 0.696 0.675 ...
  ..@ Obs     :'data.frame':	20 obs. of  25 variables:
  .. ..$ Cbias      : num [1:20] 0.999 1.016 0.98 1.028 0.993 ...
  .. ..$ Csd        : num [1:20] 0.441 0.274 0.284 0.441 0.421 ...
  .. ..$ CAA_nsamp  : num [1:20] 51 97 74 97 91 75 87 70 59 90 ...
  .. ..$ CAA_ESS    : num [1:20] 12 14 12 16 16 17 13 18 16 15 ...
  .. ..$ CAL_nsamp  : num [1:20] 86.1 92.4 77.8 91.2 89.8 ...
  .. ..$ CAL_ESS    : num [1:20] 18 15 15 12 16 16 19 18 12 16 ...
  .. ..$ Isd        : num [1:20] 0.256 0.408 0.324 0.586 0.387 ...
  .. ..$ Dbias      : num [1:20] 0.893 1.146 0.81 1.129 0.803 ...
  .. ..$ Derr       : num [1:20] 0.1669 0.1431 0.1022 0.0538 0.1462 ...
  .. ..$ Mbias      : num [1:20] 1.1 1.03 1.01 1.02 0.99 ...
  .. ..$ FMSY_Mbias : num [1:20] 0.908 0.744 0.785 1.171 0.472 ...
  .. ..$ BMSY_B0bias: num [1:20] 1.116 0.889 0.994 0.974 0.876 ...
  .. ..$ lenMbias   : num [1:20] 1.073 1.036 1.007 0.946 0.962 ...
  .. ..$ LFCbias    : num [1:20] 0.926 0.923 1.024 0.997 0.962 ...
  .. ..$ LFSbias    : num [1:20] 1.024 1.012 1.008 0.986 0.94 ...
  .. ..$ Abias      : num [1:20] 0.592 1.673 1.248 0.911 0.619 ...
  .. ..$ Aerr       : num [1:20] 0.364 0.265 0.397 0.451 0.491 ...
  .. ..$ Kbias      : num [1:20] 1.03 1.03 1.04 1.08 1.01 ...
  .. ..$ t0bias     : num [1:20] 1.019 0.97 1.079 1.037 0.946 ...
  .. ..$ Linfbias   : num [1:20] 1.105 1.123 0.977 1.042 0.951 ...
  .. ..$ hbias      : num [1:20] 1.042 1.039 0.947 1.045 1.011 ...
  .. ..$ Irefbias   : num [1:20] 1.224 1.022 0.931 0.999 1.148 ...
  .. ..$ Crefbias   : num [1:20] 0.965 0.99 0.879 0.809 1.012 ...
  .. ..$ Brefbias   : num [1:20] 1.016 1.042 1.002 0.846 0.859 ...
  .. ..$ betas      : num [1:20] 1.31 1.43 1.28 1.08 1.05 ...
  ..@ B_BMSY  : num [1:20, 1:4, 1:20] 0.407 0.28 0.414 0.305 0.354 ...
  ..@ F_FMSY  : num [1:20, 1:4, 1:20] 4.01 5.52 1.86 2.59 3.59 ...
  ..@ B       : num [1:20, 1:4, 1:20] 607 436 684 406 475 ...
  ..@ FM      : num [1:20, 1:4, 1:20] 0.1835 0.2576 0.0971 0.1803 0.1688 ...
  ..@ C       : num [1:20, 1:4, 1:20] 121.1 109.2 50.2 49.9 77.1 ...
  ..@ TAC     : num [1:20, 1:4, 1:20] 121.1 109.2 50.2 49.9 77.1 ...
  ..@ SSB_hist: num [1:20, 1:34, 1:50, 1:2] 2.57 2.94 1.82 1.12 2.99 ...
  ..@ CB_hist : num [1:20, 1:34, 1:50, 1:2] 0 0 0 0 0 0 0 0 0 0 ...
  ..@ FM_hist : num [1:20, 1:34, 1:50, 1:2] 0 0 0 0 0 0 0 0 0 0 ...
}

\examples{
data(PorgMSE2)

}
\keyword{datasets}
