\name{Pplot}
\alias{Pplot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
A projection by projection plot of F/FMSY and B/BMSY 
}
\description{
A shorter version of the plot method for MSEs that just shows the projected trends in stock status and over exploitation
}
\usage{
Pplot(MSEobj,nam=NA)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{MSEobj}{
An object of class MSE
}
  \item{nam}{
Name of the plot
}
}
\author{
T. Carruthers
}
