% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{Precise_Biased}
\alias{Precise_Biased}
\title{Precise_Biased Obs}
\format{An object of class \code{Obs} of length 1.}
\usage{
Precise_Biased
}
\description{
An object of class Obs
}
\keyword{datasets}

