% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{Precise_Unbiased}
\alias{Precise_Unbiased}
\title{Precise_Unbiased Observation}
\format{An object of class \code{Observation} of length 1.}
\usage{
Precise_Unbiased
}
\description{
An object of class Observation
}
\keyword{datasets}

