% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{projOpt_cpp}
\alias{projOpt_cpp}
\title{Rcpp version of the Projection Optimizer}
\usage{
projOpt_cpp(lnIn, Mc, hc, Mac, Wac, R0c, Vc, nyears, maxage, movc, Spat_targc,
  SRrelc, aRc, bRc, proyears, Control)
}
\arguments{
\item{lnIn}{internal}

\item{Mc}{internal}

\item{hc}{internal}

\item{Mac}{internal}

\item{Wac}{internal}

\item{R0c}{internal}

\item{Vc}{internal}

\item{nyears}{internal}

\item{maxage}{internal}

\item{movc}{internal}

\item{Spat_targc}{internal}

\item{SRrelc}{internal}

\item{aRc}{internal}

\item{bRc}{internal}

\item{proyears}{internal}

\item{Control}{internal}

\item{movc}{internal}

\item{SSBpRc}{internal}
}
\description{
Optimize for MSY and calculate MSY reference points
}
\keyword{internal}
