% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoxyPMobj.r
\docType{data}
\name{pSBMSY}
\alias{pSBMSY}
\title{pSBMSY Performance Metric Object}
\format{An object of class \code{PM} of length 1.}
\usage{
pSBMSY
}
\description{
An object of class PM
}
\keyword{datasets}
