% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFuns.r
\name{qopt}
\alias{qopt}
\title{Internal optimization function that find the catchability (q where F=qE)
value required to get to user-specified stock depletion (current biomass /
unfished biomass)}
\usage{
qopt(lnq, depc, Fc, Perrc, Mc, hc, Mac, Wac, R0c, Vc, nyears, maxage, movc,
  Spat_targc, SRrelc, aRc, bRc, opt = T)
}
\arguments{
\item{lnq}{internal parameter}

\item{depc}{internal parameter}

\item{Fc}{internal parameter}

\item{Perrc}{internal parameter}

\item{Mc}{internal parameter}

\item{hc}{internal parameter}

\item{Mac}{internal parameter}

\item{Wac}{internal parameter}

\item{R0c}{internal parameter}

\item{Vc}{internal parameter}

\item{nyears}{internal parameter}

\item{maxage}{internal parameter}

\item{movc}{internal parameter}

\item{Spat_targc}{internal parameter}

\item{SRrelc}{internal parameter}

\item{aRc}{internal parameter}

\item{bRc}{internal parameter}

\item{opt}{internal parameter}
}
\description{
The user specifies the level of stock depleiton. This funciton takes the
derived effort trajectories and finds the catchabiltiy to get the stock
there.
}
\author{
T. Carruthers
}
\keyword{internal}

