% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFuns.r
\name{Range}
\alias{Range}
\alias{range01}
\title{Standardize values}
\usage{
Range(x, Max, Min)

range01(x)
}
\arguments{
\item{x}{vector of values}

\item{Max}{Maximum value}

\item{Min}{Minimum value}
}
\description{
Function to standardize to value relative to minimum and maximum values
}
