% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLM_output.R
\docType{data}
\name{Rcontrol2}
\alias{Rcontrol2}
\title{MP using prior for intrinsic rate of increase with a quadratic approximation
to surplus production}
\format{An object of class \code{DLM_output} of length 1.}
\usage{
Rcontrol2(x, DLM_data, reps = 100, yrsmth = 10, gg = 2, glim = c(0.5,
2))
}
\arguments{
\item{x}{A position in data-limited methods data object}

\item{DLM_data}{A data-limited methods data object}

\item{reps}{The number of TAC samples}

\item{yrsmth}{The number of years for smoothing catch and biomass data}

\item{gg}{A gain parameters}

\item{glim}{Limits for the change in TAC among years}
}
\description{
An MP proposed by Carl Walters that modifies quotas according to trends in
apparent surplus production that includes information from a demographically
derived prior for intrinsic rate of increase. This is different from
Rcontrol because it includes a quadratic approximation of recent trend in
surplus production given biomass
}
\author{
C. Walters and T. Carruthers
}
\references{
Made-up for this package.
}
\keyword{datasets}

