% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{RD_RD_Rec_PR_GOM}
\alias{RD_RD_Rec_PR_GOM}
\title{RD_RD_Rec_PR_GOM OM}
\format{An object of class \code{OM} of length 1.}
\usage{
RD_RD_Rec_PR_GOM
}
\description{
An object of class OM
}
\keyword{datasets}
