% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSCAM2DLM.R
\name{read.control.file}
\alias{read.control.file}
\title{Reads iSCAM control file}
\usage{
read.control.file(file = NULL, num.gears = NULL, num.age.gears = NULL,
  verbose = FALSE)
}
\arguments{
\item{file}{File location}

\item{num.gears}{The number of gears}

\item{num.age.gears}{The number age-gears}

\item{verbose}{should detailed results be printed to console}
}
\description{
A function for returning the results of the
iscam control file
}
\author{
Chris Grandin (DFO PBS)
}
