% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSCAM2DLM.R
\name{read.data.file}
\alias{read.data.file}
\title{Reads iSCAM dat file}
\usage{
read.data.file(file = NULL, verbose = FALSE)
}
\arguments{
\item{file}{File location}

\item{verbose}{should detailed results be printed to console}
}
\description{
A function for returning the results of the
.dat iscam file
}
\author{
Chris Grandin (DFO PBS)
}
