% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSCAM2DLM.R
\name{read.mcmc}
\alias{read.mcmc}
\title{Reads iSCAM mcmc output files}
\usage{
read.mcmc(model.dir = NULL, verbose = TRUE)
}
\arguments{
\item{model.dir}{Folder name}

\item{verbose}{should detailed results be printed to console}
}
\description{
A function for returning the results of the
iscam mcmc files
}
\author{
Chris Grandin (DFO PBS)
}
