% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSCAM2DLM.R
\name{read.report.file}
\alias{read.report.file}
\title{Reads iSCAM Rep file}
\usage{
read.report.file(fn)
}
\arguments{
\item{fn}{File location}
}
\description{
A function for returning the results of the
.rep iscam file
}
\author{
Chris Grandin (DFO PBS)
}
