% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fease_Functions.R
\name{RealFease}
\alias{RealFease}
\title{MP feasibility diagnostic using real data}
\usage{
RealFease(Data = NULL)
}
\arguments{
\item{Data}{An object of class 'Data'. Optional. If Data object is included, the returned MPs are both feasible (in terms of management)
and possible (sufficient data to run MP)}
}
\value{
a vector of MP names that calculate without errors for the specific data.
}
\description{
What MPs do not return NAs from the real data
}
\author{
T. Carruthers
}
