% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{Red_Snapper_GOM}
\alias{Red_Snapper_GOM}
\title{Red_Snapper_GOM OM}
\format{An object of class \code{OM} of length 1.}
\usage{
Red_Snapper_GOM
}
\description{
An object of class OM
}
\keyword{datasets}

