% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{Red_snapper_GOM_OM}
\alias{Red_snapper_GOM_OM}
\title{Red_snapper_GOM_OM OM}
\format{An object of class \code{OM} of length 1.}
\usage{
Red_snapper_GOM_OM
}
\description{
An object of class OM (from SS using SS2DLM)
 SEDAR 31, 2014 update assessment. Author: No author provided.
}
\keyword{datasets}

