\name{replic8}
\alias{replic8}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Enlarge (replicate) a DLM data object to create an additional dimension for simulation / sensitivity testing
}
\description{
Replicates position 1 data to multiple positions for sensitivity testing etc 
}
\usage{
replic8(DLM_data, nrep)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{DLM_data}{
A data-limited methods data object
}
  \item{nrep}{
The number of positions to expand the DLM object to
}
}
\author{
T. Carruthers
}
