% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Functions.R
\name{Report}
\alias{Report}
\title{Generate a Data Report}
\usage{
Report(Data = NULL, md = NULL, name = "Data-Report",
  title = "Data Documentation", author = "Author Name",
  date = Sys.Date(), output_format = c("html_document"), open = TRUE,
  dir = NULL, overwrite = FALSE)
}
\arguments{
\item{Data}{Either an object of class \code{Data} or the file path to a valid
file to be imported with \code{XL2Data}}

\item{md}{Full file path to a valid text file documentating the Data}

\item{name}{Optional. Name of the output file}

\item{title}{Title for the Report. Title in the markdown file will override this value}

\item{author}{Author of the Report. Author in the markdown file will override this value}

\item{date}{Date of the Report. Date in the markdown file will override this value}

\item{output_format}{Currently only accepts \code{html_document}}

\item{open}{Logical. Open the report in a web browser?}

\item{dir}{Optional. Directory to save the file. Defaults to \code{getwd()}}

\item{overwrite}{Logical. Overwrite an existing file with the same name?}
}
\value{
Nothing. A Data Report is generated and saved in \code{dir}
}
\description{
A HTML Data Report is generated and opened in a web browser
}
\examples{
\dontrun{
DataInit() # generate Data Input and Documentation files
# enter fishery data in Input file and document in Documentation file
Data <- XL2Data()
Report(Data, "path.to.md.file/Data.md")
}
}
\author{
A. Hordyk
}
