% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{Rougheye_Rockfish_BC}
\alias{Rougheye_Rockfish_BC}
\title{Rougheye_Rockfish_BC OM}
\format{An object of class \code{OM} of length 1.}
\usage{
Rougheye_Rockfish_BC
}
\description{
An object of class OM
}
\keyword{datasets}

