% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{Rougheye_Rockfish_BC_SRA}
\alias{Rougheye_Rockfish_BC_SRA}
\title{Rougheye_Rockfish_BC_SRA OM}
\format{An object of class \code{OM} of length 1.}
\usage{
Rougheye_Rockfish_BC_SRA
}
\description{
An object of class OM (built using StochasticSRA)
 DLMtool generated
}
\keyword{datasets}
