% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFO.R
\name{runCOSEWIC}
\alias{runCOSEWIC}
\title{Dedicated COSEWIC MSE run using the correct MPs and projected time horizon}
\usage{
runCOSEWIC(OM)
}
\arguments{
\item{OM}{An operating model object of class OM}
}
\description{
Projects for 6x maximum age using NFref, FMSYref and curE
}
\author{
T. Carruthers
}
