% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMSEdev.r
\name{runMSEdev}
\alias{runMSEdev}
\title{Run a Management Strategy Evaluation}
\usage{
runMSEdev(OM = "1", MPs = c("AvC", "DCAC", "DD", "Fref", "CurE"),
  interval = 4, pstar = 0.5, maxF = 0.8, timelimit = 1, reps = 1,
  CheckMPs = FALSE, Hist = FALSE, ntrials = 50, fracD = 0.05)
}
\arguments{
\item{OM}{An operating model object (class OM)}

\item{MPs}{A vector of methods (character string) of class DLM_output or
DLM_input.}

\item{interval}{The assessment interval - how often would you like to update
the management system?}

\item{pstar}{The percentile of the sample of the management recommendation
for each method}

\item{maxF}{Maximum instantaneous fishing mortality rate that may be
simulated for any given age class}

\item{timelimit}{Maximum time taken for a method to carry out 10 reps
(methods are ignored that take longer)}

\item{reps}{Number of samples of the management recommendation for each
method. Note that when this is set to 1, the mean value of the data inputs
is used.}

\item{CheckMPs}{Logical to indicate if Can function should be used to check
if MPs can be run.}

\item{Hist}{Should model stop after historical simulations? Returns a list 
containing all historical data}

\item{ntrials}{Maximum of times depletion and recruitment deviations are 
resampled to optimize for depletion. After this the model stops if more than 
percent of simulations are not close to the required depletion}

\item{fracD}{maximum allowed proportion of simulations where depletion is not 
close to sampled depletion from OM before model stops with error}

\item{nsim}{Number of simulations}

\item{proyears}{Number of projected years}
}
\value{
An object of class MSE
}
\description{
A function that runs a Management Strategy Evaluation (closed-loop
simulation) for a specified operating model
}
\author{
T. Carruthers and A. Hordyk
}

