% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataObj.r
\name{Sam}
\alias{Sam}
\title{Conduct stock assessment}
\usage{
Sam(Data, MPs = NA, reps = 100, perc = 0.5)
}
\arguments{
\item{Data}{A data-limited methods data object}

\item{MPs}{A character vector of methods of DLM quota, DLM space or DLM size}

\item{reps}{The number of samples of quota recommendations by method}

\item{perc}{quantile of the recommendation to use}
}
\description{
A wrapper function that gets the OFL recommendation in cases where a method
of DLM quota has been specified
}
\author{
T. Carruthers
}

