% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_SupportingFunctions.R
\name{sample_steepness2}
\alias{sample_steepness2}
\title{Sample steepness given mean and cv}
\usage{
sample_steepness2(n, mu, cv)
}
\arguments{
\item{n}{number of samples}

\item{mu}{mean h}

\item{cv}{cv of h}
}
\description{
Sample steepness given mean and cv
}
\author{
Q. Huynh
}
