% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleOM.R
\name{SampleBioEcoPars}
\alias{SampleBioEcoPars}
\title{Sample Bio-Economic Parameters}
\usage{
SampleBioEcoPars(BioEco, nsim = NULL, cpars = NULL)
}
\arguments{
\item{BioEco}{An object of class 'BioEco' or class 'OM'}

\item{nsim}{Number of simulations. Ignored if 'BioEco' is class 'OM'}

\item{cpars}{Optional named list of custom parameters. Ignored if 'OM' is class 'OM'}
}
\value{
A named list of sampled Bio-Economic parameters
}
\description{
Sample Bio-Economic Parameters
}
\keyword{internal}
