% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleCpars.R
\name{SampleCpars}
\alias{SampleCpars}
\title{Sample custom pars}
\usage{
SampleCpars(cpars)
}
\arguments{
\item{cpars}{A named list containing custom parameters for the OM}
}
\value{
A named list of sampled custom parameters
}
\description{
Sample custom pars
}
