% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleImpPars.R
\name{SampleImpPars}
\alias{SampleImpPars}
\title{Sample Implementation Error Parameters}
\usage{
SampleImpPars(Imp, nsim = NULL)
}
\arguments{
\item{Imp}{An object of class 'Imp' or class 'OM'}

\item{nsim}{Number of simulations. Ignored if 'Stock' is class 'OM'}
}
\value{
A named list of sampled Implementation Error parameters
}
\description{
Sample Implementation Error Parameters
}
