% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleObsPars.R
\name{SampleObsPars}
\alias{SampleObsPars}
\title{Sample Observation Parameters}
\usage{
SampleObsPars(Obs, nsim = NULL)
}
\arguments{
\item{Obs}{An object of class 'Obs' or class 'OM'}

\item{nsim}{Number of simulations. Ignored if 'Stock' is class 'OM'}
}
\value{
A named list of sampled Observation parameters
}
\description{
Sample Observation Parameters
}
