% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleStockPars.R
\name{SampleStockPars}
\alias{SampleStockPars}
\title{Sample Stock parameters}
\usage{
SampleStockPars(Stock, nsim = NULL, nyears = NULL, proyears = NULL,
  cpars = NULL)
}
\arguments{
\item{Stock}{An object of class 'Stock' or class 'OM'}

\item{nsim}{Number of simulations. Ignored if 'Stock' is class 'OM'}

\item{nyears}{Number of historical years. Ignored if 'Stock' is class 'OM'}

\item{proyears}{Number of projection years. Ignored if 'Stock' is class 'OM'}

\item{cpars}{Optional named list of custom parameters. Ignored if 'Stock' is class 'OM'}
}
\value{
A named list of sampled Stock parameters
}
\description{
Sample Stock parameters
}
\examples{
SampleStockPars(DLMtool::Albacore, 10, 30, 20)
}
