% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{San_Francisco_Bay_herring_San_Francisco_Bay_herring_fleet_California}
\alias{San_Francisco_Bay_herring_San_Francisco_Bay_herring_fleet_California}
\title{San_Francisco_Bay_herring_San_Francisco_Bay_herring_fleet_California OM}
\format{An object of class \code{OM} of length 1.}
\usage{
San_Francisco_Bay_herring_San_Francisco_Bay_herring_fleet_California
}
\description{
An object of class OM
}
\keyword{datasets}
