% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoxyPMobj.r
\docType{data}
\name{SB_SB0}
\alias{SB_SB0}
\title{SB_SB0 Performance Metric Object}
\format{An object of class \code{PM} of length 1.}
\usage{
SB_SB0
}
\description{
An object of class PM
}
\keyword{datasets}
