% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoxyPMobj.r
\docType{data}
\name{SB_SBMSY}
\alias{SB_SBMSY}
\title{SB_SBMSY Performance Metric Object}
\format{An object of class \code{PM} of length 1.}
\usage{
SB_SBMSY
}
\description{
An object of class PM
}
\keyword{datasets}
