% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_Exported.R
\name{sdconv}
\alias{sdconv}
\title{Get log normal standard deviation from transformed space mean and standard deviation}
\usage{
sdconv(m, sd)
}
\arguments{
\item{m}{mean}

\item{sd}{standard deviation}
}
\value{
numeric
}
\description{
Get log normal standard deviation from transformed space mean and standard deviation
}
\author{
T. Carruthers
}
