% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscInt.r
\name{SelectFun}
\alias{SelectFun}
\title{Selectivity at length function}
\usage{
SelectFun(i, SL0.05, SL1, MaxSel, maxlens, Lens)
}
\arguments{
\item{i}{index}

\item{SL0.05}{length at 5 percent selection}

\item{SL1}{length at full selection}

\item{MaxSel}{Maximum selectivity}

\item{maxlens}{maximum length}

\item{Lens}{vector of lengths}
}
\description{
Selectivity at length function
}
\keyword{internal}
