% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_Exported.R
\name{setup}
\alias{setup}
\title{Setup parallel processing}
\usage{
setup(cpus = min(parallel::detectCores(), 3), slaveOutfile = "log.txt", ...)
}
\arguments{
\item{cpus}{number of CPUs}

\item{slaveOutfile}{Write R slave output to this file.}

\item{...}{other arguments passed to 'snowfall::sfInit'}
}
\description{
Sets up parallel processing using the snowfall package
}
