% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{SG_GOM}
\alias{SG_GOM}
\alias{SG_GOM}
\title{SG_GOM OM}
\format{An object of class \code{OM} of length 1.}
\usage{
SG_GOM

SG_GOM
}
\description{
An object of class OM

An object of class OM
}
\keyword{datasets}
