% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{SG_SG_ComLL_GOM}
\alias{SG_SG_ComLL_GOM}
\title{SG_SG_ComLL_GOM OM}
\format{An object of class \code{OM} of length 1.}
\usage{
SG_SG_ComLL_GOM
}
\description{
An object of class OM
}
\keyword{datasets}
