\name{SnapMSE}
\alias{SnapMSE}
\docType{data}
\title{
Example MSE object used in the vignette
}
\description{
A dummy example MSE object, with blue shark, generic fleet and imprecise and biased observation model, four MPs, and 16 simulations.
}
\usage{data("SnapMSE")}
\format{
  The format is:
Formal class 'MSE' [package "DLMtool"] with 17 slots
  ..@ Name    : chr "Stock:Blue_shark  Fleet:Generic_fleet  Observation model:Imprecise_Biased"
  ..@ nyears  : num 50
  ..@ proyears: num 30
  ..@ nMPs    : int 4
  ..@ MPs     : chr [1:4] "Fratio" "DCAC" "Fdem" "DD"
  ..@ nsim    : num 16
  ..@ OM      :'data.frame':	16 obs. of  34 variables:
  .. ..$ RefY        : num [1:16] 15916 7401 10480 2887 8534 ...
  .. ..$ M           : num [1:16] 0.172 0.175 0.167 0.242 0.185 ...
  .. ..$ Depletion   : num [1:16] 0.397 0.555 0.361 0.338 0.57 ...
  .. ..$ A           : num [1:16] 63144 1191 34716 118 4138 ...
  .. ..$ BMSY_B0     : num [1:16] 0.38 0.346 0.304 0.337 0.349 ...
  .. ..$ FMSY_M      : num [1:16] 0.497 0.577 0.83 0.423 0.478 ...
  .. ..$ Mgrad       : num [1:16] -0.1167 -0.2097 -0.0849 0.197 -0.0967 ...
  .. ..$ Msd         : num [1:16] 0.06411 0.08061 0.09191 0.00586 0.04776 ...
  .. ..$ procsd      : num [1:16] 0.248 0.213 0.152 0.233 0.25 ...
  .. ..$ Esd         : num [1:16] 0.315 0.398 0.214 0.333 0.38 ...
  .. ..$ dFfinal     : num [1:16] 0.00832 -0.00412 0.00492 0.01095 -0.00757 ...
  .. ..$ MSY         : num [1:16] 2573 3845 4674 2884 4262 ...
  .. ..$ qinc        : num [1:16] -0.331 0.525 -1.985 0.71 0.915 ...
  .. ..$ qcv         : num [1:16] 0.181 0.218 0.151 0.251 0.133 ...
  .. ..$ FMSY        : num [1:16] 0.0855 0.1011 0.1387 0.1025 0.0882 ...
  .. ..$ Linf        : num [1:16] 197 196 201 197 201 ...
  .. ..$ K           : num [1:16] 0.226 0.232 0.239 0.238 0.218 ...
  .. ..$ t0          : num [1:16] -1.032 -1.027 -0.971 -1.011 -0.959 ...
  .. ..$ hs          : num [1:16] 0.487 0.647 0.769 0.656 0.638 ...
  .. ..$ Linfgrad    : num [1:16] 0.1212 -0.0666 -0.1746 -0.1365 0.1226 ...
  .. ..$ Kgrad       : num [1:16] 0.0116 0.1432 -0.2061 -0.1219 -0.0717 ...
  .. ..$ Linfsd      : num [1:16] 0.0044 0.01507 0.02219 0.01226 0.00046 ...
  .. ..$ recgrad     : num [1:16] -8.58 1.41 1.21 -1.32 9.24 ...
  .. ..$ Ksd         : num [1:16] 0.0206 0.0245 0.0159 0.0173 0.0198 ...
  .. ..$ ageM        : num [1:16] 4.37 3.81 3.8 4.33 4.12 ...
  .. ..$ V26         : num [1:16] 35.7 46 31.1 33.4 44 ...
  .. ..$ V27         : num [1:16] 147 103 117 108 117 ...
  .. ..$ V28         : num [1:16] 0.164 0.456 0.297 0.797 0.695 ...
  .. ..$ LFC         : num [1:16] 55.9 48.9 60 52.2 58 ...
  .. ..$ OFLreal     : num [1:16] 2035 4356 5837 3427 4114 ...
  .. ..$ Spat_targ   : num [1:16] 1 1 1 1 1 1 1 1 1 1 ...
  .. ..$ Frac_area_1 : num [1:16] 0.0952 0.1028 0.0983 0.0997 0.1023 ...
  .. ..$ Prob_staying: num [1:16] 0.85 0.833 0.879 0.878 0.829 ...
  .. ..$ AC          : num [1:16] 0.12 0.705 0.611 0.146 0.639 ...
  ..@ Obs     :'data.frame':	16 obs. of  25 variables:
  .. ..$ Cbias      : num [1:16] 0.604 1.099 0.787 1.049 0.999 ...
  .. ..$ Csd        : num [1:16] 0.375 0.289 0.429 0.512 0.262 ...
  .. ..$ CAA_nsamp  : num [1:16] 89 80 86 65 73 88 97 54 70 57 ...
  .. ..$ CAA_ESS    : num [1:16] 14 13 17 13 15 20 15 12 15 19 ...
  .. ..$ CAL_nsamp  : num [1:16] 77.3 70.6 78.5 96.8 73.9 ...
  .. ..$ CAL_ESS    : num [1:16] 20 11 17 19 16 16 20 14 15 19 ...
  .. ..$ Isd        : num [1:16] 0.274 0.341 0.47 0.528 0.342 ...
  .. ..$ Dbias      : num [1:16] 0.762 1.424 1.64 0.697 0.599 ...
  .. ..$ Derr       : num [1:16] 0.0747 0.0884 0.0918 0.0827 0.0928 ...
  .. ..$ Mbias      : num [1:16] 0.749 0.727 2.711 0.437 0.981 ...
  .. ..$ FMSY_Mbias : num [1:16] 0.862 1.336 0.273 1.268 0.683 ...
  .. ..$ BMSY_B0bias: num [1:16] 0.685 0.941 1.192 1.006 1.446 ...
  .. ..$ lenMbias   : num [1:16] 0.828 0.734 1.033 0.672 0.816 ...
  .. ..$ LFCbias    : num [1:16] 0.922 0.881 1.1 0.993 1.093 ...
  .. ..$ LFSbias    : num [1:16] 0.884 0.958 0.86 1 1.041 ...
  .. ..$ Abias      : num [1:16] 0.538 3.303 3.903 0.969 0.975 ...
  .. ..$ Aerr       : num [1:16] 0.311 0.288 0.393 0.317 0.21 ...
  .. ..$ Kbias      : num [1:16] 1.265 0.905 1.052 0.974 0.899 ...
  .. ..$ t0bias     : num [1:16] 1.043 0.956 1.091 0.994 1.103 ...
  .. ..$ Linfbias   : num [1:16] 0.805 0.975 0.964 0.895 1.037 ...
  .. ..$ hbias      : num [1:16] 0.89 0.945 0.921 1.19 1.007 ...
  .. ..$ Irefbias   : num [1:16] 1.928 0.836 0.845 0.581 0.706 ...
  .. ..$ Crefbias   : num [1:16] 0.876 1.052 0.792 0.553 0.672 ...
  .. ..$ Brefbias   : num [1:16] 0.334 0.458 1.194 1.003 1.656 ...
  .. ..$ betas      : num [1:16] 2.988 2.751 0.453 2.727 2.308 ...
  ..@ B_BMSY  : num [1:16, 1:4, 1:30] 0.638 1.17 1.758 1.359 0.861 ...
  ..@ F_FMSY  : num [1:16, 1:4, 1:30] 0.373 2.854 1.198 0.668 0.83 ...
  ..@ B       : num [1:16, 1:4, 1:30] 34017 67403 97638 44286 49632 ...
  ..@ FM      : num [1:16, 1:4, 1:30] 0.0319 0.2886 0.1661 0.0685 0.0732 ...
  ..@ C       : num [1:16, 1:4, 1:30] 649 15647 11473 2852 3300 ...
  ..@ TAC     : num [1:16, 1:4, 1:30] 649 15647 11473 2852 3300 ...
  ..@ SSB_hist: num [1:16, 1:46, 1:50, 1:2] 3.144 2.76 12.83 19.055 0.222 ...
  ..@ CB_hist : num [1:16, 1:46, 1:50, 1:2] 0 0 0 0 0 0 0 0 0 0 ...
  ..@ FM_hist : num [1:16, 1:46, 1:50, 1:2] 0 0 0 0 0 0 0 0 0 0 ...
}

\examples{
data(SnapMSE)
}
\keyword{datasets}
