% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Output.R
\name{SPSRA}
\alias{SPSRA}
\title{Surplus Production Stock Reduction Analysis}
\usage{
SPSRA(x, Data, reps = 100)
}
\arguments{
\item{x}{A position in a data-limited methods data object}

\item{Data}{A data-limited methods data object (class DLM)}

\item{reps}{The number of samples of the TAC taken for the calculation of
the quota}
}
\description{
A surplus production equivalent of DB-SRA that uses a demographically
derived prior for intrinsic rate of increase (McAllister method, below)
}
\references{
McAllister, M.K., Pikitch, E.K., and Babcock, E.A. 2001. Using
demographic methods to construct Bayesian priors for the intrinsic rate of
increase in the Schaefer model and implications for stock rebuilding. Can.
J. Fish. Aquat. Sci. 58: 1871-1890.
}
\author{
T. Carruthers
}
