% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Output.R
\name{SPSRA_}
\alias{SPSRA_}
\alias{SPSRAopt}
\title{Surplus Production Stock Reduction Analysis Internal Function}
\usage{
SPSRA_(x, Data, reps = 100, dep = NULL)

SPSRAopt(lnK, dep, r, Ct, PE)
}
\arguments{
\item{x}{A position in a data-limited methods data object}

\item{Data}{A data-limited methods data object (class DLM)}

\item{reps}{The number of samples of the TAC taken for the calculation of
the quota}

\item{dep}{internal parameter}

\item{lnK}{internal parameter}

\item{r}{internal parameter}

\item{Ct}{internal parameter}

\item{PE}{internal parameter}
}
\description{
Surplus Production Stock Reduction Analysis Internal Function

SPSRA Optimizer
}
\section{Functions}{
\itemize{
\item \code{SPSRA_}: internal function

\item \code{SPSRAopt}: internal function
}}

\keyword{internal}
