% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessmentToOM.R, R/SS2DLM.R, R/iSCAM2DLM.R
\name{SRopt}
\alias{SRopt}
\title{Function that returns a stochastic estimate of steepness given observed stock recruitment data}
\usage{
SRopt(nsim, SSB, rec, SSBpR, plot = F, type = "BH")

SRopt(nsim, SSB, rec, SSBpR, plot = F, type = "BH")

SRopt(nsim, SSB, rec, SSBpR, plot = F, type = "BH")
}
\arguments{
\item{nsim}{number of samples of steepness to generate}

\item{SSB}{'observations' of spawning biomass}

\item{rec}{'observations' (model predictions) of recruitment}

\item{SSBpR}{spawning stock biomass per recruit at unfished conditions}

\item{plot}{should plots of model fit be produced?}

\item{type}{what type of stock recruitment curve is being fitted BH = Beverton-Holt}

\item{nsim}{number of samples of steepness to generate}

\item{SSB}{'observations' of spawning biomass}

\item{rec}{'observations' (model predictions) of recruitment}

\item{SSBpR}{spawning stock biomass per recruit at unfished conditions}

\item{plot}{should plots of model fit be produced?}

\item{type}{what type of stock recruitment curve is being fitted BH = Beverton-Holt}

\item{nsim}{number of samples of steepness to generate}

\item{SSB}{'observations' of spawning biomass}

\item{rec}{'observations' (model predictions) of recruitment}

\item{SSBpR}{spawning stock biomass per recruit at unfished conditions}

\item{plot}{should plots of model fit be produced?}

\item{type}{what type of stock recruitment curve is being fitted BH = Beverton-Holt}
}
\description{
Function that returns a stochastic estimate of steepness given observed stock recruitment data

Function that returns a stochastic estimate of steepness given observed stock recruitment data

Function that returns a stochastic estimate of steepness given observed stock recruitment data
}
\author{
T. Carruthers

T. Carruthers

T. Carruthers
}

