% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS2DLM.R
\name{SS2Data}
\alias{SS2Data}
\title{Reads data Stock Synthesis file structure into an data object using package r4ss}
\usage{
SS2Data(SSdir, Source = "No source provided", length_timestep = NA,
  Name = "", Author = "No author provided", printstats = F, verbose = T)
}
\arguments{
\item{SSdir}{A folder with Stock Synthesis input and output files in it}

\item{Source}{Reference to assessment documentation e.g. a url}

\item{length_timestep}{The duration (in years) of each timestep in the model (if an quarterly model is used this is 0.25)}

\item{Name}{The name of the operating model}

\item{Author}{Who did the assessment}

\item{printstats}{Should the r4ss function SS_output return info on data that was read in?}

\item{verbose}{Should the r4ss function SS_ouput return detailed messages?}
}
\description{
A function that uses the file location of a fitted SS3 model including input files to population the various slots of an data object
}
\author{
T. Carruthers
}
