% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{Albacore}
\alias{Albacore}
\alias{Blue_shark}
\alias{Bluefin_tuna}
\alias{Bluefin_tuna_WAtl}
\alias{Butterfish}
\alias{Herring}
\alias{Mackerel}
\alias{Porgy}
\alias{Rockfish}
\alias{Snapper}
\alias{Sole}
\alias{Toothfish}
\alias{Stock-class-objects}
\title{Stock class objects}
\format{An object of class \code{Stock} of length 1.}
\usage{
Albacore

Blue_shark

Bluefin_tuna

Bluefin_tuna_WAtl

Butterfish

Herring

Mackerel

Porgy

Rockfish

Snapper

Sole

Toothfish
}
\description{
Example objects of class Stock
}
\examples{
Toothfish@Source

avail("Stock")
}
\keyword{datasets}
