% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{StockDescription}
\alias{StockDescription}
\title{StockDescription}
\format{
An object of class \code{data.frame} with 34 rows and 2 columns.
}
\usage{
StockDescription
}
\description{
A data.frame with description of slots for class Stock
}
\keyword{datasets}
