% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMobjects.R
\name{STY}
\alias{STY}
\title{Performance Metric: Probability Short-Term Yield > 0.5 Relative Yield}
\usage{
STY(MSEobj = NULL)
}
\arguments{
\item{MSEobj}{An object of class MSE}

\item{Ref}{Reference Point for short-term yield (fraction of reference yield)}
}
\value{
An object of class PMobj
}
\description{
Performance Metric: Probability Short-Term Yield > 0.5 Relative Yield
}
\examples{
\dontrun{
STY(myMSE)
}
}
