% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubOM.r
\name{Sub.OM}
\alias{Sub.OM}
\title{Subset a Stock, Fleet, or Obs object from an OM object}
\usage{
Sub.OM(OM, Sub = c("Stock", "Fleet", "Obs", "Imp"))
}
\arguments{
\item{OM}{An operating model object (class OM)}

\item{Sub}{A character string specifying what object type to strip out
"Stock", "Fleet", or "Obs"}
}
\value{
An object of class Stock, Fleet, or Obs
}
\description{
A function that strips out a Stock, Fleet, Obs, or Imp object from a 
complete OM object. Mainly used for internal functions.
}
\author{
A. Hordyk
}
