% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFO.R
\name{SubCpars}
\alias{SubCpars}
\title{Subset an OM cpars slot}
\usage{
SubCpars(OM, sims)
}
\arguments{
\item{OM}{An object of class OM}

\item{sims}{A logical vector OM@nsim long of simulations to either retain (TRUE) or remove (FALSE)}
}
\value{
An object of class OM
}
\description{
Subset the custom parameters of an operating model
}
\author{
T. Carruthers
}
