% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_functions.R
\name{SubOM}
\alias{SubOM}
\title{Subset a Stock, Fleet, Obs, or Imp object from an OM object}
\usage{
SubOM(OM, Sub = c("Stock", "Fleet", "Obs", "Imp", "BioEco"))
}
\arguments{
\item{OM}{An operating model object (class OM)}

\item{Sub}{A character string specifying what object type to strip out
"Stock", "Fleet", "Obs", "Imp", or "BioEco"}
}
\value{
An object of class Stock, Fleet, Obs, Imp, or BioEco
}
\description{
A function that strips out a Stock, Fleet, Obs, or Imp object from a
complete OM object. Mainly used for internal functions.
}
\examples{
Stock <- SubOM(DLMtool::testOM, "Stock")
class(Stock)
}
\author{
A. Hordyk
}
