% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineClasses.r
\docType{methods}
\name{summary,DLM_data-method}
\alias{summary,DLM_data-method}
\title{Summary of DLM_data object}
\usage{
\S4method{summary}{DLM_data}(object)
}
\arguments{
\item{object}{object of class DLM_data}
}
\description{
Summary of DLM_data object
}

