% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefinePMClasses.R
\docType{methods}
\name{summary,MSE-method}
\alias{summary,MSE-method}
\title{Summary of MSE object}
\usage{
\S4method{summary}{MSE}(object, ...)
}
\arguments{
\item{object}{object of class MSE}

\item{...}{a list of PM methods}
}
\description{
Summary of MSE object
}
