% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{Swordfish_OM}
\alias{Swordfish_OM}
\title{Swordfish_OM OM}
\format{An object of class \code{OM} of length 1.}
\usage{
Swordfish_OM
}
\description{
An object of class OM
 Author: Alex Hanke (DFO)  Assessment: http://iccat.org/Documents/SCRS/Manual/CH2/2_1_9_SWO_ENG.pdf
}
\keyword{datasets}
