% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataObj.r
\name{TAC}
\alias{TAC}
\title{Calculate TAC recommendations for more than one MP}
\usage{
TAC(DLM_data, MPs = NA, reps = 100, timelimit = 1)
}
\arguments{
\item{DLM_data}{A data-limited methods data object}

\item{MPs}{optional vector of MP names}

\item{reps}{Number of repititions}

\item{timelimit}{The maximum time (seconds) taken to complete 10 reps}
}
\description{
A function that returns the stochastic TAC recommendations from a vector of
data-limited MPs (DLM_output) given a data-limited data object DLM_data
}
\author{
T. Carruthers
}
