% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_SupportingFunctions.R
\name{TACfilter}
\alias{TACfilter}
\title{TAC Filter}
\usage{
TACfilter(TAC)
}
\arguments{
\item{TAC}{A numeric vector of TAC recommendations}
}
\description{
Filters vector of TAC recommendations by replacing negatives with NA and
and values beyond five standard deviations from the mean as NA
}
\author{
T. Carruthers
}
