% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{Target_All_Fish}
\alias{Target_All_Fish}
\title{Target_All_Fish Fleet}
\format{An object of class \code{Fleet} of length 1.}
\usage{
Target_All_Fish
}
\description{
An object of class Fleet
}
\keyword{datasets}
