% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{Targeting_Small_Fish}
\alias{Targeting_Small_Fish}
\title{Targeting_Small_Fish Fleet}
\format{An object of class \code{Fleet} of length 1.}
\usage{
Targeting_Small_Fish
}
\description{
An object of class Fleet
}
\keyword{datasets}

