% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFuns.r
\name{tdlnorm}
\alias{tdlnorm}
\title{Calculate density of log-normally distributed random numbers}
\usage{
tdlnorm(x, mu, cv)
}
\arguments{
\item{x}{vector}

\item{mu}{mean}

\item{cv}{coefficient of variation}
}
\value{
numeric
}
\description{
Calculate density of log-normally distributed random numbers
}
\author{
T. Carruthers
}

