% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{testFunction}
\alias{testFunction}
\title{Example name}
\usage{
testFunction(x, Data, reps = 100)
}
\arguments{
\item{x}{A position in data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of TAC samples}
}
\description{
A brief description of the management procedure.
}
\details{
more details in paragraphs

Second paragraph

italics text: \emph{italics}
bold text:  \strong{bold}
r code: \code{r_function_call(with = "arguments")}, \code{NULL}, \code{TRUE}
}
\seealso{
links to other functions or urls
  \code{\link{prod}} for products, \code{\link{cumsum}} for cumulative
  sums, and \code{\link{colSums}}/\code{\link{rowSums}} marginal sums over
  high-dimensional arrays.
}
\author{
T. Carruthers
}
