% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{tiny}
\alias{tiny}
\title{tiny}
\format{An object of class \code{numeric} of length 1.}
\usage{
tiny
}
\description{
A tiny value used in DLMtool to avoid zeros for numerical reasons
}
\keyword{internal}
