% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{Toothfish}
\alias{Toothfish}
\title{Toothfish Stock}
\format{An object of class \code{Stock} of length 1.}
\usage{
Toothfish
}
\description{
An object of class Stock
}
\keyword{datasets}
