\name{Tplot}
\alias{Tplot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{A trade-off plot for an MSE object }
\description{
A shorter version of the plot method for MSEs that just shows the overall trade-offs
}
\usage{
Tplot(MSEobj,nam=NA)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{MSEobj}{
An object of class 'MSE'
}
\item{nam}{
Name of the plot
}
}
\author{
T. Carruthers
}

