\name{Tplot2}
\alias{Tplot2}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{A shorter version of the plot method for MSEs that just shows the overall trade-offs}
\description{
A trade-off plot for an MSE object that compares long-term yield (LTY: fraction of simulations getting over half FMSY yield in the last ten years of the projection), short-term yield (STY: fraction of simulations getting over half FMSY yield in the first ten years of the projection), variability in yield (VY: fraction of simulations where average annual variability in yield is less than 10 per cent) and biomass level (B10: the fraction of simulations in which biomass stays above 10 percent of BMSY).
}
\usage{
Tplot2(MSEobj,nam=NA)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{MSEobj}{
An object of class 'MSE'
}
\item{nam}{
Name of the plot
}
}
\author{
T. Carruthers
}

