% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE_Plotting.R
\name{Tplot3}
\alias{Tplot3}
\title{Test Trade-Off Plot}
\usage{
Tplot3(MSEobj, ..., lims = c(0.2, 0.2, 0.8, 0.8))
}
\arguments{
\item{MSEobj}{An object of class MSE}

\item{...}{Names of PM methods to plot}

\item{lims}{Numeric vector of satisficing limits. Recycled to number of PM methods}
}
\value{
produces a plot
}
\description{
Test Trade-Off Plot
}
\examples{
\dontrun{
 Tplot3(myMSE)
}
}
\author{
A. Hordyk
}
