% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE_Plotting_PMs.R
\name{TradePlot_n}
\alias{TradePlot_n}
\alias{Tplot_n}
\alias{Tplot2_n}
\alias{Tplot3_n}
\title{Generic Trade-Plot Function - IN DEVELOPMENT}
\usage{
TradePlot_n(MSEobj, ..., lims = c(0.2, 0.2, 0.8, 0.8), point.size = 2,
  lab.size = 4, axis.title.size = 12, axis.text.size = 10,
  legend.title.size = 12, position = c("right", "bottom"),
  fill = "gray80", alpha = 0.4, PMlist = NULL, Refs = NULL)

Tplot_n(MSEobj, lims = c(0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5), ...)

Tplot2_n(MSEobj, lims = c(0.2, 0.2, 0.8, 0.8), ...)

Tplot3_n(MSEobj, lims = c(0.5, 0.5, 0.8, 0.5), ...)
}
\arguments{
\item{MSEobj}{An object of class \code{MSE}}

\item{...}{Names of Performance Metrics (PMs). First PM is recycled if number of PMs is not even}

\item{lims}{A numeric vector of acceptable risk/minimum probability thresholds. Recycled if not equal to number of PMs.}

\item{point.size}{Numeric. Size of the MP points}

\item{lab.size}{Numeric. Size of MP label}

\item{axis.title.size}{Numeric. Size of axis titles}

\item{axis.text.size}{Numeric. Size of axis text}

\item{legend.title.size}{Numeric. Size of legend title text}

\item{position}{Character. Position of legend - 'right' or 'bottom'}

\item{fill}{Character. Color of the fill}

\item{alpha}{Numeric. Transparency of fill}

\item{PMlist}{Optional list of PM names. Overrides any supplied in ... above}

\item{Refs}{An optional named list (matching the PM names) with numeric values to override the default \code{Ref} values. See examples.}
}
\value{
A summary table of MP performance
}
\description{
Generic Trade-Plot Function - IN DEVELOPMENT
}
\section{Functions}{
\itemize{
\item \code{Tplot_n}: A trade-off plot showing probabilities that:
\itemize{
\item overfishing (POF) against long-term yield is > 50\% of reference yield (LTY)
\item spawning biomass is below BMSY (P100) against LTY
\item spawning biomass is below 0.5BMSY (P50) against LTY
\item spawning biomass is below 0.1BMSY (P10) against LTY
}

\item \code{Tplot2_n}: A trade-off plot showing probabilities that:
\itemize{
\item short-term yield is > 50% of reference yield(STY) against long-term yield is > 50\% of reference yield (LTY)
\item spawning biomass is below 0.1BMSY (P10) against average annual variability in yield is < 20\% (AAVY)
}

\item \code{Tplot3_n}: A trade-off plot showing probabilities that:
\itemize{
\item overfishing (POF) against long-term yield is > 50% of reference yield (LTY)
\item spawning biomass is below 0.1BMSY (P10) against average annual variability in yield is < 20\% (AAVY)
}
}}

\author{
A. Hordyk
}
