% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFuns.r
\name{trlnorm}
\alias{trlnorm}
\title{Generate log-normally distributed random numbers}
\usage{
trlnorm(reps, mu, cv)
}
\arguments{
\item{reps}{number of random numbers}

\item{mu}{mean}

\item{cv}{coefficient of variation}
}
\value{
numeric
}
\description{
Generate log-normally distributed random numbers
}
\author{
T. Carruthers
}

