% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscInt.r
\name{TwoSidedFun}
\alias{TwoSidedFun}
\title{Internal Two sided selectivity curve}
\usage{
TwoSidedFun(L1, s1, s2, lens)
}
\arguments{
\item{L1}{length one}

\item{s1}{s1}

\item{s2}{s2}

\item{lens}{vector of lengths}

\item{tst}{s2}
}
\description{
Internal Two sided selectivity curve
}
\keyword{internal}

