\name{updateMSE}
\alias{updateMSE}
\title{
Update an MSE object with new slots
}
\description{
Updates an existing MSE object (class MSE) from a previous version of the DLMtool to include the new slots.  The slots will be empty, but avoids the 'slot doesn't exist' error that sometimes occurs.
}
\usage{
updateMSE(MSEobj)
}
\arguments{
  \item{MSEobj}{
A MSE object from a previous version of the DLMtool
}
}
\value{
An object of class \code{MSE} 
}
\author{
A. Hordyk
}

