% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_Exported.R
\name{userguide}
\alias{userguide}
\title{Open the DLMtool User Guide}
\usage{
userguide()
}
\description{
Opens the DLMtool User Guide website (requires internet connection)
}
\examples{
\dontrun{
userguide()
}
}
