% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Functions.R
\name{Uses}
\alias{Uses}
\title{Find the Management Procedures that use a particular data slot}
\usage{
Uses(slot, silent = FALSE)
}
\arguments{
\item{slot}{A slot from an object of class \code{Data}. Character string.}

\item{silent}{Logical. Should messages be printed?}
}
\value{
A character string of MPs that use the slot.
}
\description{
Find the Management Procedures that use a particular data slot
}
\examples{
Uses("Mort")
}
\author{
A. Hordyk
}
