% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleOM.R
\name{validcpars}
\alias{validcpars}
\title{Valid custom parameters (cpars)}
\usage{
validcpars(print = TRUE)
}
\arguments{
\item{print}{Print the valid names for cpars?}
}
\value{
invisibly returns vector of valid cpars names
}
\description{
Valid custom parameters (cpars)
}
