% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{WM_Generic_fleet_GOM}
\alias{WM_Generic_fleet_GOM}
\title{WM_Generic_fleet_GOM OM}
\format{An object of class \code{OM} of length 1.}
\usage{
WM_Generic_fleet_GOM
}
\description{
An object of class OM
}
\keyword{datasets}
