% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Functions.R
\name{XL2Data}
\alias{XL2Data}
\title{Import a Data object from Excel file}
\usage{
XL2Data(name = "Data")
}
\arguments{
\item{name}{Name of the data file, with or without file extension.}
}
\value{
An object of class 'Data'
}
\description{
Import a Data object from Excel file
}
\examples{
\dontrun {
Data <- XL2Data('Data.xlsx')
}
}
\author{
A. Hordyk
}
