% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{Yellowfin_OM}
\alias{Yellowfin_OM}
\title{Yellowfin_OM OM}
\format{An object of class \code{OM} of length 1.}
\usage{
Yellowfin_OM
}
\description{
An object of class OM
}
\keyword{datasets}
