% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{Yellowfin_Tuna_IO}
\alias{Yellowfin_Tuna_IO}
\title{Yellowfin_Tuna_IO OM}
\format{An object of class \code{OM} of length 1.}
\usage{
Yellowfin_Tuna_IO
}
\description{
An object of class OM
}
\keyword{datasets}

