% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{Yellowfin_tuna_IO_OM}
\alias{Yellowfin_tuna_IO_OM}
\title{Yellowfin_tuna_IO_OM OM}
\format{An object of class \code{OM} of length 1.}
\usage{
Yellowfin_tuna_IO_OM
}
\description{
An object of class OM (from SS using SS2DLM)
 SEDAR 31, 2014 update assessment. Author: No author provided.
}
\keyword{datasets}

