% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{YM_YM_Rec_comb_GOM}
\alias{YM_YM_Rec_comb_GOM}
\title{YM_YM_Rec_comb_GOM OM}
\format{An object of class \code{OM} of length 1.}
\usage{
YM_YM_Rec_comb_GOM
}
\description{
An object of class OM
}
\keyword{datasets}
