% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Output.R
\name{YPR_}
\alias{YPR_}
\alias{YPRopt}
\title{Internal function for YPR MPs}
\usage{
YPR_(x, Data, reps = 100, Abun = NULL)

YPRopt(Linfc, Kc, t0c, Mdb, a, b, LFS, maxage, reps = 100)
}
\arguments{
\item{x}{Iteration number}

\item{Data}{Object of class Data}

\item{reps}{Number of reps}

\item{Abun}{Optional numeric of length \code{reps} of abundance}

\item{Linfc}{von Bertalanffy  Linf}

\item{Kc}{von Bertalanffy K}

\item{t0c}{von Bertalanffy t0}

\item{a}{alpha parameter of L-W relationship}

\item{b}{beta parameter of L-W relationship}

\item{LFS}{Length at full selection}

\item{maxage}{Maximum age}
}
\description{
Internal function for YPR MPs
}
\section{Functions}{
\itemize{
\item \code{YPR_}: Internal function for YPR MPs

\item \code{YPRopt}: YPR calculation function
}}

\note{
Yield per recruit estimate of FMSY Meaghan Bryan 2013
}
\keyword{internal}
