% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{YPR_ML}
\alias{YPR_ML}
\title{Yield Per Recruit analysis to get FMSY proxy F01 paired with a mean-length
estimate of current stock size}
\usage{
YPR_ML(x, Data, reps = 100)
}
\arguments{
\item{x}{A position in a data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of samples of the TAC}
}
\description{
A simple yield per recruit approximation to FMSY (F01) which is the position
of the ascending YPR curve for which dYPR/dF = 0.1(dYPR/d0) A mean-length
estimate of recent Z is used to infer current abundance
}
\note{
The mean length extension was programmed by Gary Nelson as part of his
excellent R package 'fishmethods'
}
\author{
Meaghan Bryan and T. Carruthers
}
