\name{avail}
\alias{avail}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
What objects of this class are available
}
\description{
Generic class finder
}
\usage{
avail(classy)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{classy}{
A class of object (character string, e.g. 'Fleet')
}
}
\details{
Finds objects of the specified class in the global environment or the package:DLMtool
}
\author{
T. Carruthers
}