% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_Exported.R
\name{avail}
\alias{avail}
\title{What objects of this class are available}
\usage{
avail(classy, builtin = FALSE)
}
\arguments{
\item{classy}{A class of object (character string, e.g. 'Fleet')}

\item{builtin}{Logical. Only return Objects of class 'classy' from DLMtool & DLMextra packages?}
}
\description{
Generic class finder
}
\details{
Finds objects of the specified class in the global environment or the
DLMtool package.
}
\examples{
avail("OM")
Stocks <- avail("Stock")
Fleets <- avail("Fleet")
MPs <- avail("MP")
}
\seealso{
\link{Can} \link{Cant} \link{avail}
}
\author{
T. Carruthers
}
