% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLM_output.R
\name{AvC}
\alias{AvC}
\title{Average Catch}
\usage{
AvC(x, DLM_data, reps = 100)
}
\arguments{
\item{x}{A position in a data-limited methods data object}

\item{DLM_data}{A data-limited methods data object}

\item{reps}{The number of stochastic samples of the TAC recommendation}
}
\description{
A simple average catch MP that is included to demonstrate a 'status quo' management option
}
\author{
T. Carruthers
}
