% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Mixed.R
\name{AvC_MLL}
\alias{AvC_MLL}
\title{Average Catch with a size limit}
\usage{
AvC_MLL(x, Data, reps = 100)
}
\arguments{
\item{x}{A position in a data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of stochastic samples of the TAC recommendation}
}
\description{
Included demonstration purposes of a mixed control MP
}
