% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output.r
\name{BK_CC}
\alias{BK_CC}
\title{Beddington and Kirkwood life-history method combined with catch curve
analysis}
\usage{
BK_CC(x, Data, reps = 100, Fmin=0.005)
}
\arguments{
\item{x}{Position in a data-limited methods data object}

\item{Data}{A data-limited methods data object (class Data)}

\item{reps}{The number of samples of the TAC recommendation}

\item{Fmin}{The minimum fishing mortality rate that is derived from the
catch-curve (interval censor)}
}
\description{
Calculates an OFL using a catch curve estimate of current F and an
approximation of FMSY based on length at first capture.
}
\references{
Beddington, J.R., Kirkwood, G.P., 2005. The estimation of
potential yield and stock status using life history parameters. Philos.
Trans. R. Soc. Lond. B Biol. Sci. 360, 163-170.
}
\author{
T. Carruthers
}
