% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output.r
\docType{data}
\name{BK_ML}
\alias{BK_ML}
\title{Beddington and Kirkwood life-history analysis with mean-length estimator of
current abundance}
\format{An object of class \code{Output} of length 1.}
\usage{
BK_ML(x, Data, reps = 100)
}
\arguments{
\item{x}{Position in a data-limited methods data object}

\item{Data}{A data-limited methods data object (class Data)}

\item{reps}{The number of samples of the TAC recommendation}
}
\description{
Uses an approximation to FMSY based on length at first capture and an
estimate of current abundance based on a mean-length estimator.
}
\note{
The mean length extension was programmed by Gary Nelson as part of his
excellent R package 'fishmethods'
}
\author{
T. Carruthers
}
\references{
Beddington, J.R., Kirkwood, G.P., 2005. The estimation of
potential yield and stock status using life history parameters. Philos.
Trans. R. Soc. Lond. B Biol. Sci. 360, 163-170.
}
\keyword{datasets}

