% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataObj.r
\name{boxplot.Data}
\alias{boxplot.Data}
\title{Boxplot of TAC recommendations}
\usage{
\method{boxplot}{Data}(x, outline = FALSE, ...)
}
\arguments{
\item{x}{An object of class MSE}

\item{outline}{Logical. Include outliers in plot?}

\item{...}{Optional additional arguments passed to \code{boxplot}}
}
\value{
Returns a data frame containing the information shown in the plot
}
\description{
Boxplot of TAC recommendations
}
\author{
A. Hordyk
}
