% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Plotting.R
\name{boxplot.Data}
\alias{boxplot.Data}
\title{Boxplot of TAC recommendations}
\usage{
\method{boxplot}{Data}(x, upq = 0.9, lwq = 0.1, ylim = NULL, outline = FALSE, col = NULL, ...)
}
\arguments{
\item{x}{An object of class MSE}

\item{upq}{Upper quantile of TACs for max ylim}

\item{lwq}{Lower quantile of TACs for min ylim}

\item{ylim}{Optional numeric vector of length 2 to specify limits of y-axis.}

\item{outline}{Logical. Include outliers in plot?}

\item{col}{Optional colours to pass to \code{boxplot}}

\item{...}{Optional additional arguments passed to \code{boxplot}}
}
\value{
Returns a data frame containing the information shown in the plot
}
\description{
Boxplot of TAC recommendations
}
\author{
A. Hordyk
}
