% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMobjects.R
\name{calcMean}
\alias{calcMean}
\title{Calculate Mean Probability}
\usage{
calcMean(Prob)
}
\arguments{
\item{Prob}{Prob slot from an object of class PMobj}
}
\description{
Calculate Mean Probability
}
\keyword{internal}
