% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataObj.r
\name{Can}
\alias{Can}
\title{What data-limited methods can be applied to this DLM_data object?}
\usage{
Can(DLM_data, timelimit = 1)
}
\arguments{
\item{DLM_data}{A data-limited methods data object (class DLM_data)}

\item{timelimit}{The maximum time (seconds) taken for a method to undertake
10 reps (this filters out methods that are too slow)}
}
\description{
An diagnostic tool that looks up the slot requirements of each method and
compares this to the data available to limit the analysis to methods that
have the correct data, do not produce errors and run within a time limit.
Time limit is the maximum time taken to carry out five reps (stochastic
samples) of a given method and is in units of seconds.
}
