% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLM_output.R
\name{CC1}
\alias{CC1}
\title{Constant catch management procedure of Geromont and Butterworth (2014)}
\usage{
CC1(x, DLM_data, reps = 100, yrsmth = 5, xx=0)
}
\arguments{
\item{x}{A position in data-limited methods data object}

\item{DLM_data}{A data-limited methods data object}

\item{reps}{The number of TAC samples}

\item{yrsmth}{Years over which to calculate mean catches}

\item{xx}{Parameter controlling the TAC. Mean catches are multiplied by
(1-xx)}
}
\value{
A numeric vector of TAC recommendations
}
\description{
The TAC is the average catch over last yrsmth years.
}
\details{
This is one of four constant catch rules of Geromont and Butterworth 2014.
}
\references{
Geromont, H.F., Butterworth, D.S. 2014. Generic management
procedures for data-poor fisheries; forecasting with few data. ICES J. Mar.
Sci. doi:10.1093/icesjms/fst232
}
\author{
T. Carruthers
}
