% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSEObj.r
\name{checkMSE}
\alias{checkMSE}
\title{Check that MSE object includes all slots}
\usage{
checkMSE(MSEobj)
}
\arguments{
\item{MSEobj}{A MSE object.}

\item{MPs}{A vector MPs names or MP numbers to subset the MSE object.}
}
\description{
Check that an MSE object includes all slots in the latest version of DLMtool
Use `updateMSE` to update the MSE object
}
\author{
A. Hordyk
}

