% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerfMetrics.r
\name{ChkYrs}
\alias{ChkYrs}
\title{Internal function to check or correct years in performance metrics}
\usage{
ChkYrs(MSEobj, y1 = NULL, y2 = NULL)
}
\arguments{
\item{MSEobj}{An object of class 'MSE'}

\item{y1}{A numeric value indicating the first year to summarize performance}

\item{y2}{A numeric value indicating the last year to summarize performance}
}
\value{
A data.frame containing y1 and y2
}
\description{
Internal function to check or correct years in performance metrics
}
