% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sketch_Functions.R
\name{ChooseEffort}
\alias{ChooseEffort}
\title{Manually map the historical relative fishing effort trajectory.}
\usage{
ChooseEffort(FleetObj, Years=NULL)
}
\arguments{
\item{FleetObj}{A fleet object.}

\item{Years}{An optional vector of years. Should be nyears long.}
}
\description{
Interactive plot which allows users to specify the relative trajectory and
variability in the historical fishing effort and populates Fleet object.
}
\author{
A. Hordyk
}
