% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output.r
\docType{data}
\name{CompSRA}
\alias{CompSRA}
\title{Age-composition-based estimate of current stock depletion given constant Z
linked to an FMSY estimate to provide OFL}
\format{An object of class \code{Output} of length 1.}
\usage{
CompSRA(x, Data, reps = 100)
}
\arguments{
\item{x}{A position in a data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of stochastic samples of the TAC.}
}
\description{
Estimates an OFL based on a Stock Reduction analysis fitted to current
age-composition data. Knife-edge vulnerability at age at maturity allows for
an FMSY estimate. OFL=FMSY*F/C
}
\note{
Given a fixed historical F, What level of depletion gives you this
length composition?
}
\author{
T. Carruthers
}
\keyword{datasets}

