% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Output.R
\name{CompSRA}
\alias{CompSRA}
\alias{CompSRA4010}
\title{Age-composition-based estimate of current stock depletion given constant Z
linked to an FMSY estimate to provide OFL}
\usage{
CompSRA(x, Data, reps = 100, plot = FALSE)

CompSRA4010(x, Data, reps = 100)
}
\arguments{
\item{x}{A position in a data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of stochastic samples of the TAC recommendation}

\item{plot}{Logical. Show the plot?}
}
\value{
An object of class \code{\link[=Rec-class]{Rec}} with the \code{TAC} slot populated with a numeric vector of length \code{reps}

A Rec object with a vector of TAC recommendation
}
\description{
Estimates an OFL based on a Stock Reduction analysis fitted to current
age-composition data. Knife-edge vulnerability at age at maturity allows for
an FMSY estimate.
}
\section{Functions}{
\itemize{
\item \code{CompSRA}: Basic control rule where OFL=FMSY*F/C

\item \code{CompSRA4010}: With a 40-10 control rule
}}

\note{
See the \href{https://dlmtool.github.io/DLMtool/reference/index.html}{online documentation} for correctly rendered equations.

Given a fixed historical F, What level of depletion gives you this
length composition?
}
\author{
T. Carruthers
}
