% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input.r
\docType{data}
\name{curE}
\alias{curE}
\title{Fishing at current effort levels}
\format{An object of class \code{Input} of length 1.}
\usage{
curE(x, Data, ...)
}
\arguments{
\item{x}{A position in a data-limited methods data object.}

\item{Data}{A data-limited methods data object.}

\item{...}{Optional additional arguments that are ignored. Note arguments
\code{reps} or \code{...} are required for all input controls}
}
\description{
Constant fishing effort set at final year of historical simulations subject
to changes in catchability determined by OM@qinc and interannual variability
in catchability determined by OM@qcv. This MP is intended to represent a
'status quo' management approach.
}
\note{
Made up for this package.
}
\author{
T. Carruthers.
}
\keyword{datasets}

