\name{curE}
\alias{curE}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Fishing at current effort levels
}
\description{
Constant fishing effort set at final year of historical simulations subject to changes in catchability determined by OM@qinc and interannual variability in catchability determined by OM@qcv. This MP is intended to represent a 'status quo' management approach. 
}
\usage{
curE(x, DLM_data, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A position in a data-limited methods data object.
}
  \item{DLM_data}{
A data-limited methods data object.
}
\item{...}{Optional additional arguments that are ignored. Note arguments \code{reps} or \code{...} are required for all input controls }
}
\author{
T. Carruthers.
}
\note{
Made up for this package. 
}
