% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{curE75}
\alias{curE75}
\title{Fishing at 75 per cent of current effort levels}
\usage{
curE75(x, Data, ...)
}
\arguments{
\item{x}{A position in a data-limited methods data object.}

\item{Data}{A data-limited methods data object.}

\item{...}{Optional additional arguments that are ignored. Note arguments
\code{reps} or \code{...} are required for all input controls}
}
\description{
Constant fishing effort set at 75 per cent of final year of historical
simulations subject to changes in catchability determined by OM@qinc and
interannual variability in catchability determined by OM@qcv. This MP is
intended to represent a 'status quo' management approach.
}
\note{
Made up for this package.
}
\author{
T. Carruthers.
}
