% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Functions.R
\name{DataInit}
\alias{DataInit}
\title{Initialize an empty Data workbook or CSV}
\usage{
DataInit(name = "Data", ext = c("xlsx", "csv"), overwrite = FALSE,
  dir = NULL)
}
\arguments{
\item{name}{Name of the data file. Default is Data.xlsx}

\item{ext}{Optional file extension. 'xlsx' (default) or 'csv'}

\item{overwrite}{Logical. Overwrite existing files?}

\item{dir}{Optional directory path to create the Data file. Default is `getwd()``}
}
\value{
Nothing. Creates a data file in the working directory.
}
\description{
Initialize an empty Data workbook or CSV
}
\examples{
\dontrun{
DataInit("MyData")
}
}
\author{
A. Hordyk
}
