% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Output.R
\name{DBSRA}
\alias{DBSRA}
\alias{DBSRA_40}
\alias{DBSRA4010}
\title{Depletion-Based Stock Reduction Analysis}
\usage{
DBSRA(x, Data, reps = 100, plot = FALSE)

DBSRA_40(x, Data, reps = 100, plot = FALSE)

DBSRA4010(x, Data, reps = 100, plot = FALSE)
}
\arguments{
\item{x}{A position in a data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of stochastic samples of the TAC recommendation}

\item{plot}{Logical. Show the plot?}
}
\value{
An object of class \code{\link[=Rec-class]{Rec}} with the \code{TAC} slot populated with a numeric vector of length \code{reps}
}
\description{
User prescribed BMSY/B0, M, FMSY/M are used to find B0 and therefore the OFL
by back-constructing the stock to match a user specified level of stock
depletion (OFL = M * FMSY/M * depletion* B0).
}
\section{Functions}{
\itemize{
\item \code{DBSRA}: Base version. You specify a range of stock depletion and,
given historical catches DB-SRA calculates what unfished biomass must have
been to get you here given samples for M, FMSY relative to M and also BMSY
relative to Bunfished.

\item \code{DBSRA_40}: Assumes 40 percent current depletion (Bcurrent/B0 = 0.4), which is
more or less the most optimistic state for a stock (ie very close to BMSY/B0 for many stocks).

\item \code{DBSRA4010}: Base version paired with the 40-10 rule that throttles
back the TAC to zero at 10 percent of unfished biomass.
}}

\note{
This is set up to return the OFL (FMSY * current biomass).

You may have noticed that you -the user- specify three of the factors that
make the quota recommendation. So this can be quite a subjective method.

Also the DB-SRA method of this package isn't exactly the same as the
original method of Dick and MacCall (2011) because it has to work for
simulated depletions above BMSY/B0 and even on occasion over B0. Also it
doesn't have the modification for flatfish life histories that has
previously been applied by Dick and MacCall.
}
\section{Required \link[=Data-class]{Data}}{

\code{DBSRA}: BMSY_B0, Cat, Dep, FMSY_M, L50, Mort, vbK, vbLinf, vbt0


\code{DBSRA_40}: BMSY_B0, Cat, Dep, FMSY_M, L50, Mort, vbK, vbLinf, vbt0


\code{DBSRA4010}: BMSY_B0, Cat, Dep, FMSY_M, L50, Mort, vbK, vbLinf, vbt0
}

\section{Rendered Equations}{
 
See \href{https://dlmtool.github.io/DLMtool/reference/DBSRA.html}{Online Documentation} for correctly rendered equations




}

\references{
Dick, E.J., MacCall, A.D., 2010. Estimates of sustainable yield for 50 data-poor
stocks in the Pacific Coast groundfish fishery management plan. Technical memorandum.
Southwest fisheries Science Centre, Santa Cruz, CA. National Marine Fisheries Service,
National Oceanic and Atmospheric Administration of the U.S. Department of Commerce.
NOAA-TM-NMFS-SWFSC-460.

Dick, E.J., MacCall, A.D., 2011. Depletion-Based Stock Reduction
Analysis: A catch-based method for determining sustainable yields for
data-poor fish stocks. Fish. Res. 110, 331-341.
}
\author{
T. Carruthers
}
