% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{DBSRA_ML}
\alias{DBSRA_ML}
\title{Depletion-Based Stock Reduction Analysis using mean length estimator of
stock depletion}
\usage{
DBSRA_ML(x, Data, reps = 100)
}
\arguments{
\item{x}{A position in a data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of stochastic samples of the quota recommendation}
}
\description{
DBSRA using the mean length estimator to calculate current stock depletion.
}
\note{
The mean length extension was programmed by Gary Nelson as part of his
excellent R package 'fishmethods'
}
\references{
Dick, E.J., MacCall, A.D., 2011. Depletion-Based Stock Reduction
Analysis: A catch-based method for determining sustainable yields for
data-poor fish stocks. Fish. Res. 110, 331-341.
}
\author{
T. Carruthers
}
