% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Output.R
\name{DCAC}
\alias{DCAC}
\alias{DCAC_40}
\alias{DCAC4010}
\alias{DCAC_ML}
\alias{DAAC}
\alias{HDAAC}
\title{Depletion Corrected Average Catch}
\usage{
DCAC(x, Data, reps = 100, plot = FALSE)

DCAC_40(x, Data, reps = 100, plot = FALSE)

DCAC4010(x, Data, reps = 100, plot = FALSE)

DCAC_ML(x, Data, reps = 100, plot = FALSE)

DAAC(x, Data, reps = 100)

HDAAC(x, Data, reps = 100)
}
\arguments{
\item{x}{A position in a data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of stochastic samples of the TAC recommendation}

\item{plot}{Logical. Show the plot?}
}
\value{
An object of class \code{\link[=Rec-class]{Rec}} with the \code{TAC} slot populated with a numeric vector of length \code{reps}
}
\description{
A method of calculating an MSY proxy (FMSY * BMSY and therefore the OFL at
most productive stock size) based on average catches accounting for the
windfall catch that got the stock down to BMSY levels.
}
\section{Functions}{
\itemize{
\item \code{DCAC_40}: This variant assumes that current stock biomass is exactly
40 per cent of unfished levels. The 40 percent depletion assumption may not
really affect DCAC that much as it already makes TAC recommendations that are
quite MSY-like.

\item \code{DCAC4010}: In this method, DCAC is paired with the 40-10 rule that throttles
back the OFL to zero at 10 percent of unfished stock size (the OFL is not subject to downward
adjustment above 40 percent unfished). DCAC can overfish below BMSY levels. The 40-10
harvest control rule largely resolves this problem providing an MP with surprisingly good
performance even at low stock levels.

\item \code{DCAC_ML}: This variant uses the mean length estimator to calculate current stock
depletion. The mean length extension was programmed by Gary Nelson as part of his
excellent R package 'fishmethods'.

\item \code{DAAC}: Depletion Adjusted Average Catch: essentially DCAC multiplied
by 2*depletion and divided by BMSY/B0 (Bpeak) (Harford and Carruthers, 2017).

\item \code{HDAAC}: Hybrid Depletion Adjusted Average Catch: essentially DCAC
multiplied by 2*depletion and divided by BMSY/B0 (Bpeak) when below BMSY,
and DCAC above BMSY (Harford and Carruthers 2017).
}}

\note{
See the \href{https://dlmtool.github.io/DLMtool/reference/index.html}{online documentation} for correctly rendered equations.

It's probably worth noting that DCAC TAC recommendations do not tend
to zero as depletion tends to zero. It adjusts for depletion only in
calculating historical average catch. It follows that at stock levels much
below BMSY, DCAC tends to chronically overfish.
}
\examples{
DCAC(1, DLMtool::Atlantic_mackerel, plot=TRUE) 

DCAC_40(1, DLMtool::Atlantic_mackerel, plot=TRUE) 

DCAC4010(1, DLMtool::Atlantic_mackerel, plot=TRUE) 

DCAC_ML(1, DLMtool::SimulatedData, plot=TRUE) 

}
\references{
MacCall, A.D., 2009. Depletion-corrected average catch: a simple
formula for estimating sustainable yields in data-poor situations. ICES J.
Mar. Sci. 66, 2267-2271.

Harford W. and Carruthers, T. 2017. Interim and long-term performance of
static and adaptive management procedures. Fish. Res. 190, 84-94.
}
\seealso{
Other Average Catch MPs: \code{\link{AvC}}
}
\author{
T. Carruthers
}
