% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{DCAC_ML}
\alias{DCAC_ML}
\title{Depletion-Based Stock Reduction Analysis using mean-length estimator of
current depletion}
\usage{
DCAC_ML(x, Data, reps = 100)
}
\arguments{
\item{x}{A position in a data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of stochastic samples of the TAC recommendation}
}
\description{
DCAC that uses the mean length estimator to calculate current stock
depletion.
}
\note{
The mean length extension was programmed by Gary Nelson as part of his
excellent R package 'fishmethods'
}
\references{
MacCall, A.D., 2009. Depletion-corrected average catch: a simple
formula for estimating sustainable yields in data-poor situations. ICES J.
Mar. Sci. 66, 2267-2271.
}
\author{
T. Carruthers
}
