% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Input.R, R/MPs_Output.R
\name{DDe}
\alias{DDe}
\alias{DDes}
\alias{DDe75}
\alias{DD}
\alias{DD4010}
\title{Delay - Difference Stock Assessment with UMSY and MSY as leading parameters}
\usage{
DDe(x, Data, reps = 100)

DDes(x, Data, reps = 100, LB = 0.9, UB = 1.1)

DDe75(x, Data, reps = 100)

DD(x, Data, reps = 100)

DD4010(x, Data, reps = 100)
}
\arguments{
\item{x}{A position in a data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of stochastic samples of the TAC recommendation}

\item{LB}{The lowest permitted factor of previous fishing effort}

\item{UB}{The highest permitted factor of previous fishing effort}
}
\value{
A Rec object of either TAC or effort recommendations
}
\description{
A simple delay-difference assessment that estimates the TAC using a
time-series of catches and a relative abundance index. Conditioned on effort.
}
\section{Functions}{
\itemize{
\item \code{DDe}: Effort-control version. The recommended effort is the ratio of
UMSY to current U.

\item \code{DDes}: Variant of \code{DDe} that limits the maximum change in effort to 10 percent.

\item \code{DDe75}: Variant of \code{DDe} where the recommended effort is 75\%
of the ratio of UMSY to current U.

\item \code{DD}: Base version where the TAC = UMSY * Current Biomass.

\item \code{DD4010}: In this version, a 40-10 rule is imposed over the TAC recommendation.
}}

\note{
This DD model is observation error only and has does not estimate
process error (recruitment deviations). Similar to many other assessment
models it depends on a whole host of dubious assumptions such as temporally
stationary productivity and proportionality between the abundance index and
real abundance. Unsurprisingly the extent to which these assumptions are
violated tends to be the biggest driver of performance for this method.
}
\references{
Carruthers, T, Walters, C.J,, and McAllister, M.K. 2012. Evaluating methods that classify
fisheries stock status using only fisheries catch data. Fisheries Research 119-120:66-79.

Hilborn, R., and Walters, C. 1992. Quantitative Fisheries Stock Assessment: Choice,
Dynamics and Uncertainty. Chapman and Hall, New York.
}
\author{
T. Carruthers
}
