% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLM_output.R
\docType{data}
\name{DD4010}
\alias{DD4010}
\title{Delay - Difference Stock Assessment with UMSY and MSY leading coupled with a
40-10 harvest control rule}
\format{An object of class \code{DLM_output} of length 1.}
\usage{
DD4010(x, DLM_data, reps = 100)
}
\arguments{
\item{x}{A position in a data-limited methods data object}

\item{DLM_data}{A data-limited methods data object}

\item{reps}{The number of stochastic samples of the TAC recommendation}
}
\value{
A numeric vector of TAC recommendations
}
\description{
A simple delay-difference assessment that estimates the OFL using a
time-series of catches and a relative abundance index. In this version of
the DD MP a 40-10 rule is imposed over the OFL recommendation.
}
\author{
T. Carruthers
}
\references{
Method based on equations of Carl Walters
}
\keyword{datasets}

