% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output.r
\name{DD_R}
\alias{DD_R}
\title{Internal optimization function for delay-difference MPs}
\usage{
DD_R(params, opty, So_DD, Alpha_DD, Rho_DD, ny_DD, k_DD, wa_DD, E_hist, C_hist,
  UMSYprior)
}
\arguments{
\item{params}{vector of length 3 with log parameters}

\item{opty}{optimization option}

\item{So_DD}{internal parameter}

\item{Alpha_DD}{internal parameter}

\item{Rho_DD}{internal parameter}

\item{ny_DD}{internal parameter}

\item{k_DD}{internal parameter}

\item{wa_DD}{internal parameter}

\item{E_hist}{internal parameter}

\item{C_hist}{internal parameter}

\item{UMSYprior}{internal parameter}
}
\description{
Internal optimization function for delay-difference MPs
}
\author{
T. Carruthers
}
\keyword{internal}
