% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Output.R
\name{DD_R}
\alias{DD_R}
\title{Delay-Difference Internal Function}
\usage{
DD_R(
  params,
  opty,
  So_DD,
  Alpha_DD,
  Rho_DD,
  ny_DD,
  k_DD,
  wa_DD,
  E_hist,
  C_hist,
  UMSYprior
)
}
\arguments{
\item{params}{internal parameter}

\item{opty}{internal parameter}

\item{So_DD}{internal parameter}

\item{Alpha_DD}{internal parameter}

\item{Rho_DD}{internal parameter}

\item{ny_DD}{internal parameter}

\item{k_DD}{internal parameter}

\item{wa_DD}{internal parameter}

\item{E_hist}{internal parameter}

\item{C_hist}{internal parameter}

\item{UMSYprior}{internal parameter}
}
\description{
Delay-Difference Internal Function
}
\keyword{internal}
