% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Input.R
\name{DDe75}
\alias{DDe75}
\title{Effort control version of DD - Delay - Difference Stock Assessment with UMSY
and MSY leading that fishes at 75 per cent of FMSY}
\usage{
DDe75(x, Data, reps = 100)
}
\arguments{
\item{x}{A position in a data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of stochastic samples of the TAC recommendation}
}
\description{
A simple delay-difference assessment that estimates and recommends 75 per
cent FMSY using a time-series of catches and a relative abundance index.
}
\note{
This DD model is observation error only and has does not estimate
process error (recruitment deviations). Similar to many other assessment
models it depends on a whole host of dubious assumptions such as temporally
stationary productivity and proportionality between the abundance index and
real abundance. Unsurprisingly the extent to which these assumptions are
violated tends to be the biggest driver of performance for this method.
}
\references{
Carruthers, T, Walters, C.J,, and McAllister, M.K. 2012. Evaluating methods that classify
fisheries stock status using only fisheries catch data. Fisheries Research 119-120:66-79.

Hilborn, R., and Walters, C., 1992. Quantitative Fisheries Stock Assessment: Choice,
Dynamics and Uncertainty. Chapman and Hall, New York.
}
\author{
T. Carruthers
}
