\name{DepF}
\alias{DepF}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Depletion Corrected Fratio
}
\description{
The Fratio method with a controller that reduces F according to the production curve given an estimate of current stock depletion. 
}
\usage{
DepF(x, DLM, reps = 100)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A position in data-limited methods data object DLM
}
  \item{DLM}{
A data-limited methods data object
}
  \item{reps}{
The number of quota samples
}
}
\value{
A numeric vector of quota recommendations
}
\references{
Made-up for this package. 
}
\author{
T. Carruthers
}
