\name{DepF}
\alias{DepF}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Depletion Corrected Fratio}
\description{
The Fratio MP with a harvest control rule that reduces F according to the production curve given an estimate of current stock depletion. 
}
\usage{
DepF(x, DLM_data, reps = 100)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A position in data-limited methods data object DLM
}
  \item{DLM_data}{
A data-limited methods data object
}
  \item{reps}{
The number of TAC samples
}
}
\value{
A numeric vector of TAC recommendations
}
\references{
Made-up for this package. 
}
\author{
T. Carruthers
}
