\name{DLMDataDir}
\alias{DLMDataDir}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Directory of the installed package on your computer
}
\description{
A way of locating where the package was installed so you can find example data files and code etc.
}
\usage{
DLMDataDir(stock=NA)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{stock}{
  Character string representing the name of a .csv file e.g. 'Snapper', 'Rockfish'
}
}
\author{
T. Carruthers
}
