% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_Exported.R
\name{DLMDataDir}
\alias{DLMDataDir}
\title{Directory of the installed package on your computer}
\usage{
DLMDataDir(stock=NA)
}
\arguments{
\item{stock}{Character string representing the name of a .csv file e.g.
'Snapper', 'Rockfish'}
}
\description{
A way of locating where the package was installed so you can find example
data files and code etc.
}
\examples{
\dontrun{
tilefish_location <- DLMDataDir("Gulf_blue_tilefish")
tilefish_Data <- new("Data", tilefish_location)
}
}
\author{
T. Carruthers
}
