% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLMtool-package.R
\docType{package}
\name{DLMtool-package}
\alias{DLMtool}
\alias{DLMtool-package}
\title{DLMtool: Data-Limited Methods Toolkit}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A collection of data-limited management procedures that can be evaluated with management strategy evaluation with the 'MSEtool' package, or applied to fishery data to provide management recommendations.
}
\author{
\strong{Maintainer}: Adrian Hordyk \email{adrian@bluematterscience.com}

Authors:
\itemize{
  \item Tom Carruthers \email{tom@bluematterscience.com}
  \item Quang Huynh \email{quang@bluematterscience.com}
}

Other contributors:
\itemize{
  \item M. Bryan [contributor]
  \item HF. Geremont [contributor]
  \item C. Grandin [contributor]
  \item W. Harford [contributor]
  \item Q. Huynh [contributor]
  \item C. Walters [contributor]
}

}
\keyword{internal}
