% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLMtool-package.R
\docType{package}
\name{DLMtool-package}
\alias{DLMtool-package}
\alias{DLMtool}
\title{Data-Limited Methods Toolkit}
\description{
Simulation testing and implementation of data-limited fishery stock
assessment methods
}
\section{Additional Information}{

See the \href{https://dlmtool.github.io/DLMtool/userguide/introduction.html}{DLMtool User Guide} for
a detailed description of how to use the DLMtool package.

The help documentation for the DLMtool package can also be accessed \href{https://dlmtool.github.io/DLMtool/}{here}.

See the \href{http://www.datalimitedtoolkit.org/}{Data-Limited Toolkit Website} for more information on the DLMtool,
including an interactive demo of the main features of the toolkit, information on case studies where the toolkit has
been applied, and more about the history and development of the DLMtool.
}

\references{
Carruthers, T.R., Punt, A.E., Walters, C.J., MacCall, A.,
McAllister, M.K., Dick, E.J., Cope, J. 2014. Evaluating methods for setting
catch limits in data-limited fisheries. Fisheries Research. 153: 48-68.

Carruthers, T.R., Kell, L.T., Butterworth, D.S., Maunder, M.N., Geromont,
H.F., Walters, C., McAllister, M.K., Hillary, R., Levontin, P., Kitakado,
T., Davies, C.R. Performance review of simple management procedures. ICES
Journal of Marine Science.
}
\author{
Tom Carruthers \email{t.carruthers@fisheries.ubc.ca}

Adrian Hordyk \email{a.hordyk@oceans.ubc.ca}
}
\keyword{assessment}
\keyword{data-limited}
\keyword{data-poor}
\keyword{evaluation}
\keyword{fisheries}
\keyword{management}
\keyword{stock}
\keyword{strategy}
