% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSEObj.r
\name{DOM}
\alias{DOM}
\title{How dominant is an MP?}
\usage{
DOM(MSEobj, MPtg=NA)
}
\arguments{
\item{MSEobj}{An object of class 'MSE'}

\item{MPtg}{A character vector of management procedures for cross
examination}
}
\value{
A matrix of performance comparisons length(MPtg) rows by MSE@nMPs
columns
}
\description{
The DOM function examines how consistently an MP outperforms another. For
example DCAC might provide higher yield than AvC on average but outperforms
AvC in less than half of simulations.
}
\author{
A. Hordyk
}

