% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input.r
\name{EtargetLopt}
\alias{EtargetLopt}
\title{Effort MP: adjust effort up/down if mean length above/below Ltarget}
\usage{
EtargetLopt(x, Data, reps = 100, yrsmth=3, buffer=0.1)
}
\arguments{
\item{x}{A position in data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of TAC samples}

\item{yrsmth}{Number of years to calculate average length}

\item{buffer}{Parameter controlling the fraction of mean catch to set the
reference (or target) TAC level - acts as a precautionary buffer}
}
\value{
An adjustment for fishing effort
}
\description{
Effort MP: adjust effort up/down if mean length above/below Ltarget
}
\author{
HF Geromont
}
