% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{Example_datafile}
\alias{Example_datafile}
\title{Example_datafile Data}
\format{An object of class \code{Data} of length 1.}
\usage{
Example_datafile
}
\description{
An object of class Data
}
\keyword{datasets}
