% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFuns.r
\name{Fease}
\alias{Fease}
\title{MP feasibility diagnostic}
\usage{
Fease(feaseobj,outy='table')
}
\arguments{
\item{feaseobj}{An object of class 'Fease'}

\item{outy}{Determines whether you would like a full table or some column of
the table for a specific case of the feasibility object. When set equal to
table, the full table is produced. When set equal to an integer number the
names of MPs that are feasible for that case are returned.}
}
\description{
What MPs may be run (best case scenario) for various data-availability
scenarios?
}
\author{
T. Carruthers
}
