% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{Feasibility}
\alias{Feasibility}
\title{Feasibility DLM_fease}
\format{An object of class \code{DLM_fease} of length 1.}
\usage{
Feasibility
}
\description{
An object of class DLM_fease
}
\keyword{datasets}
