% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Reference.R
\name{FMSYref}
\alias{FMSYref}
\alias{FMSYref50}
\alias{FMSYref75}
\alias{NFref}
\title{Reference management procedures}
\usage{
FMSYref(x, Data, reps = 100)

FMSYref50(x, Data, reps = 100)

FMSYref75(x, Data, reps = 100)

NFref(x, Data, reps = 100)
}
\arguments{
\item{x}{A position in data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of TAC samples}
}
\description{
Several reference MPs for your operating model to use in the management strategy
evaluation. FMSYref (and related) assume perfect information about FMSY (FMSY
is taken from the operating model stored at Data@OM$FMSY). NFref sets annual catch
to zero (or close to it). Used for looking at variability in stock with no fishing.
}
\details{
Note that you can out-perform \code{FMSYref} easily. The requirement for fixed
F is actually quite strict and is by no means the upper limit in terms of
yield. Don't panic if your method beats this one for yield, especially for
short-lived species of high temporal variability in productivity!

\code{NFref} sets annual catch to 0.01.
}
\section{Functions}{
\itemize{
\item \code{FMSYref}: A reference FMSY method that fishes at FMSY

\item \code{FMSYref50}: A reference FMSY method that fishes at half of FMSY

\item \code{FMSYref75}: A reference FMSY method that fishes at 75\% of FMSY

\item \code{NFref}: A reference MP that sets annual catch to zero
(or very close to it).
}}

\author{
T. Carruthers, A. Hordyk
}
