% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output.r
\name{FMSYref}
\alias{FMSYref}
\title{A reference FMSY method (uses perfect information about FMSY)}
\usage{
FMSYref(x, Data, reps = 100)
}
\arguments{
\item{x}{A position in data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of TAC samples}
}
\description{
FMSY is taken from the operating model stored at DLM@OM$FMSY
}
\details{
Note that you can out-perform this MP even though it has perfect
information of FMSY and current abundance. The requirment for fixed F is
actually quite strict and is by no means the upper limit in terms of yield.
Don't panic if your method beats this one for yield, especially for
short-lived species of high temporal variability in productivity!
}
\author{
T. Carruthers
}
