% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output.r
\name{Fratio}
\alias{Fratio}
\title{An FMSY/M ratio method}
\usage{
Fratio(x, Data, reps = 100)
}
\arguments{
\item{x}{A position in a data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of samples of the TAC recommendation}
}
\value{
A numeric vector of TAC recommendations
}
\description{
Calculates the OFL based on a fixed ratio of FMSY to M multiplied by a
current estimate of abundance.
}
\details{
A simple method that tends to outperform many other approaches alarmingly
often even when current biomass is relatively poorly known. The low stock
crash potential is largely due to the quite large difference between Fmax
and FMSY for most stocks.
}
\references{
Gulland, J.A., 1971. The fish resources of the ocean. Fishing
News Books, West Byfleet, UK.

Martell, S., Froese, R., 2012. A simple method for estimating MSY from catch
and resilience. Fish Fish. doi: 10.1111/j.1467-2979.2012.00485.x.
}
\author{
T. Carruthers
}
