% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output.r
\name{Fratio4010}
\alias{Fratio4010}
\title{An FMSY/M ratio method paired with the 40-10 rule}
\usage{
Fratio4010(x, Data, reps = 100)
}
\arguments{
\item{x}{A position in data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of TAC samples}
}
\description{
Calculates the OFL based on a fixed ratio of FMSY to M multiplied by a
current estimate of abundance. In this method DBSRA is paired with the 40-10
rule that throttles back the OFL to zero at 10 percent of unfished biomass.
}
\references{
Gulland, J.A., 1971. The fish resources of the ocean. Fishing
News Books, West Byfleet, UK.

Martell, S., Froese, R., 2012. A simple method for estimating MSY from catch
and resilience. Fish Fish. doi: 10.1111/j.1467-2979.2012.00485.x.
}
\author{
T. Carruthers
}
