% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLM_output.R
\docType{data}
\name{GB_slope}
\alias{GB_slope}
\title{Geromont and Butterworth index slope Harvest Control Rule}
\format{An object of class \code{DLM_output} of length 1.}
\usage{
GB_slope(x, DLM_data, reps = 100, yrsmth = 5, lambda = 1)
}
\arguments{
\item{x}{A position in data-limited methods data object}

\item{DLM_data}{A data-limited methods data object}

\item{reps}{The number of TAC samples}

\item{yrsmth}{Number of years for evaluating slope in relative abundance
index}

\item{lambda}{A gain parameter}
}
\description{
An MP similar to SBT1 that modifies a time-series of catch recommendations
and aims for a stable catch rates.
}
\details{
Note that this is my interpretation of their approach and is now stochastic.
Currently it is generalized and is not 'tuned' to more detailed assessment
data which might explain why in some cases it leads to stock declines.
}
\author{
T. Carruthers
}
\references{
Geromont, H.F. and Butterworth, D.S. 2014. Complex assessment or
simple management procedures for efficient fisheries management: a
comparative study. ICES J. Mar. Sci. doi:10.1093/icesjms/fsu017
}
\keyword{datasets}

