% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{GB_target}
\alias{GB_target}
\title{Geromont and Butterworth target CPUE and catch MP}
\usage{
GB_target(x, Data, reps = 100, w = 0.5)
}
\arguments{
\item{x}{A position in data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of quota samples}

\item{w}{A gain parameter}
}
\description{
An MP similar to SBT2 that modifies a time-series of catch recommendations
and aims for target catch rate and catch level based on BMSY/B0 and MSY,
respectively.
}
\details{
Note that this is my interpretation of their MP and is now stochastic.
Currently it is generalized and is not 'tuned' to more detailed assessment
data which might explain why in some cases it leads to stock declines.
}
\references{
Geromont, H.F. and Butterworth, D.S. 2014. Complex assessment or
simple management procedures for efficient fisheries management: a
comparative study. ICES J. Mar. Sci. doi:10.1093/icesjms/fsu017
}
\author{
T. Carruthers
}
