% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{Gcontrol}
\alias{Gcontrol}
\title{G-control MP}
\usage{
Gcontrol(x, Data, reps = 100, yrsmth = 10, gg = 2, glim = c(0.5,
2))
}
\arguments{
\item{x}{A position in data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of quota samples}

\item{yrsmth}{The number of years over which to smooth catch and biomass
data}

\item{gg}{A gain parameter}

\item{glim}{A constraint limiting the maximum level of change in quota
recommendations}
}
\description{
A harvest control rule proposed by Carl Walters that uses trajectory in
inferred surplus production to make upward/downward adjustments to TAC
recommendations
}
\references{
Made-up for this package. Carruthers et al. 2015. Performance of
Simple Management Procedures.
}
\author{
C. Walters and T. Carruthers
}
