% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{Generic_fleet}
\alias{Generic_fleet}
\title{Generic_fleet Fleet}
\format{An object of class \code{Fleet} of length 1.}
\usage{
Generic_fleet
}
\description{
An object of class Fleet
}
\keyword{datasets}
