% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{Generic_Obs}
\alias{Generic_Obs}
\title{Generic_Obs Obs}
\format{An object of class \code{Obs} of length 1.}
\usage{
Generic_Obs
}
\description{
An object of class Obs
}
\keyword{datasets}
