% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscInt.r
\name{getSlope1}
\alias{getSlope1}
\title{Internal function to calculate slope}
\usage{
getSlope1(s1, L1, L0.05)
}
\arguments{
\item{s1}{s1}

\item{L1}{length one}

\item{L0.05}{length at 5 percent selection}
}
\description{
Internal function to calculate slope
}
\keyword{internal}

