% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscInt.r
\name{getSlope2}
\alias{getSlope2}
\title{Internal function to calculate slope}
\usage{
getSlope2(s2, LFS, s1, maxlen, MaxSel)
}
\arguments{
\item{s2}{desceding slope}

\item{LFS}{length one}

\item{s1}{ascending slope}

\item{maxlen}{length of oldest age class}

\item{MaxSel}{selectivity of maxlen}
}
\description{
Internal function to calculate slope
}
\keyword{internal}

