% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_Internal.R
\name{gettempvar}
\alias{gettempvar}
\title{Creates a time series per simulation that has gradient grad and random normal walk with sigma}
\usage{
gettempvar(targ, targsd, targgrad, nyears, nsim, rands = NULL)
}
\arguments{
\item{targ}{mean}

\item{targsd}{standard deviation}

\item{targgrad}{gradient}

\item{nyears}{number of historical years}

\item{nsim}{number of simulations}
}
\description{
Creates a time series per simulation that has gradient grad and random normal walk with sigma
}
\keyword{internal}
