% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{Imprecise_Biased}
\alias{Imprecise_Biased}
\title{Imprecise_Biased Observation}
\format{An object of class \code{Observation} of length 1.}
\usage{
Imprecise_Biased
}
\description{
An object of class Observation
}
\keyword{datasets}
