% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{Imprecise_Unbiased}
\alias{Imprecise_Unbiased}
\title{Imprecise_Unbiased Observation}
\format{An object of class \code{Observation} of length 1.}
\usage{
Imprecise_Unbiased
}
\description{
An object of class Observation
}
\keyword{datasets}
