% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Functions.R
\name{Input}
\alias{Input}
\title{Function to run a set of input control methods}
\usage{
Input(Data, MPs = NA, reps = 100, timelimit = 10, CheckMPs = TRUE, msg = TRUE)
}
\arguments{
\item{Data}{A Data object}

\item{MPs}{A list of input MPs, if NA all available input MPs are run}

\item{reps}{Number of repetitions (for those methods that use them)}

\item{timelimit}{Maximum timelimit to run MP (in seconds)}

\item{CheckMPs}{Logical, the Can function is run if this is TRUE}

\item{msg}{Logical. Should messages be printed?}
}
\description{
Runs a set of input control methods are returns the output in a single table
}
\examples{
Input(DLMtool::Cobia)
}
\author{
A. Hordyk
}
