% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSCAM2DLM.R
\name{iSCAM2DLM}
\alias{iSCAM2DLM}
\title{Reads MLE estimates from Stock Synthesis file structure into an operating model using package r4ss}
\usage{
iSCAM2DLM(iSCAMdir, nsim = 48, proyears = 50, Name = NULL,
  Source = "No source provided", Author = "No author provided")
}
\arguments{
\item{nsim}{The number of simulations to take for parameters with uncertainty (for OM@cpars custom parameters)}

\item{Name}{The name of the operating model}

\item{Source}{Reference to assessment documentation e.g. a url}

\item{Author}{Who did the assessment}

\item{SSdir}{A folder with Stock Synthesis input and output files in it}
}
\description{
A function that uses the file location of a fitted SS3 model including input files to population the various slots of an operating model with MLE parameter estimates
}
\author{
T. Carruthers
}

