% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output.r
\name{IT10}
\alias{IT10}
\title{Index Target 10}
\usage{
IT10(x, Data, reps = 100,yrsmth=5,mc=0.1)
}
\arguments{
\item{x}{A position in a data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of stochastic samples of the quota recommendation}

\item{yrsmth}{The number of historical years over which to average the index}

\item{mc}{The maximum fractional change in the TAC among years.}
}
\value{
A numeric vector of TAC recommendations
}
\description{
An index target MP where the TAC is modified according to current index
levels (mean index over last 5 years) relative to a target level. Maximum
annual changes are 10 per cent.
}
\author{
T. Carruthers
}
