% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Input.R, R/MPs_Output.R
\name{ItargetE1}
\alias{ItargetE1}
\alias{ItargetE4}
\alias{Itarget1}
\alias{Itarget4}
\title{A management procedure that incrementally adjusts the TAC (starting from
reference level that is a fraction of mean recent catches) or effort (from
a reference level) to reach a target CPUE / relative abundance index}
\usage{
ItargetE1(x, Data, reps = 100, yrsmth = 5, xx = 0, Imulti = 1.5)

ItargetE4(x, Data, reps = 100, yrsmth = 5, xx = 0, Imulti = 2.5)

Itarget1(x, Data, reps = 100, yrsmth = 5, xx = 0, Imulti = 1.5,
  plot = FALSE, ...)

Itarget4(x, Data, reps = 100, yrsmth = 5, xx = 0.3, Imulti = 2.5,
  plot = FALSE, ...)
}
\arguments{
\item{x}{A position in data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of TAC samples}

\item{yrsmth}{Years over which to smooth recent estimates of surplus
production}

\item{xx}{Parameter controlling the fraction of mean catch to start using in
first year}

\item{Imulti}{Parameter controlling how much larger target CPUE / index is
compared with recent levels.}

\item{plot}{Logical. Show the plot?}

\item{...}{other arguments passed to plotting function}
}
\value{
A Rec object
}
\description{
Four index/CPUE target MPs proposed by Geromont and Butterworth 2014.
Tested by Carruthers et al. 2015.
}
\section{Functions}{
\itemize{
\item \code{ItargetE1}: The less precautionary effort-based MP

\item \code{ItargetE4}: The most biologically precautionary effort-based MP

\item \code{Itarget1}: The less precautionary TAC-based MP

\item \code{Itarget4}: The most biologically precautionary TAC-based MP
}}

\references{
Carruthers et al. 2015. Performance evaluation of simple
management procedures. ICES J. Mar Sci. 73, 464-482.

Geromont, H.F., Butterworth, D.S. 2014. Generic management procedures for
data-poor fisheries; forecasting with few data. ICES J. Mar. Sci. 72, 251-261.
doi:10.1093/icesjms/fst232
}
\seealso{
Other Index methods: \code{\link{GB_slope}},
  \code{\link{GB_target}}, \code{\link{Gcontrol}}
}
\author{
T. Carruthers
}
