% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Mixed.R
\name{Itarget1_MPA}
\alias{Itarget1_MPA}
\title{A example mixed control MP that uses the Itarget1 output control MP together with a
spatial closure.}
\usage{
Itarget1_MPA(x, Data, reps = 100, yrsmth = 5, xx = 0, Imulti = 1.5)
}
\arguments{
\item{x}{A position in data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of TAC samples}

\item{yrsmth}{Years over which to smooth recent estimates of surplus
production}

\item{xx}{Parameter controlling the fraction of mean catch to start using in
first year}

\item{Imulti}{Parameter controlling how much larger target CPUE / index is
compared with recent levels.}
}
\description{
Included demonstration purposes of a mixed control MP
}
