% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Input.R
\name{ITe5}
\alias{ITe5}
\title{Index Target Effort-Based 5}
\usage{
ITe5(x, Data, reps = 100, yrsmth = 5, mc = 0.05)
}
\arguments{
\item{x}{A position in a data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of stochastic samples of the quota recommendation}

\item{yrsmth}{The number of historical years over which to average the index}

\item{mc}{The maximum fractional change in the effort among years.}
}
\value{
A numeric vector of input controls
}
\description{
An index target MP where the Effort is modified according to current index
levels (mean index over last 5 years) relative to a target level. Maximum
annual changes are 5 per cent.
}
\author{
T. Carruthers
}
