% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE_Plotting.R
\name{join_plots}
\alias{join_plots}
\title{Plot several plots with a shared legend}
\usage{
join_plots(
  plots,
  ncol = length(plots),
  nrow = 1,
  position = c("right", "bottom"),
  legend = TRUE
)
}
\arguments{
\item{plots}{list of plot objects of class \code{gg} or \code{ggplot}}

\item{ncol}{Optional number of columns}

\item{nrow}{Optional number of rows}

\item{position}{position of the legend ("bottom" or "right")}

\item{legend}{Logical. Use a legend?}
}
\description{
Plot several plots with a shared legend
}
\note{
modified from https://github.com/tidyverse/ggplot2/wiki/share-a-legend-between-two-ggplot2-graphs
}
