\name{Kplot}
\alias{Kplot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
KOBE plot: a projection by projection plot of F/FMSY and B/BMSY
}
\description{
A standard KOBE plot by each method that also shows the percentage of methods that ended up in each quadrant.
}
\usage{
Kplot(MSEobj,maxsim=60,nam=NA)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{MSEobj}{
An object of class MSE
}
\item{maxsim}{
Maximum number of simulations (lines) to plot on each panel. 
}
\item{nam}{
The name of the plot
}
}
\author{
T. Carruthers 
}
\note{
Apologies for the nauseating shading. 
}
