% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE_Plotting.R
\name{Kplot}
\alias{Kplot}
\title{KOBE plot: a projection by projection plot of F/FMSY and B/BMSY}
\usage{
Kplot(
  MSEobj,
  maxsim = 60,
  MPs = NA,
  sims = NULL,
  maxMP = 9,
  nam = NA,
  cex.leg = 1.5
)
}
\arguments{
\item{MSEobj}{An object of class MSE}

\item{maxsim}{Maximum number of simulations (lines) to plot on each panel.}

\item{MPs}{Optional subset MSE object by MP}

\item{sims}{Optional subset MSE object by simulation}

\item{maxMP}{Maximum number of MPs to include in plot}

\item{nam}{The name of the plot}

\item{cex.leg}{Size of legend}
}
\description{
A standard KOBE plot by each method that also shows the percentage of
methods that ended up in each quadrant.
}
\note{
Apologies for the nauseating shading.
}
\author{
T. Carruthers with some additions from A. Hordyk
}
