% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFuns.r
\name{L2A}
\alias{L2A}
\title{Length to age conversion}
\usage{
L2A(t0c,Linfc,Kc,Len,maxage)
}
\arguments{
\item{t0c}{Theoretical age at length zero}

\item{Linfc}{Maximum length}

\item{Kc}{Maximum growth rate}

\item{Len}{Length}

\item{maxage}{Maximum age}
}
\value{
An age (vector of ages, matrix of ages) corresponding with Len
}
\description{
Simple deterministic length to age conversion given inverse von Bertalanffy
growth.
}
\author{
T. Carruthers
}

