% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_Exported.R
\name{L2A}
\alias{L2A}
\title{Length to age conversion}
\usage{
L2A(t0c, Linfc, Kc, Len, maxage, ploty = F)
}
\arguments{
\item{t0c}{Theoretical age at length zero}

\item{Linfc}{Maximum length}

\item{Kc}{Maximum growth rate}

\item{Len}{Length}

\item{maxage}{Maximum age}

\item{ploty}{Should a plot be included}
}
\value{
An age (vector of ages, matrix of ages) corresponding with Len
}
\description{
Simple deterministic length to age conversion given inverse von Bertalanffy
growth.
}
\author{
T. Carruthers
}
\keyword{internal}
