% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LBSPR_MPs.r
\name{LBSPR_ItEff}
\alias{LBSPR_ItEff}
\title{Length-based SPR model with HCR that iteratively adjusts Effort}
\usage{
LBSPR_ItEff(x, DLM_data, yrsmth=1,reps=5, ...)
}
\arguments{
\item{x}{Simulation number}

\item{DLM_data}{DLM_data object}

\item{yrsmth}{Number of years to smooth length data - not currently used}

\item{reps}{Number of repetitions. Not currently used}

\item{...}{ignored}
}
\description{
Iteratively adjusts Effort based on distance between estimated and target
SPR (40\%), and slope of recent SPR estimates.
}
\author{
A. Hordyk
}
