% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StochasticSRA.R
\name{LSRA_opt}
\alias{LSRA_opt}
\title{Internal estimation function for LSRA and LSRA2 functions}
\usage{
LSRA_opt(param, FF_a, Chist, M_a, Mat_age_a, Wt_age_a, sel_a, Recdevs_a, h_a,
  Umax = 0.2, mode = 1)
}
\arguments{
\item{param}{a numeric value representing log(R0)}

\item{FF_a}{numeric value, recent fishign mortality rate (apical F)}

\item{Chist}{a vector of historical catch observations [nyears]}

\item{M_a}{numeric value, natural mortality rate}

\item{Mat_age_A}{a vector of maturity at age [nage]}

\item{Wt_age}{a vector of weight at age [nage]}

\item{sel}{a vector of selectivity at age [nage]}

\item{Recdevs}{a vector of recruitment deviations [nyears]}

\item{h}{a numeric value of steepness values of the Bev-Holt Stock-Recruitment relationship}
}
\value{
depends on mode but could be 1:objective function 2:trajectory of Fs 3: SSB depletion 4:log(R0) 5:diagnostic plots
}
\description{
Internal estimation function for LSRA and LSRA2 functions
}
\author{
T. Carruthers
}

