% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Input.R, R/MPs_Output.R
\name{LtargetE1}
\alias{LtargetE1}
\alias{LtargetE4}
\alias{Ltarget1}
\alias{Ltarget4}
\title{A management procedure that incrementally adjusts the TAC or effort to reach
a target mean length in catches.}
\usage{
LtargetE1(x, Data, reps = 100, yrsmth = 5, xx = 0, xL = 1.05)

LtargetE4(x, Data, reps = 100, yrsmth = 5, xx = 0, xL = 1.15)

Ltarget1(x, Data, reps = 100, yrsmth = 5, xx = 0, xL = 1.05)

Ltarget4(x, Data, reps = 100, yrsmth = 5, xx = 0.2, xL = 1.15)
}
\arguments{
\item{x}{A position in data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of TAC samples}

\item{yrsmth}{Years over which to smooth recent estimates of surplus
production}

\item{xx}{Parameter controlling the fraction of mean catch to start using in
first year}

\item{xL}{Parameter controlling the magnitude of the target mean length of
catches relative to average length in catches.}
}
\value{
A \linkS4class{Rec} object
}
\description{
Four target length MPs proposed by Geromont and Butterworth 2014.
Tested by Carruthers et al. 2015.
}
\section{Functions}{
\itemize{
\item \code{LtargetE1}: The least biologically precautionary effort-based MP.

\item \code{LtargetE4}: The most biologically precautionary effort-based MP.

\item \code{Ltarget1}: The least biologically precautionary TAC-based MP.

\item \code{Ltarget4}: The most biologically precautionary TAC-based MP.
}}

\references{
Carruthers et al. 2015. Performance evaluation of simple
management procedures. ICES J. Mar Sci. 73, 464-482.

Geromont, H.F., Butterworth, D.S. 2014. Generic management procedures for
data-poor fisheries; forecasting with few data. ICES J. Mar. Sci.
doi:10.1093/icesjms/fst232
}
\seealso{
\link{L95target}
}
\author{
T. Carruthers
}
