% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMobjects.R
\name{LTY}
\alias{LTY}
\title{Performance Metric: Probability Long-Term Yield > 0.5 Relative Yield}
\usage{
LTY(MSEobj = NULL)
}
\arguments{
\item{MSEobj}{An object of class MSE}
}
\value{
An object of class PMobj
}
\description{
Performance Metric: Probability Long-Term Yield > 0.5 Relative Yield
}
\examples{
\dontrun{
LTY(myMSE)
}
}
