% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_Exported.R
\name{makePerf}
\alias{makePerf}
\title{Convert a OM object to one without observation or process error}
\usage{
makePerf(OMin, except = NULL)
}
\arguments{
\item{OMin}{An object of class \code{OM}}

\item{except}{An optional vector of slot names in the OM that will not be
changed (not tested perfectly so watch out!)}
}
\value{
A new \code{OM} object
}
\description{
Note: This function has been replaced with \code{tinyErr} and will soon be removed from
the package
}
\details{
Takes an existing OM object and converts it to one without any observation
error, and very little process error.  Used for debugging and testing that
MPs perform as expected under perfect conditions.
}
\author{
A. Hordyk
}
