% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{matlenlim}
\alias{matlenlim}
\title{A data-limited method in which fishing retention is set according to the
maturity curve NEW}
\usage{
matlenlim(x, Data, ...)
}
\arguments{
\item{x}{A position in a data-limited methods object}

\item{Data}{A data-limited methods object}

\item{...}{Optional additional arguments that are ignored. Note arguments
\code{reps} or \code{...} are required for all input controls}
}
\value{
A input control recommendation object
}
\description{
An example of the implementation of input controls in the DLM toolkit, where
retention-at-length is set equivalent to maturity-at-length
}
\references{
Made-up for this package
}
\author{
T. Carruthers
}
