% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLM_input.R
\name{matlenlim}
\alias{matlenlim}
\title{A data-limited method in which fishing vulnerability is set according to the
maturity curve}
\usage{
matlenlim(x, DLM_data, ...)
}
\arguments{
\item{x}{A position in a data-limited methods object}

\item{DLM_data}{A data-limited methods object}

\item{...}{Optional additional arguments that are ignored. Note arguments
\code{reps} or \code{...} are required for all input controls}
}
\value{
A vector of input control recommendations, with values for length at
first capture and full selection
}
\description{
An example of the implementation of input controls in the DLM toolkit, where
selectivity-at-length is set equivalent to maturity-at-length
}
\references{
Made-up for this package
}
\author{
T. Carruthers
}
