% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Output.R
\name{MCD}
\alias{MCD}
\title{Mean Catch Depletion}
\usage{
MCD(x, Data, reps = 100)
}
\arguments{
\item{x}{A position in a data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of stochastic samples of the quota recommendation}
}
\value{
A numeric vector of TAC recommendations
}
\description{
A simple average catch-depletion MP that was included to demonstrate just
how informative an estimate of current stock depletion can be. TAC=2*D*AvC
}
\author{
T. Carruthers
}
