% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{MCD4010}
\alias{MCD4010}
\title{Mean Catch Depletion}
\usage{
MCD4010(x, Data, reps = 100)
}
\arguments{
\item{x}{A position in a data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of stochastic samples of the quota recommendation}
}
\value{
A numeric vector of TAC recommendations
}
\description{
A simple average catch-depletion MP linked to a 40-10 harvest controle rule
that was included to demonstrate just how informative an estimate of current
stock depletion can be. TAC=d(1-d)AvC
}
\author{
T. Carruthers
}
