% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_Exported.R
\name{ML2D}
\alias{ML2D}
\title{Depletion and F estimation from mean length of catches}
\usage{
ML2D(OM, ML, nsim = 100, ploty = T, Dlim = c(0.05, 0.6))
}
\arguments{
\item{OM}{An object of class 'OM'}

\item{ML}{A estimate of current mean length of catches}

\item{nsim}{Number of simulations}

\item{ploty}{Produce a plot of depletion and F}

\item{Dlim}{Limits on the depletion that is returned as a fraction of
unfished biomass.}
}
\value{
An object of class 'OM' with 'D' slot populated
}
\description{
A highly dubious means of getting very uncertain estimates of current stock
biomass and (equilibrium) fishing mortality rate from growth, natural
mortality rate, recruitment and fishing selectivity.
}
\author{
T. Carruthers
}
