% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS2DLM.R
\name{movdistil}
\alias{movdistil}
\title{Simplified a multi-area transition matrix into the best 2 x 2 representation}
\usage{
movdistil(movtab)
}
\arguments{
\item{movtab}{a table of estimated movements}
}
\description{
A Function that takes a larger movement matrix, identifies the most parsimonious representation of 2 non-mixed areas and returns the final unfished movement matrix
}
\author{
T. Carruthers
}
