% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE_Plotting.R
\name{MPStats}
\alias{MPStats}
\title{Calculate Statistics for MP Performance}
\usage{
MPStats(
  MSEobj,
  PMRefs = list(B_BMSY = 0.5, SSB_SSB0 = 0.2, F_FMSY = 1, AAVY = 30, AAVE = 30),
  lastYrs = 10,
  UseMean = TRUE,
  msg = TRUE
)
}
\arguments{
\item{MSEobj}{An object of class MSE}

\item{PMRefs}{A list of reference points for the performance metrics (must
be named)}

\item{lastYrs}{The last number of years in the projection to calculate the
statistics}

\item{UseMean}{Logical. Calculate mean (TRUE) or median (FALSE)?}

\item{msg}{Logical. Print out messages?}
}
\description{
Function calculates probabilities and other statistics for a range of
performance metrics
}
\author{
A. Hordyk
}
