% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input.r
\name{MRnoreal}
\alias{MRnoreal}
\title{An marine reserve in area 1 with no spatial reallocation of fishing effort}
\usage{
MRnoreal(x, Data, ...)
}
\arguments{
\item{x}{A position in data / simulation object DLM}

\item{Data}{A data limited methods data object}

\item{...}{Optional additional arguments that are ignored. Note arguments
\code{reps} or \code{...} are required for all input controls}
}
\description{
A spatial control that prevents fishing in area 1 and does not reallocate
this fishing effort to area 2.
}
\author{
T. Carruthers
}
