% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Input.R
\name{MRreal}
\alias{MRreal}
\title{A marine reserve in area 1 with full reallocation of fishing effort}
\usage{
MRreal(x, Data, ...)
}
\arguments{
\item{x}{A position in data / simulation object DLM}

\item{Data}{A data limited methods data object}

\item{...}{Optional additional arguments that are ignored. Note arguments
\code{reps} or \code{...} are required for all input controls}
}
\description{
A spatial control that prevents fishing in area 1 and reallocates this
fishing effort to area 2 (or over other areas).
}
\author{
T. Carruthers
}
