\name{MRreal}
\alias{MRreal}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
An marine reserve in area 1 with full reallocation of fishing effort
}
\description{
A spatial control that prevents fishing in area 1 and reallocates this fishing effort to area 2.  
}
\usage{
MRreal(x, DLM_data, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A position in data / simulation object DLM
}
  \item{DLM_data}{
A data limited methods data object
}
\item{...}{Optional additional arguments that are ignored. Note arguments \code{reps} or \code{...} are required for all input controls }
}
\author{
T. Carruthers
}


