% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_Exported.R
\name{NAor0}
\alias{NAor0}
\title{Is a value NA or zero.}
\usage{
NAor0(x)
}
\arguments{
\item{x}{A numeric value.}
}
\value{
TRUE or FALSE
}
\description{
As title
}
\author{
T. Carruthers
}
\keyword{internal}
