\name{Needed}
\alias{Needed}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Data needed to get DLM methods running
}
\description{
Wrapper function for DLMdiag that lists what data are needed to run data-limited methods that are current not able to run given a DLM data object
}
\usage{
Needed(DLM, timelimit=1)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{DLM}{
A data-limited methods data object
}
  \item{timelimit}{
The maximum time (seconds) taken to complete 10 reps
}
}
\author{
T. Carruthers
}


