% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataObj.r
\name{Needed}
\alias{Needed}
\title{Data needed to get MPs running}
\usage{
Needed(Data, timelimit=1)
}
\arguments{
\item{Data}{A data-limited methods data object}

\item{timelimit}{The maximum time (seconds) taken to complete 10 reps}
}
\description{
Wrapper function for DLMdiag that lists what data are needed to run
data-limited methods that are current not able to run given a DLM_cdata
object
}
\author{
T. Carruthers
}

