\name{Needed}
\alias{Needed}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Data needed to get MPs running}
\description{
Wrapper function for DLMdiag that lists what data are needed to run data-limited methods that are current not able to run given a DLM_cdata object
}
\usage{
Needed(DLM_data, timelimit=1)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{DLM_data}{
A data-limited methods data object
}
  \item{timelimit}{
The maximum time (seconds) taken to complete 10 reps
}
}
\author{
T. Carruthers
}


