% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_Exported.R
\name{optCPU}
\alias{optCPU}
\title{Determine optimal number of cpus}
\usage{
optCPU(thresh = 5, plot = TRUE, nsim = 48)
}
\arguments{
\item{thresh}{Recommended n cpus is what percent of the fastest time?}

\item{plot}{Logical. Show the plot?}

\item{nsim}{Numeric. Number of simulations.}
}
\description{
Determine optimal number of cpus
}
\author{
A. Hordyk
}
