% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_Exported.R
\name{optCPU}
\alias{optCPU}
\title{Determine optimal number of cpus}
\usage{
optCPU(nsim = 96, thresh = 5, plot = TRUE, msg = TRUE, maxn = NULL)
}
\arguments{
\item{nsim}{Numeric. Number of simulations.}

\item{thresh}{Recommended n cpus is what percent of the fastest time?}

\item{plot}{Logical. Show the plot?}

\item{msg}{Logical. Should messages be printed to console?}

\item{maxn}{Optional. Maximum number of cpus. Used for demo purposes}
}
\description{
Determine optimal number of cpus
}
\note{
See relevant section of the \href{https://dlmtool.github.io/DLMtool/userguide/parallel-processing.html#determining-optimal-number-of-processors}{DLMtool User Guide} for more information.
}
\examples{
\dontrun{
optCPU()
}
}
\seealso{
\link{setup}
}
\author{
A. Hordyk
}
