% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineClasses.r, R/dataObjects.r
\docType{data}
\name{ourReefFish}
\alias{ourReefFish}
\alias{ourReefFish}
\title{Example data object}
\format{An object of class \code{Data} of length 1.}
\usage{
data('ourReefFish')

ourReefFish
}
\description{
Example data object with a number of output control MPs run on it, and
includes resulting distributions of TACs

An object of class Data
}
\examples{
\dontrun{ 
data(ourReefFish)
str(ourReefFish)  
plot(ourReefFish) 
}

}
\keyword{datasets}
