% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMobjects.R
\name{PerformanceMetric}
\alias{PerformanceMetric}
\alias{P10}
\alias{P50}
\alias{P100}
\alias{POF}
\alias{LTY}
\alias{STY}
\alias{Yield}
\alias{AAVY}
\title{Performance Metrics}
\usage{
P10(MSEobj = NULL, Ref = 0.1, Yrs = NULL)

P50(MSEobj = NULL, Ref = 0.5, Yrs = NULL)

P100(MSEobj = NULL, Ref = 1, Yrs = NULL)

POF(MSEobj = NULL, Ref = 1, Yrs = NULL)

LTY(MSEobj = NULL, Ref = 0.5, Yrs = c(max(MSEobj@proyears - 9, 1),
  MSEobj@proyears))

STY(MSEobj = NULL, Ref = 0.5, Yrs = c(1, 10))

Yield(MSEobj = NULL, Ref = 1, Years = NULL)

AAVY(MSEobj = NULL, Ref = 0.2, Yrs = NULL)
}
\arguments{
\item{MSEobj}{An object of class MSE}

\item{Ref}{Reference point for calculating the performance metric. See details.}

\item{Yrs}{Numeric vector of length 2 with year indices to summarize performance.
If NULL, the performance is summarized over all projection years.}
}
\value{
An object of class PMobj
}
\description{
Performance metrics (PMs) for your management strategy evaluation.
}
\details{
Performance Metric definitions:

\tabular{ll}{
\code{P10} \tab Probability B > 0.1 BMSY \cr
\code{P50} \tab Probability B > 0.5 BMSY \cr
\code{P100} \tab Probability B > BMSY \cr
\code{POF} \tab Probability F < FMSY \cr
\code{LTY} \tab Probability Long-Term Yield > 0.5 Relative Yield \cr
\code{STY} \tab Probability Short-Term Yield > 0.5 Relative Yield \cr
\code{AAVY} \tab Probability AAVY < 0.2 (Average Annual Variability in Yield) \cr
\code{Yield} \tab Average Yield (relative to Reference Yield) \cr
}

Argument \code{Ref} provides the ratio relative to the reference point for calculating
the performance metric. For biomass-based PMs (P10, P50, P100), this is the fraction of
BMSY. For POF, the fraction of FMSY. For Yield (and LTY/STY), the fraction of the
Reference Yield.

Long-Term Yield is the Yield in the last ten years of the projection period in the MSE.
Short-Term Yield is that in the first 10 years of the projection period
}
\examples{
\dontrun{
P10(myMSE)
P50(myMSE)
P100(myMSE)
POF(myMSE)
LTY(myMSE)
STY(myMSE)
AAVY(myMSE)
Yield(myMSE)
}
}
