% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineClasses.r
\docType{methods}
\name{plot,Data,ANY-method}
\alias{plot,Data,ANY-method}
\title{Plot Data object}
\usage{
\S4method{plot}{Data,ANY}(x, y, ...)
}
\arguments{
\item{x}{object of class Data}

\item{funcs}{MPs}

\item{maxlines}{maximum number of lines}

\item{perc}{percentile of TAC recommendation}

\item{xlims}{limits of x-axis}
}
\description{
Plot Data object
}
