% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineClasses.r
\docType{methods}
\name{plot,Data,ANY-method}
\alias{plot,Data,ANY-method}
\title{Plot Data object}
\usage{
\S4method{plot}{Data,ANY}(x, upq = 0.9, lwq = 0.1, outline = FALSE, ...)
}
\arguments{
\item{x}{object of class Data}

\item{upq}{Upper quantile of TACs for max ylim}

\item{lwq}{Lower quantile of TACs for min ylim}

\item{outline}{Logical. Include outliers in plot?}

\item{...}{Optional additional arguments passed to \code{boxplot}}
}
\description{
Plot Data object
}
