% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineClasses.r
\docType{methods}
\name{plot,OM,ANY-method}
\alias{plot,OM,ANY-method}
\title{Plot OM object}
\usage{
\S4method{plot}{OM,ANY}(x)
}
\arguments{
\item{x}{object of class OM}
}
\description{
Plot OM object
}

