% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_Plotting.R
\name{plot.Stock}
\alias{plot.Stock}
\title{Title}
\usage{
\method{plot}{Stock}(x, nsamp = 3, nsim = 200, nyears = 50,
  proyears = 28, output_file = NULL, output_dir = getwd(),
  quiet = TRUE, tabs = TRUE, title = NULL, date = NULL,
  plotPars = NULL, open = TRUE, dev = FALSE)
}
\arguments{
\item{x}{}

\item{nsamp}{}

\item{nsim}{}

\item{nyears}{}

\item{proyears}{}

\item{output_file}{}

\item{output_dir}{}

\item{quiet}{}

\item{tabs}{}

\item{title}{}

\item{date}{}

\item{plotPars}{}

\item{open}{}

\item{dev}{}
}
\description{
Title
}
