% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscFuns.r
\name{plotFun}
\alias{plotFun}
\title{Print out plotting functions}
\usage{
plotFun(class = c('MSE', 'Data'), msg=TRUE)
}
\arguments{
\item{class}{Character string. Prints out the plotting functions for objects
of this class.}

\item{msg}{Logical. Should the functions be printed to screen?}
}
\description{
This function prints out the available plotting functions for objects of
class MSE or Data
}
\note{
Basically the function looks for any functions in the DLMtool that
have the word `plot` in them.  There is a chance that some plotting
functions are missed. Let us know if you find any and we will add them.
}
\author{
A. Hordyk
}
