\name{plotOFL}
\alias{plotOFL}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
A generic OFL plot for NOAA use
}
\description{
As title. 
}
\usage{
plotOFL(DLM_data,xlims=NA,perc=0.5)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{DLM_data}{
An object of class DLM_data that has been run though TAC()
}
\item{xlims}{
x axis limits
}
\item{perc}{
The percentile of the OFL distribution to be plotted
}
}
\value{
A table of performance metrics.
}
\author{
T. Carruthers
}

