% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Functions.R
\name{plotOFL}
\alias{plotOFL}
\title{A generic OFL plot for NOAA use}
\usage{
plotOFL(Data, xlims = NA, perc = 0.5)
}
\arguments{
\item{Data}{An object of class Data that has been run though TAC()}

\item{xlims}{x axis limits}

\item{perc}{The percentile of the OFL distribution to be plotted}
}
\value{
A table of performance metrics.
}
\description{
As title.
}
\author{
T. Carruthers
}
