% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_Plotting.R
\name{plotSelect}
\alias{plotSelect}
\title{Plot the vulnerability and retention curves}
\usage{
plotSelect(OM, Pars = NULL, pyears = 4, sim = NA, type = "l")
}
\arguments{
\item{OM}{An object of class 'OM'}

\item{Pars}{Named list of sampled parameters}

\item{pyears}{number of years to plot}

\item{sim}{the simulation to plot. default is NA to plot a random simulation.
Set to 1 for reproducible plot}

\item{type}{plot type - line "l", point "p", or both "b"}
}
\description{
Plot the vulnerability and retention curves
}
\author{
A. Hordyk
}
