% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMtable.R
\name{PMObj}
\alias{PMObj}
\title{Create a table of Performance Objectives}
\usage{
PMObj(MSE, ..., out.file = NULL, output_format = "html_document",
  openFile = TRUE, quiet = TRUE, dir = NULL, RMDfile = NULL,
  font_size = 14, full_width = TRUE, enableSearch = TRUE,
  PMlist = NULL, build = TRUE)
}
\arguments{
\item{MSE}{An object of class 'MSE'}

\item{...}{PM objects to be used as performance metrics. Characters (i.e names of PM objects)}

\item{out.file}{Name of the output file. If none provided, output file will be named 'PerfLimTable'}

\item{output_format}{Output file format. Currently only 'html_document' is supported}

\item{openFile}{Logical. Should the file be opened in browser?}

\item{quiet}{Logical. An option to suppress printing of the pandoc command line.}

\item{dir}{Optional. Directory for output file. Default is working directory.}

\item{RMDfile}{Optional. RMD template file}

\item{font_size}{Numeric. Font size for text in the table}

\item{full_width}{Logical. Should table be full width?}

\item{enableSearch}{Logical. Should search be enabled in the html table?}

\item{PMlist}{Optional. List of PM names.}

\item{build}{Logical. Build the html table?}
}
\value{
Nothing is returned.
}
\description{
Create a table of Performance Objectives
}
