% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSEObj.r
\name{Pplot}
\alias{Pplot}
\title{A projection by projection plot of F/FMSY and B/BMSY}
\usage{
Pplot(MSEobj, nam=NA)
}
\arguments{
\item{MSEobj}{An object of class MSE}

\item{nam}{Title of plot}
}
\description{
A shorter version of the plot method for MSEs that just shows the projected
trends in stock status and over exploitation
}
\author{
T. Carruthers
}

