% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Output.R
\name{Rcontrol2}
\alias{Rcontrol2}
\title{MP using prior for intrinsic rate of increase with a quadratic approximation
to surplus production}
\usage{
Rcontrol2(x, Data, reps = 100, yrsmth = 10, gg = 2, glim = c(0.5, 2))
}
\arguments{
\item{x}{A position in data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of TAC samples}

\item{yrsmth}{The number of years for smoothing catch and biomass data}

\item{gg}{A gain parameters}

\item{glim}{Limits for the change in TAC among years}
}
\description{
An MP proposed by Carl Walters that modifies quotas according to trends in
apparent surplus production that includes information from a demographically
derived prior for intrinsic rate of increase. This is different from
Rcontrol because it includes a quadratic approximation of recent trend in
surplus production given biomass
}
\references{
Made-up for this package.
}
\author{
C. Walters and T. Carruthers
}
