% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{Red_snapper}
\alias{Red_snapper}
\title{Red_snapper DLM_data}
\format{An object of class \code{DLM_data} of length 1.}
\usage{
Red_snapper
}
\description{
An object of class DLM_data
}
\keyword{datasets}
