% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Functions.R
\name{replic8}
\alias{replic8}
\title{Enlarge (replicate) a DLM data object to create an additional dimension for
simulation / sensitivity testing}
\usage{
replic8(Data, nrep)
}
\arguments{
\item{Data}{A data-limited methods data object}

\item{nrep}{The number of positions to expand the DLM object to}
}
\description{
Replicates position 1 data to multiple positions for sensitivity testing etc
}
\author{
T. Carruthers
}
