\name{Required}
\alias{Required}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{What methods need what data}
\description{
A function that finds all methods in the environment and searches the function text for slots in the DLM data object
}
\usage{
Required(funcs = NA)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{funcs}{
A character vector of possible methods of class DLM quota, DLM space or DLM size
}
}
\author{
T. Carruthers
}
