% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_Exported.R
\name{Required}
\alias{Required}
\title{What methods need what data}
\usage{
Required(funcs = NA)
}
\arguments{
\item{funcs}{A character vector of possible methods of class DLM quota, DLM
space or DLM size}
}
\description{
A function that finds all methods in the environment and searches the
function text for slots in the DLM data object
}
\author{
T. Carruthers
}
