% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataObj.r
\name{runInMP}
\alias{runInMP}
\title{Runs input control MPs on a DLM_data object.}
\usage{
runInMP(DLM_data, MPs = NA, reps = 100)
}
\arguments{
\item{DLM_data}{A object of class DLM_data}

\item{MPs}{A vector of MPs of class 'DLM_input'}

\item{reps}{Number of stochastic repititions - often not used in input
control MPs.}
}
\description{
Function runs a MP (or MPs) of class 'DLM_input' and returns a list: input
control recommendation(s) in element 1 and DLM_data object in element 2.
}
\author{
A. Hordyk
}
