% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popdyn.R
\name{runInMP}
\alias{runInMP}
\title{Runs input control MPs on a Data object.}
\usage{
runInMP(Data, MPs = NA, reps = 100)
}
\arguments{
\item{Data}{A object of class Data}

\item{MPs}{A vector of MPs of class 'Input'}

\item{reps}{Number of stochastic repititions - often not used in input
control MPs.}
}
\description{
Function runs a MP (or MPs) of class 'Input' and returns a list: input
control recommendation(s) in element 1 and Data object in element 2.
}
\author{
A. Hordyk
}
