% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleOM.R
\name{SampleCpars}
\alias{SampleCpars}
\title{Sample custom pars}
\usage{
SampleCpars(cpars, nsim = 48, msg = TRUE)
}
\arguments{
\item{cpars}{A named list containing custom parameters for the OM}

\item{nsim}{number of simulations}

\item{msg}{logical - print the names of the cpars? Turn off when using the function in a loop}
}
\value{
A named list of sampled custom parameters
}
\description{
Sample custom pars
}
\keyword{internal}
