% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleOM.R
\name{SampleObsPars}
\alias{SampleObsPars}
\title{Sample Observation Parameters}
\usage{
SampleObsPars(Obs, nsim = NULL, cpars = NULL)
}
\arguments{
\item{Obs}{An object of class 'Obs' or class 'OM'}

\item{nsim}{Number of simulations. Ignored if 'Obs' is class 'OM'}

\item{cpars}{Optional named list of custom parameters. Ignored if 'OM' is class 'OM'}
}
\value{
A named list of sampled Observation parameters
}
\description{
Sample Observation Parameters
}
\keyword{internal}
