% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{SBT2}
\alias{SBT2}
\title{SBT complex MP}
\usage{
SBT2(x, Data, reps = 100,
epsB=0.25,epsR=0.75,tauR=5,tauB=7,gamma=1)
}
\arguments{
\item{x}{A position in a data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of samples of the TAC}

\item{epsB}{Control parameter}

\item{epsR}{Control parameter}

\item{tauR}{Control parameter}

\item{tauB}{Control parameter}

\item{gamma}{Control parameter}
}
\description{
An MP that makes incremental adjustments to TAC recommendations based on
index levels relative to target levels (BMSY/B0) and catch levels relative
to target levels (MSY)
}
\details{
This isn't exactly the same as the proposed methods and is stochastic in
this implementation. The method doesn't tend to work too well under many
circumstances possibly due to the lack of 'tuning' that occurs in the real
SBT assessment environment. You could try asking Rich Hillary at CSIRO about
this approach.
}
\references{
http://www.ccsbt.org/site/recent_assessment.php
}
\author{
T. Carruthers
}
