% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_Exported.R
\name{setup}
\alias{setup}
\title{Setup parallel processing}
\usage{
setup(cpus = parallel::detectCores(), type = "SOCK")
}
\arguments{
\item{cpus}{number of CPUs}

\item{type}{Type of cluster. Can be 'SOCK', 'MPI', 'PVM' or 'NWS'. Default is 'SOCK'.}
}
\description{
Sets up parallel processing using the snowfall package
}
