% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_Exported.R
\name{setup}
\alias{setup}
\title{Setup parallel processing}
\usage{
setup(cpus = parallel::detectCores() * 0.5, ...)
}
\arguments{
\item{cpus}{number of CPUs}

\item{...}{other arguments passed to 'snowfall::sfInit'}
}
\description{
Sets up parallel processing using the snowfall package
}
\examples{
\dontrun{
setup() # set-up half the available processors
setup(6) # set-up 6 processors
}
}
