% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_Exported.R
\name{setup}
\alias{setup}
\title{Setup parallel processing}
\usage{
setup(cpus = parallel::detectCores(), type = "SOCK")
}
\arguments{
\item{cpus}{number of CPUs}
}
\description{
Sets up parallel processing using the snowfall package
}
