% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_Internal.R
\name{simCAA}
\alias{simCAA}
\title{Simulate Catch-at-Age Data}
\usage{
simCAA(nsim, nyears, maxage, Cret, CAA_ESS, CAA_nsamp)
}
\arguments{
\item{nsim}{Number of simulations}

\item{nyears}{Number of years}

\item{maxage}{Maximum age}

\item{Cret}{Retained Catch at age in numbers - array(sim, years, maxage)}

\item{CAA_ESS}{CAA effective sample size}

\item{CAA_nsamp}{CAA sample size}
}
\value{
CAA array
}
\description{
CAA generated with a multinomial observation model from retained catch-at-age
data
}
