% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_Internal.R
\name{simCAL}
\alias{simCAL}
\title{Simulate Catch-at-Length Data}
\usage{
simCAL(nsim, nts, nyears, maxage, CAL_ESS, CAL_nsamp, nCALbins,
  CAL_binsmid, vn, retL, Linfarray, Karray, t0array, LenCV)
}
\arguments{
\item{nsim}{Number of simulations}

\item{nts}{Number of sub-year time-steps}

\item{nyears}{Number of years}

\item{maxage}{Maximum age - not used}

\item{CAL_ESS}{CAA effective sample size}

\item{CAL_nsamp}{CAA sample size}

\item{nCALbins}{number of CAL bins}

\item{CAL_binsmid}{mid-points of CAL bins}

\item{vn}{Vulnerable numbers-at-age}

\item{retL}{Retention at length curve}

\item{Linfarray}{Array of Linf values by simulation and year}

\item{Karray}{Array of K values by simulation and year}

\item{t0array}{Array of t0 values by simulation and year}

\item{LenCV}{CV of length-at-age#'}
}
\value{
named list with CAL array and LFC, ML, & Lc vectors
}
\description{
Simulate CAL and calculate length-at-first capture (LFC),
mean length (ML), modal length (Lc), and mean length over modal length (Lbar)
}
