% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{SimulatedData}
\alias{SimulatedData}
\alias{SimulatedData}
\title{SimulatedData Data}
\format{An object of class \code{Data} of length 1.}
\usage{
SimulatedData

SimulatedData
}
\description{
An object of class Data

An object of class Data
}
\keyword{datasets}
