% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataObjects.r
\docType{data}
\name{Simulation_1}
\alias{Simulation_1}
\title{Simulation_1 DLM_data}
\format{An object of class \code{DLM_data} of length 1.}
\usage{
Simulation_1
}
\description{
An object of class DLM_data
}
\keyword{datasets}

