% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{slotlim}
\alias{slotlim}
\title{An data-limited method which sets a slot limit}
\usage{
slotlim(x, Data, ...)
}
\arguments{
\item{x}{A position in a data-limited methods object}

\item{Data}{A data-limited methods object}

\item{...}{Optional additional arguments that are ignored. Note arguments
\code{reps} or \code{...} are required for all input controls}
}
\value{
An object of class 'Rec'
}
\description{
An example of the implementation of input controls in the DLM toolkit, where
selectivity-at-length is set using a slot limit; that is, a minimum and
maximum legal length.  The maximum limit is set here, quite arbitrarily, as
the 75th percentile between the new minimum legal length and the estimated
asymptotic length.
}
\references{
Made-up for this package
}
\author{
A. Hordyk
}
