% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_Output.R
\name{SPmod}
\alias{SPmod}
\title{Surplus production based catch-limit modifier}
\usage{
SPmod(x, Data, reps = 100, alp = c(0.8, 1.2), bet = c(0.8, 1.2))
}
\arguments{
\item{x}{A position in data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of quota samples}

\item{alp}{Condition for modifying the TAC (bounds on change in abundance)}

\item{bet}{Limits for how much the TAC can change among years}
}
\value{
A numeric vector of TAC recommendations
}
\description{
An MP that makes incremental adjustments to TAC recommendations based on the
apparent trend in surplus production. Based on the theory of Mark Maunder
(IATTC)
}
\details{
Note that this isn't exactly what Mark has previously suggested and is
stochastic in this implementation.
}
\references{
http://www.iattc.org/Meetings/Meetings2014/MAYSAC/PDFs/SAC-05-10b-Management-Strategy-Evaluation.pdf
}
\author{
T. Carruthers
}
