% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLM_output.R
\docType{data}
\name{SPslope}
\alias{SPslope}
\title{Slope in surplus production MP}
\format{An object of class \code{DLM_output} of length 1.}
\usage{
SPslope(x, DLM_data, reps = 100, yrsmth = 4, alp = c(0.9, 1.1), bet =
c(1.5, 0.9))
}
\arguments{
\item{x}{A position in data-limited methods data object}

\item{DLM_data}{A data-limited methods data object}

\item{reps}{The number of quota samples}

\item{yrsmth}{Years over which to smooth recent estimates of surplus
production}

\item{alp}{Condition for modifying the DLM_data (bounds on change in
abundance)}

\item{bet}{Limits for how much the DLM_data can change among years}
}
\value{
A numeric vector of DLM_data recommendations
}
\description{
A management procedure that makes incremental adjustments to TAC
recommendations based on the apparent trend in recent surplus production.
Based on the theory of Mark Maunder (IATTC)
}
\details{
Note that this isn't exactly what Mark has previously suggested and is
stochastic in this implementation.
}
\author{
T. Carruthers
}
\references{
http://www.iattc.org/Meetings/Meetings2014/MAYSAC/PDFs/SAC-05-10b-Management-Strategy-Evaluation.pdf
}
\keyword{datasets}

