% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StochasticSRA.R
\name{SRAcomp}
\alias{SRAcomp}
\title{Plot simulation test of Stochastic SRA method}
\usage{
SRAcomp(sim, OM, maxplot = 10, outfile = NA)
}
\arguments{
\item{sim}{The output list object of SRAsim() function.}

\item{sam}{The output object of StochasticSRA() function.}

\item{patchy}{The fraction of years that have catch at age data}

\item{nCAA}{The number of annual catch at age observations (same among all years)}
}
\value{
A list: Chist = historical catch series,Recdevs = historical recruitment deviations (mean = 1), CAA = catch at age matrix, N = numbers at age matrix, SSB = annual spawning biomass, FM = Fishing mortality rate at age matrix, M = natural mortality rate \code{classy}
}
\description{
Plots simulation variables versus estimation variables for Stochastic SRA methods of conditioning operating models.
}
\examples{
setup()
OM<-new('OM',Albacore,Generic_fleet,Generic_obs)
sim<-SRAsim(OM,qmult=1,patchy=0.8)
CAA<-sim$CAA
Chist<-sim$Chist
OM<-StochasticSRA(OM,CAA,Chist,nsim=30,nits=500)
SRAcomp(sim,OM) 
}
\author{
T. Carruthers (Canadian DFO grant)
}

