% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssessmentToOM.R
\name{SS2DLM}
\alias{SS2DLM}
\title{Function to read MLE estimates from SS3 file structure into an operating model}
\usage{
SS2DLM(SS3dir, nsim = 48, proyears = 50, Name = NULL,
  Source = "No source provided", Author = "No author provided")
}
\arguments{
\item{SS3dir}{A folder with SS3 input and output files in it}

\item{nsim}{The number of simulations to take for parameters with uncertainty (for OM@cpars custom parameters)}

\item{Name}{The name of the operating model}

\item{Source}{Reference to assessment documentation e.g. a url}

\item{Author}{Who did the assessment}
}
\description{
A function that uses the file location of a fitted SS3 model including input files to population the various slots of an operating model with MLE parameter estimates
}
\author{
T. Carruthers
}
