% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubOM.r
\name{SubOM}
\alias{SubOM}
\title{Subset a Stock, Fleet, or Obs object from an OM object}
\usage{
SubOM(OM, Sub = c("Stock", "Fleet", "Observation"))
}
\arguments{
\item{OM}{An operating model object (class OM)}

\item{Sub}{A character string specifying what object type to strip out
"Stock", "Fleet", or "Observation"}
}
\value{
An object of class Stock, Fleet, or Obs
}
\description{
A function that strips out a Stock, Fleet, or Obs object from a 
complete OM object. Mainly used for internal functions.
}
\author{
A. Hordyk
}

