% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineClasses.r
\docType{methods}
\name{summary,Data-method}
\alias{summary,Data-method}
\title{Summary of Data object}
\usage{
\S4method{summary}{Data}(object)
}
\arguments{
\item{object}{object of class Data}
}
\description{
Summary of Data object
}
