% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefinePMClasses.R
\docType{methods}
\name{summary,MSE-method}
\alias{summary,MSE-method}
\title{Summary of MSE object}
\usage{
\S4method{summary}{MSE}(object, ..., silent = FALSE)
}
\arguments{
\item{object}{object of class MSE}

\item{...}{a list of names of PM methods}

\item{silent}{Should summary be printed to console? Logical.}
}
\description{
Summary of MSE object
}
