% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineClasses.r
\docType{methods}
\name{summary,MSE-method}
\alias{summary,MSE-method}
\title{Summary of MSE object}
\usage{
\S4method{summary}{MSE}(object)
}
\arguments{
\item{object}{object of class MSE}
}
\description{
Summary of MSE object
}
