% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Functions.R
\name{TAC}
\alias{TAC}
\title{Calculate TAC recommendations for more than one MP}
\usage{
TAC(Data, MPs = NA, reps = 100, timelimit = 1)
}
\arguments{
\item{Data}{A data-limited methods data object}

\item{MPs}{optional vector of MP names}

\item{reps}{Number of repititions}

\item{timelimit}{The maximum time (seconds) taken to complete 10 reps}
}
\description{
A function that returns the stochastic TAC recommendations from a vector of
data-limited MPs (Output) given a data-limited data object Data
}
\examples{
\dontrun{
Data <- TAC(DLMtool::Cobia)
plot(Data)
}
}
\author{
T. Carruthers
}
