\name{Tplot}
\alias{Tplot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
A trade-off plot for an MSE object 
}
\description{
A shorter version of the plot method for MSEs that just shows the overall trade-offs
}
\usage{
Tplot(MSEobj)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{MSEobj}{
An object of class 'MSE'
}
}
\author{
T. Carruthers
}

