% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSEObj.r
\name{Tplot}
\alias{Tplot}
\title{A trade-off plot for an MSE object}
\usage{
Tplot(MSEobj,nam=NA)
}
\arguments{
\item{MSEobj}{An object of class 'MSE'}

\item{nam}{Name of the plot}
}
\description{
A shorter version of the plot method for MSEs that just shows the overall
trade-offs
}
\author{
T. Carruthers
}
