% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE_Plotting.R
\name{Tplot_old}
\alias{Tplot_old}
\alias{Tplot2_old}
\title{Trade-off plots for an MSE object}
\usage{
Tplot_old(MSEobj, nam = NA)

Tplot2_old(MSEobj, nam = NA)
}
\arguments{
\item{MSEobj}{An object of class 'MSE'}

\item{nam}{Name of the plot}
}
\description{
Three figures showing trade-offs between fishing mortality, biomass, and yield.
}
\section{Functions}{
\itemize{
\item \code{Tplot_old}: Used in the plot method for MSE objects that shows trade-off between
yield versus probability of overfishing and biomass levels (relative to BMSY).

\item \code{Tplot2_old}: Simpler plot that compares long-term yield (LTY:
fraction of simulations getting over half FMSY yield in the last ten years
of the projection), short-term yield (STY: fraction of simulations getting
over half FMSY yield in the first ten years of the projection), variability
in yield (VY: fraction of simulations where average annual variability in
yield is less than 10 per cent) and biomass level (B10: the fraction of
simulations in which biomass stays above 10 percent of BMSY).
}}

\seealso{
\link{TradePlot} \link{PerformanceMetric}
}
\author{
T. Carruthers & A. Hordyk
}
