% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_Exported.R
\name{trlnorm}
\alias{trlnorm}
\alias{tdlnorm}
\title{Lognormal distribution for DLMtool}
\usage{
trlnorm(reps, mu, cv)

tdlnorm(x, mu, cv)
}
\arguments{
\item{reps}{number of random numbers}

\item{mu}{mean}

\item{cv}{coefficient of variation}

\item{x}{vector}
}
\value{
numeric
}
\description{
Variant of rlnorm which returns the mean when reps = 1.
}
\section{Functions}{
\itemize{
\item \code{trlnorm}: Generate log-normally distributed random numbers

\item \code{tdlnorm}: Calculate density of log-normally distributed random numbers
}}

\author{
T. Carruthers
}
\keyword{internal}
