% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscInt.r
\name{TwoSidedFun}
\alias{TwoSidedFun}
\title{Internal Two sided selectivity curve}
\usage{
TwoSidedFun(LFS, s1, s2, lens)
}
\arguments{
\item{LFS}{first length at fulll selection}

\item{s1}{ascending slope}

\item{s2}{descending slope}

\item{lens}{vector of lengths}
}
\description{
Internal Two sided selectivity curve
}
\keyword{internal}

