% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_Exported.R
\name{updateMSE}
\alias{updateMSE}
\title{Update an MSE object with new slots}
\usage{
updateMSE(MSEobj)
}
\arguments{
\item{MSEobj}{A MSE object from a previous version of the DLMtool.
Also works with Stock, Fleet, Obs, Imp, and Data objects.}
}
\value{
An object of class matching class(MSEobj)
}
\description{
Updates an existing MSE object (class MSE) from a previous version of the
DLMtool to include the new slots.  The slots will be empty, but avoids the
'slot doesn't exist' error that sometimes occurs. Also works with Stock, Fleet,
Obs, Imp, and Data objects.
}
\author{
A. Hordyk
}
