% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSEObj.r
\name{updateMSE}
\alias{updateMSE}
\title{Update an MSE object with new slots}
\usage{
updateMSE(MSEobj)
}
\arguments{
\item{MSEobj}{A MSE object from a previous version of the DLMtool}
}
\value{
An object of class \code{MSE}
}
\description{
Updates an existing MSE object (class MSE) from a previous version of the
DLMtool to include the new slots.  The slots will be empty, but avoids the
'slot doesn't exist' error that sometimes occurs.
}
\author{
A. Hordyk
}
