% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleOM.R
\name{validcpars}
\alias{validcpars}
\title{Valid custom parameters (cpars)}
\usage{
validcpars(
  type = c("all", "Stock", "Fleet", "Obs", "Imp", "internal"),
  valid = TRUE
)
}
\arguments{
\item{type}{What cpars to show? 'all', 'Stock', 'Fleet', 'Obs', 'Imp', or 'internal'}

\item{valid}{Logical. Show valid cpars?}
}
\value{
a HTML datatable with variable name, description and type of valid cpars
}
\description{
Valid custom parameters (cpars)
}
\examples{
\dontrun{
validcpars() # all valid cpars

validcpars("Obs", FALSE) # invalid Obs cpars
}

}
