% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE_Plotting.R
\name{VOI}
\alias{VOI}
\title{Calculate Value Of Information}
\usage{
VOI(
  MSEobj,
  ncomp = 6,
  nbins = 8,
  maxrow = 8,
  Ut = NA,
  Utnam = "Utility",
  plot = TRUE
)
}
\arguments{
\item{MSEobj}{An object of class MSE}

\item{ncomp}{Maximum number of variables to examine per MP}

\item{nbins}{Number of percentile bins for sampled parameters of the
operating model or observation model, which is used for calculating
variability in utility across the sampled range of each parameter}

\item{maxrow}{maximum number of MPs per plot}

\item{Ut}{A matrix of user-specified utility values of nsim rows and nMPs
columns}

\item{Utnam}{The name of the utility measure for plotting}

\item{plot}{Logical. Show the plot?}
}
\description{
A function that relates operating model parameters and parameters of the
observation model to yield (by default). A user can also specific their own
utility values (Ut) which is arranged in a matrix of nsim rows and nMP
columns.
}
\author{
T. Carruthers
}
