% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMobjects.R
\name{Yield}
\alias{Yield}
\title{Performance Metric: Average Yield}
\usage{
Yield(MSEobj = NULL)
}
\arguments{
\item{MSEobj}{An object of class MSE}
}
\value{
An object of class PMobj
}
\description{
Performance Metric: Average Yield
}
\examples{
\dontrun{
Yield(myMSE)
}
}
