% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output.r
\name{YPR}
\alias{YPR}
\title{Yield Per Recruit analysis to get FMSY proxy F01}
\usage{
YPR(x, Data, reps = 100)
}
\arguments{
\item{x}{A position in a data-limited methods data object}

\item{Data}{A data-limited methods data object}

\item{reps}{The number of samples of the TAC}
}
\value{
A numeric vector of TAC samples
}
\description{
A simple yield per recruit approximation to FMSY (F01) which is the position
of the ascending YPR curve for which dYPR/dF = 0.1(dYPR/d0)
}
\note{
Based on the code of Meaghan Bryan
}
\references{
Beverton and Holt. 1954.
}
\author{
Meaghan Bryan and Tom Carruthers
}
